% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{avonet}
\alias{avonet}
\title{Table of species morphological traits}
\format{
An object of class \code{data.frame} with 10661 rows and 4 columns.
}
\source{
\url{https://doi.org/10.6084/m9.figshare.16586228.v5}
}
\usage{
avonet
}
\description{
A dataset containing the species name, wing length, secondary length and mass from 10661 species.
}
\keyword{datasets}
