% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_create.R
\name{bird_create}
\alias{bird_create}
\title{Create bird flight traits}
\usage{
bird_create(
  scientific_name,
  mass = NULL,
  wing_span = NULL,
  wing_aspect = NULL,
  wing_area = NULL,
  body_frontal_area = NULL,
  species_name = lifecycle::deprecated()
)
}
\arguments{
\item{scientific_name}{Scientific name of the species}

\item{mass}{Mass of the bird in kilograms}

\item{wing_span}{Wing span in meters}

\item{wing_aspect}{Wing aspect ratio (no unit)}

\item{wing_area}{Wing area in meter square}

\item{body_frontal_area}{Body frontal area in meter square}

\item{species_name}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{scientific_name} instead}
}
\value{
List containing mass, wing span, wing aspect ratio, and body frontal area.
}
\description{
This function returns a list with the four morphological information necessary to construct the
power curve: mass, wing span, wing aspect ratio, and body frontal area.

When any of these variables are missing, we query the
AVONET database (\doi{10.6084/m9.figshare.16586228.v5}) using the scientific name from
\href{https://www.birds.cornell.edu/clementschecklist/}{the Clements Checklist}.
}
\examples{
# Using AVONET dataset
bird_create("Acrocephalus arundinaceus")

# Using AVONET dataset + custom values
bird_create("Acrocephalus arundinaceus", wing_aspect = 8)

# Import your own bird. You will need mass, and at least two of wing_span,
# wing_aspect or wing_area.
bird_create("Madynuss nutshell", mass = 8, wing_span = 0.2, wing_aspect = 4)

}
\seealso{
Other bird:
\code{\link[=print.bird]{print.bird()}}

Other movement:
\code{\link[=graph_set_movement]{graph_set_movement()}},
\code{\link[=plot_graph_movement]{plot_graph_movement()}},
\code{\link[=speed2prob]{speed2prob()}},
\code{\link[=tag_download_wind]{tag_download_wind()}}
}
\concept{bird}
\concept{movement}
