% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM.R
\name{classify_flap}
\alias{classify_flap}
\title{Classify flapping flight}
\usage{
classify_flap(dta, period = 12, tz = "UTC")
}
\arguments{
\item{dta}{data stored as a list see str(data(PAM_data)) for example format}

\item{period}{number of timepoints after which behaviour is considered migratory e.g. for hoopoes, 3x5min = 15 minutes of intense activity is considered flight}

\item{tz}{timezone, default is "UTC"}
}
\value{
timetable: a timetable for when the species was migrating or not,

classification: a classification timeseries where datetime corresponds to activity, and

no_movement: the value in classification which corresponds to no movement

low_movement: the value in classification which corresponds to low activity

high_movement: the value in classification which corresponds to high activity

migration: the value in classification which corresponds to migratory flapping flight

threshold: the threshold between low and high activity
}
\description{
This function uses activity data to classify migratory flapping flight.
}
