% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{classify_pam}
\alias{classify_pam}
\title{Automatic classification of pam}
\usage{
classify_pam(pam, min_duration = 30)
}
\arguments{
\item{pam}{data list}

\item{min_duration}{duration in minutes}
}
\value{
pam
}
\description{
This function uses activity data to classify migratory flapping flight.
Inspired by \code{\link{classify_flap}}
(https://github.com/KiranLDA/pamLr/blob/master/R/classify_flap.R) from
\href{https://github.com/KiranLDA/pamLr}{pamLr}
}
\seealso{
\code{\link[pamLr]{classify_flap}}
}
