% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM.R
\name{create_crop}
\alias{create_crop}
\title{Crop all sensor data to the same timeframe}
\usage{
create_crop(dta, start, end)
}
\arguments{
\item{dta}{path where files are stored}

\item{start}{posicxt object for date that PAM data should start}

\item{end}{posicxt object for date that PAM data should end}
}
\value{
shortened PAM data
}
\description{
Get rid of excess data. e.g. when a logger is kept in a rucksack or a lab before being downloaded.
}
