% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM.R
\name{create_import}
\alias{create_import}
\title{Import PAM data}
\usage{
create_import(
  pathname = pathname,
  measurements = c(".pressure", ".glf", ".acceleration", ".temperature", ".magnetic")
)
}
\arguments{
\item{pathname}{path where files are stored}

\item{measurements}{a series of measurements logged by the PAM logger which are to be imported. Currently supports these file extentions: ".pressure", ".glf", ".gle",".acceleration", ".temperature", "AirTemperature", ".BodyTemperature" and ".magnetic"}
}
\value{
a list of measurements for the one individual
}
\description{
Imports and formats many datasets into one big nested list containing all the data from the different sensors. A subset of sensors can be selected using \code{measurements}.
}
