% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{crop_pam}
\alias{crop_pam}
\title{Crop pam in time}
\usage{
crop_pam(pam, crop_start, crop_end)
}
\arguments{
\item{pam}{data list}

\item{crop_start}{posicxt object for date that pam data should start}

\item{crop_end}{posicxt object for date that pam data should end}
}
\value{
shortened pam data
}
\description{
Crop the timeserie of all sensor to a given time frame
}
