% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight.R
\name{find_twilights}
\alias{find_twilights}
\title{Returns twilights for each day based on a threshold of light}
\usage{
find_twilights(light, threshold = NA, shift_k = NA)
}
\arguments{
\item{light}{a dataframe with columns \code{date} and \code{obs} that are the sequence of sample
times (as POSIXct) and light levels recorded by the tag.}

\item{threshold}{the light threshold that defines twilight. If not provided, it uses the first
light (i.e, \code{obs>0}).}

\item{shift_k}{shift of the middle of the night compared to 00:00 UTC (in seconds). If not
provided, it will take the middle of all nights.}
}
\value{
A data.frame with columns \code{twilight} (date-time of twilights) and \code{rise} (logical)
}
\description{
Search for pairs of sunset, sunrise that correspond to a given light threshold. Function inspired
from \href{https://rdrr.io/github/slisovski/TwGeos/man/findTwilights.html}{\code{findTwilights}} of the
package \code{TwGeos}.
}
\examples{
pam <- pam_read(
  pathname = system.file("extdata/0_PAM/18LX", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
twl <- find_twilights(pam$light)
head(twl)
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/light-map.html#twilight-annotation}{GeoPressureManual | Light Map }
}
