% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flight_prob}
\alias{flight_prob}
\title{Movement model}
\usage{
flight_prob(
  speed,
  method = "gamma",
  shape = 7,
  scale = 7,
  bird = NA,
  fun_power = function(power) {     (1/power)^3 },
  low_speed_fix = 15
)
}
\arguments{
\item{speed}{airspeed or groundspeed in km/h}

\item{method}{method used to convert the speed to probability ("gamma" or
"power")}

\item{bird}{list of basic morphological trait necessary: mass, wing span,
wing aspect ratio and body frontal area. It is best practice to create bird
with \code{flight_bird()}.}
}
\value{
Probability values corresponding to the speed provided
}
\description{
Compute the mechanical power required for a specific bird flying as at a
given airspeed in km/h. \code{bird} (created with \code{flight_bird()})
}
\examples{
speed <- seq(1, 120)
low_speed_fix <- 20 # minimum speed allowed
prob <- flight_prob(speed,
  method = "gamma", shape = 7, scale = 7,
  low_speed_fix = low_speed_fix
)
plot(speed, prob,
  type = "l", xlab = "Groundspeed [km/h]", ylab =
    "Probability"
)
abline(v = low_speed_fix)
bird <- flight_bird("Acrocephalus arundinaceus")
prob <- flight_prob(speed, method = "power", bird = bird)
plot(speed, prob, type = "l", xlab = "Airspeed [km/h]", ylab = "Probability")
}
