% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_expand.R
\name{geo_expand}
\alias{geo_expand}
\title{Compute grid from \code{extent} and \code{scale}}
\usage{
geo_expand(extent, scale)
}
\arguments{
\item{extent}{Geographical extent of the map on which the likelihood and graph model will be
computed. Vector of length 4 \code{c(xmin, xmax, ymin, ymax)} or \code{c(W, E, S, N)}.}

\item{scale}{Number of pixels per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and \code{scale=4} for a resolution of 0.25° (~30km). To avoid
interpolating the ERA5 data, the scale should be smaller than 10. Read more about scale on the \href{https://developers.google.com/earth-engine/guides/scale}{Google earth Engine documentation}.}
}
\value{
A list containing:
\itemize{
\item \code{extent} same as input
\item \code{scale} same as input
\item \code{lat} vector of latitude
\item \code{lon} vector of longitude
\item \code{dim} vector of length 2 of the dimension of the map (number of pixel in lat and lon)
}
}
\description{
Take the two GEOgraphic parameters used to define the grid in \code{geostap} and \code{graph} and return
all variable of interest for the spatial grid (e.g., grid dimension, latitude and longitude...)
}
\examples{
str(geo_expand(extent = c(0, 10, 0, 5), scale = 1))

str(geo_expand(extent = c(-16, 23, 0, 50), scale = 10))
}
