% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight_light2mat.R
\name{geolight_light2mat}
\alias{geolight_light2mat}
\title{Convert light data in matrix format}
\usage{
geolight_light2mat(light, twl_offset = 0)
}
\arguments{
\item{light}{A dataframe with columns \code{date} and \code{value} that are the sequence of sample
times (as POSIXct) and light levels recorded by the tag.}

\item{twl_offset}{Shift of the middle of the night compared to 00:00 UTC (in seconds). If not
provided, it uses the middle of all nights.}
}
\value{
A dataframe with columns \code{date} and \code{value}.
}
\description{
Convert light data in matrix format
}
