% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight.R
\name{geolight_likelihood}
\alias{geolight_likelihood}
\title{Compute likelihood map for light data}
\usage{
geolight_likelihood(
  twl,
  lon_calib,
  lat_calib,
  map,
  stap = NA,
  adjust_calib = 1.4,
  w_llp = 0.1,
  fit_z_return = F
)
}
\arguments{
\item{twl}{A data.frame with columns \code{twilight} (date-time of twilights), \code{calib} (logical)
indicating if the twilight occurs during the calibration period and \code{discard} .
See \code{\link[=geolight_twilight]{geolight_twilight()}}.}

\item{lon_calib}{longitude of the calibration site.}

\item{lat_calib}{latitude of the calibration site.}

\item{map}{SpatRaster on which the likelihood map should be computed.}

\item{stap}{stationary periods on which the likelihood maps should be computed.}

\item{adjust_calib}{smoothing parameter for the kernel density. See \code{\link[stats:kernel]{stats::kernel()}}}

\item{w_llp}{Log-linear pooling aggregation weight. See [GeoPressureManual | Probability}

\item{fit_z_return}{logical forcing to return the fit of the calibration.}
}
\value{
list of SpatRaster of the likelihood map for each stationary period. Or a kernel fit
(see \code{\link[stats:kernel]{stats::kernel()}}) if \code{fit_z_return} is true.
}
\description{
Instead of calibrating the twilight errors in terms of duration, we directly model the zenith
angle error
}
\details{
The adjust parameter allows to manually set how smooth you want the fit to be. Because the
zenith angle error model is fitted with data only at the calibration site and that we are using
it for all locations of the bird’s journey, it is safer to assume a broader/smoother
distribution.
}
