% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight.R
\name{geolight_refracted}
\alias{geolight_refracted}
\title{Atmospheric Refraction}
\usage{
geolight_refracted(zenith)
}
\arguments{
\item{zenith}{zenith angle (degrees) to adjust.}
}
\value{
vector of zenith angles (degrees) adjusted for atmospheric refraction.
}
\description{
Adjust the solar zenith angle for atmospheric refraction.
}
\details{
Given a vector of solar zeniths computed by \code{\link{geolight_zenith}}, \code{\link{geolight_refracted}}
calculates the solar zeniths adjusted for the effect of atmospheric refraction.
}
