% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight.R
\name{geolight_zenith}
\alias{geolight_zenith}
\title{Solar Zenith Angle}
\usage{
geolight_zenith(sun, lon, lat)
}
\arguments{
\item{sun}{list of solar time and declination computed by \code{geolight_solar}.}

\item{lon}{vector of longitudes.}

\item{lat}{vector latitudes.}
}
\value{
A vector of solar zenith angles (degrees) for the given locations and times.
}
\description{
Calculate the solar zenith angle for given times and locations
}
\details{
\code{geolight_zenith} uses the solar time and declination calculated by \code{geolight_solar} to compute
the solar zenith angle for given times and locations, using the same methods as
\url{https://gml.noaa.gov/grad/solcalc/}.  This function does not adjust for atmospheric
refraction see \code{\link{geolight_refracted}}.
}
\examples{
# Approx location of Sydney Harbour Bridge
lon <- 151.211
lat <- -33.852
# Solar zenith angle for noon on the first of May 2000
# at the Sydney Harbour Bridge
s <- geolight_solar(as.POSIXct("2000-05-01 12:00:00", "EST"))
geolight_zenith(s, lon, lat)
}
\seealso{
\code{\link{geolight_solar}}
}
