% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure.Probmap}
\alias{geopressure.Probmap}
\title{Convert MSE to probability}
\usage{
geopressure.Probmap(rasterList, s = 1, thr = 0.9)
}
\arguments{
\item{rasterList}{list of raster loaded from \code{geopressure.map()}}

\item{s}{standardeviation of the pressure error}

\item{thr}{threashold of the percentage of datapoint outside the elevation range to be considered not possible}
}
\value{
List of the probability raster map
}
\description{
This function
}
