% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure.timeseries}
\alias{geopressure.timeseries}
\title{Request and download surface pressure timeseries at location}
\usage{
geopressure.timeseries(
  lon,
  lat,
  pressure = NULL,
  endTime = NULL,
  startTime = NULL
)
}
\arguments{
\item{lon}{longitude to query (-180째 to 180째).}

\item{lat}{latitude to query (0째 to 90째).}

\item{pressure}{pressure list from PAM logger dataset list (see https://github.com/KiranLDA/PAMLr for details).}

\item{endTime}{same as starttime}

\item{startTime}{if pressure not provided, then the starttime of the timeserie return is needed}
}
\value{
Timeserie of date, pressure and optionally altitude
}
\description{
This function return the surfrace atmospheric pressure timeseries from ERA5
at a particualy location specify by lat and lon. I uses SRTM-30 to translate
the pressure for the exact elevation of the ground level, accounting for
both temporal varation of pressure and temperature.
}
\details{
If you supply the pressure (and time) of the geolocator, it will additionally
return the elevation of the geolocator above sea level considering that the bird was located at the location specify
}
