% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_export.R
\name{geopressure_export}
\alias{geopressure_export}
\title{Export GeoPressure Variable}
\usage{
geopressure_export(
  tag = NA,
  geostap = NA,
  path = NA,
  directory = glue::glue("data/export/{tag$id}")
)
}
\arguments{
\item{tag}{Data logger list with label information. See \code{\link[=tag_label]{tag_label()}} for the required input.}

\item{geostap}{List of the geographical and stationary period information. See
\code{\link[=geostap_create]{geostap_create()}} for details.}

\item{path}{Data.frame containing.}

\item{directory}{Folder in which to export the variable}

\item{file}{Name of the file for the exported variable}

\item{edge}{See \code{\link[=path2edge]{path2edge()}}}
}
\description{
Set of functions to export the core variable of the process:
}
\details{
\itemize{
\item \code{tag}: export all sensors data.frame and \code{stap} data.frame, thus containing the \code{stap_id} and
\code{label} information.
\item \code{map}: export map (e.g., \code{geostap$map_pressure} or \code{geostap$map_marginal})
\item \code{path}: Export a \code{path} data.frame into a csv.
\item \code{edge}:
}
}
