% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_likelihood}
\alias{geopressure_likelihood}
\title{Compute Likelihood Map}
\usage{
geopressure_likelihood(
  pressure_mismatch,
  s = 1,
  thr = 0.9,
  fun_w = function(n) {
     log(n)/n
 }
)
}
\arguments{
\item{pressure_mismatch}{List of SpatRaster built with \code{\link[=geopressure_mismatch]{geopressure_mismatch()}}.}

\item{s}{Standard deviation of the pressure error.}

\item{thr}{Threshold of the percentage of data point outside the elevation range to be considered
not possible.}

\item{fun_w}{function taking the number of sample of the timeseries used to compute the
probability map and return the log-linear pooling weight (see the
\href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation })}
}
\value{
List of the likelihood map for each stationary period. See \code{\link[=geopressure_mismatch]{geopressure_mismatch()}} for
description of the output list.
}
\description{
This function convert the map of \eqn{MSE} and altitude threshold \eqn{z_{thr}} computed
by \code{\link[=geopressure_mismatch]{geopressure_mismatch()}} into a likelihood map with,
}
\details{
\eqn{p = \exp \left(-w \frac{MSE}{s} \right) \left[z_{thr}>thr \right],}

where \eqn{s} is the standard deviation of pressure and \eqn{thr} is the threshold. Because the
auto-correlation of the timeseries is not accounted for in this equation, we use a log-linear
pooling weight \eqn{w=\log(n)/n}, with \eqn{n} is the number of data point in the timeserie.
}
\examples{
# See `geopressure_mismatch()` for generating pressure_mismatch
\dontrun{
pressure_likelihood <- geopressure_likelihood(
  pressure_mismatch,
  s = 0.4,
  thr = 0.9
)
pressure_likelihood_1 <- pressure_likelihood[[1]]
}
pressure_likelihood_1 <- readRDS(system.file("extdata/1_pressure/",
  "18LX_pressure_likelihood_1.rda",
  package = "GeoPressureR"
))
pressure_likelihood_1
terra::plot(pressure_likelihood_1$map,
  main = "Probability",
  xlim = c(5, 20), ylim = c(42, 50)
)
}
\seealso{
\code{\link[=geopressure_mismatch]{geopressure_mismatch()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map }
}
