% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_map2path}
\alias{geopressure_map2path}
\title{Return the most likely path from a probability map}
\usage{
geopressure_map2path(map, interp = 0, format = "lonlat")
}
\arguments{
\item{map}{List of raster containing probability map of each stationary period. The metadata of
\code{map} needs to include the start and end time of the stationary period .}

\item{interp}{The position of the stationary period shorter than \code{interp} will be
replace by a linear average from other position (in days) .}

\item{format}{One of \code{"lonlat"}, \code{"ind"}, \code{"arr.ind"}. return the path in lon-lat or indices}
}
\value{
a data.frame of the position containing latitude (\code{lat}), longitude (\code{lon}) and the
stationary period id (\code{sta_id}) as column. Optionally, if indexes were requested, it will be
return. You will need to use \code{which.max(as.matrix(raster))} and not \code{which.max(raster)} to get
the correct location.
}
\description{
Find the location of the highest value in the map and return a path data.frame containing the
latitude and longitude. \code{interp} can be used to interpolate unrealistic position from short
stationary period based on the position of the longer ones. The interpolation assumes that the
first and last stationary period can be safely estimated from the probability map.
}
\examples{
# See `geopressure_prob_map()` for generating pressure_prob
pressure_prob_1 <- readRDS(system.file("extdata/1_pressure/", "18LX_pressure_prob_1.rda",
  package = "GeoPressureR"
))
geopressure_map2path(list(pressure_prob_1))
}
\seealso{
\code{\link[=geopressure_prob_map]{geopressure_prob_map()}}, \code{\link[=geopressure_ts_path]{geopressure_ts_path()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#compute-altitude}{GeoPressureManual | Pressure Map }
}
