% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_map_likelihood.R
\name{geopressure_map_likelihood}
\alias{geopressure_map_likelihood}
\title{Convert mismatch map into likelihood map}
\usage{
geopressure_map_likelihood(
  geostap,
  sd = 1,
  thr_mask = 0.9,
  log_linear_pooling_weight = function(n) log(n)/n
)
}
\arguments{
\item{geostap}{List of the geographical and stationary period information. See
\code{\link[=geostap_create]{geostap_create()}} for details.}

\item{sd}{Standard deviation of the pressure error.}

\item{thr_mask}{Threshold of the percentage of data points outside the elevation range to be considered
not possible.}

\item{log_linear_pooling_weight}{Weighting function of the log-linear pooling, taking the number of samples of the
stationary periods used and returning the weight of the aggregation. See
\href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation } for more details.}
}
\description{
This function converts the mismatch maps (MSE and mask) into a likelihood map.
}
\details{
We convert the map of the mean square error \eqn{MSE} and altitude mask \eqn{z_{mask}} computed
by \code{\link[=geopressure_map_mismatch]{geopressure_map_mismatch()}} into a likelihood map with,

\deqn{L = \exp \left(-w \frac{MSE}{\sigma} \right) \left[z_{mask}>T \right],}

where \eqn{\sigma} is the standard deviation of pressure and \eqn{T} is the mask threshold.

Because the auto-correlation of the timeseries is not accounted for in this equation, we use a
log-linear pooling weight of \eqn{w=\log(n)/n} by default, where \eqn{n} is the number of samples
in the timeserie (i.e., data points used to compute the MSE).

For more background and details on the algorithm, please read the \href{https://doi.org/10.1111/2041-210X.14043}{associated scientific publication } and the \href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation }.
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, and Felix Liechti. 2023.
“Global Positioning with Animal‐borne Pressure Sensors.” \emph{Methods in Ecology and Evolution}.
\url{https://doi.org/10.1111/2041-210X.14043}.}
}
\seealso{
Other geopressure_map: 
\code{\link{geopressure_map_mismatch}()},
\code{\link{geopressure_map}()}
}
\concept{geopressure_map}
