% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_mismatch}
\alias{geopressure_mismatch}
\title{Request and download mismatch and threshold maps of pressure}
\usage{
geopressure_mismatch(
  tag,
  extent,
  scale = 10,
  max_sample = 250,
  margin = 30,
  timeout = 60 * 5,
  workers = 90
)
}
\arguments{
\item{tag}{data logger dataset list (see \code{\link[=tag_read]{tag_read()}}). This list needs to contains a
\code{pressure} data.frame nwith variable \code{date} as POSIXt, \code{value} in hPa, \code{stap} grouping
observation measured during the same stationary period and \code{label} to label observation which
need to be discarded. In addition also need to contains \code{stap} with exact time of the stationary
period and in between flights.}

\item{extent}{Geographical extent of the map to query as a list ordered by North, West, South,
East  (e.g. \code{c(50,-16,0,20)}).}

\item{scale}{Number of pixel per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and 4 for a resolution of 0.25° (~30km). To avoid interpolating the
ERA5 data, scale should be smaller than 10. Read more about \href{https://developers.google.com/earth-engine/guides/scale}{scale on Google earth Engine documentation}.}

\item{max_sample}{The computation of the maps is only performed on \code{max_sample} datapoints of
pressure to reduce computational time. The samples are randomly (uniformly) selected on the
timeserie.}

\item{margin}{The margin is used in the threshold map to accept some measurement error. unit in
meter. (1hPa~10m)}

\item{timeout}{duration (sec) before the code is interrupted both for the request on
GeoPressureAPI and GEE. See \code{\link[httr:timeout]{httr::timeout()}}}

\item{workers}{number of parrellel request on GEE. Between 1 and 99.}
}
\value{
List of the misfit map for each stationary period, containing:
\itemize{
\item \code{map}: list of the MSE and threashold (see description above)
\item \code{stap} index of stationary period.
\item \code{nb_sample} number of pressure datapoint used.
\item \code{temporal_extent} datetime of the start and end of the stationary period
}
}
\description{
This function returns for each stationary period (1) a map of mismatch between the pressure
measured by a geolocator and the ERA5 pressure database and (2) a map altitude threshold using
a fine scale digital elevation model.
}
\details{
These maps are generated on Google Earth Engine via the Pressure map entry point of the
\href{https://raphaelnussbaumer.com/GeoPressureAPI/#description}{GeoPressure API}. The computation
performed by this function consists of the following:
\enumerate{
\item Send a query to generate the Google Earth Engine (GEE) url of the code producing the maps for
each stationary periods separately.
\item Download and read these geotiff maps as SpatRaster
}

The maps of each stationary period are returned in two layers:
\enumerate{
\item The mismatch between the input pressure timeseries and the reanalysis one at each location.
This is computed with a mean square error (MSE) where the mean error is removed. The mean error
is removed because we assume no specific altitude of the geolocator, thus allowing an
altitudinal shift of the pressure timeserie.
\item The proportion of datapoint of the input pressure timeseries corresponding to altitude value
which fall within the min and max ground elevation found at each location. The altitude value
of the geolocator pressure timeseries is computed with the barometric formula accounting for the
temporal variation of pressure (surface-pressure) and temperature (2m-temperature) based on
ERA5 data. The min and max ground elevation of each pixel is computed from SRTM-90.
}
}
\examples{
# See `tag_stap()` for generating tag
\dontrun{
pressure_mismatch <- geopressure_mismatch(
  tag$pressure,
  extent = c(50, -16, 0, 23),
  scale = 4,
  max_sample = 250,
  margin = 30
)
pressure_mismatch_1 <- pressure_mismatch$map[[1]]
}
pressure_mismatch_1 <- readRDS(system.file("extdata/1_pressure/", "18LX_pressure_mismatch_1.rda",
  package = "GeoPressureR"
))
pressure_mismatch_1
terra::plot(pressure_mismatch_1$map,
  main = c("Mean Square Error", "Mask of pressure")
)
}
\seealso{
\code{\link[=geopressure_likelihood]{geopressure_likelihood()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map }
}
