% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_timeseries_latlon.R
\name{geopressure_timeseries_latlon}
\alias{geopressure_timeseries_latlon}
\title{Request and download pressure timeseries at a given location}
\usage{
geopressure_timeseries_latlon(
  lat,
  lon,
  pressure = NULL,
  start_time = NULL,
  end_time = NULL,
  timeout = 60 * 5,
  quiet = FALSE
)
}
\arguments{
\item{lat}{Latitude to query (0째 to 90째).}

\item{lon}{Longitude to query (-180째 to 180째).}

\item{pressure}{Data.frame of the pressure measurements, usually \code{tag$pressure}.}

\item{start_time}{If \code{pressure} is not provided, \code{start_time} defines the start time of
the timeseries as POSIXlt.}

\item{end_time}{If \code{pressure} is not provided, \code{end_time} defines the end time of
the timeseries as POSIXlt.}

\item{timeout}{Duration (sec) before the code is interrupted both for the request on
GeoPressureAPI and GEE. See \code{\link[httr:timeout]{httr::timeout()}}.}

\item{quiet}{Hides the progress of the query (logical).}
}
\value{
A data.frame containing
\itemize{
\item \code{date} POSIXct date time
\item \code{pressure_era5} pressure (hPa)
\item \code{lon} same as input \code{lon} except if over water
\item \code{lat} same as input \code{lat} except if over water.
\item \code{pressure_era5_norm} only if \code{pressure} is provided as input
\item \code{altitude} only if \code{pressure} is provided as input
}
}
\description{
This function returns the surface atmospheric pressure timeseries from ERA5 at any requested
location.
}
\details{
If the location queried is over water, the location will be moved to the closest onshore location.

The ERA5 pressure timeseries of the response \eqn{P_{ERA}} will be provided on a hourly basis between
\code{start_time} and \code{end_time} or the same as \code{pressure$date} if \code{pressure} is supplied.

If you supply the \code{pressure} of the geolocator \eqn{P_{gl}}, the function will
additionally return the altitude of the geolocator above sea level \eqn{z_{gl}} using the
barometric equation,
\deqn{ z_{{gl}}(x)=z_{ERA5}(x) + \frac{T_{ERA5}(x)}{L_b}  \left( \frac{P_{gl}}{P_{ERA5}(x)}
\right)^{\frac{RL_b}{g M}-1},}
where \eqn{z_{ERA}}, \eqn{T_{ERA}}, and \eqn{P_{ERA}} respectively correspond to the ground level
elevation, temperature at 2m, and ground level pressure of ERA5, \eqn{L_b}  is the standard
temperature lapse rate, \eqn{R} is the universal gas constant, \eqn{g} is the gravity constant
and  \eqn{M} is the molar mass of air. See more information at
\href{https://raphaelnussbaumer.com/GeoPressureAPI/#description-1}{the GeoPressureAPI documentation}.

To be able to compare the temporal variation of the retrieved pressure of ERA5 \eqn{P_{ERA}} to
the geolocator pressure \eqn{P_{gl}}, the function also returns the ERA pressure normalized with
the geolocator mean pressure measurement as \code{pressure_era5_norm}.
\deqn{ P_{ERA5,0}(\boldsymbol{x})[t] = \left( P_{ERA5}(\boldsymbol{x})[t]-P_{gl}[t]\right) -
\left( \frac{1}{n}\sum_{i=1}^{n} P_{ERA5}(\boldsymbol{x})[i]-P_{gl}[i] \right).}

See \href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation}
for more information on the meaning of this value.
}
\examples{
# Request pressure at a given location
path_pres <- geopressure_timeseries_latlon(
  lat = 46, lon = 6,
  start_time = "2017-01-01 00:00",
  end_time = "2017-01-02 00:00"
)

str(path_pres)

plot(path_pres$date, path_pres$pressure,
  type = "b", ylab = "Pressure (hPa)", xlab = "Datetime"
)

# Retrieve the altitude of a bird being at this location adding random noise on the sensor.
path_pres <- geopressure_timeseries_latlon(
  lat = 46, lon = 6,
  pressure = data.frame(
    data.frame(
      date = path_pres$date,
      value = path_pres$pressure_era5 + rnorm(nrow(path_pres))
    )
  )
)

str(path_pres)

plot(path_pres$date, path_pres$altitude,
  type = "b", ylab = "Altitude (m)", xlab = "Datetime"
)

}
\seealso{
\code{\link[=geopressure_timeseries]{geopressure_timeseries()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map}
}
