% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoPressureR.R
\docType{package}
\name{GeoPressureR}
\alias{GeoPressureR}
\title{GeoPressureR: Global positioning by atmospheric pressure.}
\description{
GeoPressureR is a R package which help researchers determine the position of a bird based on the
data retrieved from multi-sensor geolocators.
}
\details{
The main novelty of this package is to match atmospheric pressure measurements to weather
reanalysis database (ERA-5) and produce a probability maps of position.

In addition, we use a graphical model to combine pressure, light, acceleration and windspeed
data into a trajectory model of the bird.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GeoPressure/GeoPressureR}
  \item \url{https://geopressure.org/GeoPressureR/}
  \item Report bugs at \url{https://github.com/GeoPressure/GeoPressureR/issues}
}

}
\author{
\strong{Maintainer}: Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-1020}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-1020}{ORCID}) [copyright holder]
  \item Mathieu Gravey \email{mathieu@mgravey.com} (\href{https://orcid.org/0000-0002-0871-1507}{ORCID})
}

}
