% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostap2path.R
\name{geostap2path}
\alias{geostap2path}
\title{Return the path from a likelihood map}
\usage{
geostap2path(geostap, likelihood = NA, interp = -1)
}
\arguments{
\item{geostap}{List of the geographical and stationary period information. See
\code{\link[=geostap_create]{geostap_create()}} for details.}

\item{likelihood}{Field of the \code{geostap} list containing the likelihood map (character). Default
\code{NA} is to take the product of \code{map_pressure} and \code{map_light} if available.}

\item{interp}{The position of the stationary period shorter than \code{interp} will be
replace by a linear average from other position (in days) .}
}
\value{
A path data.frame
\itemize{
\item \code{stap_id} stationary period
\item \code{ind} indices of the coordinate in the 2D grid. Useful to retrieve map or graph information.
\item \code{lat} Latitude,
\item \code{lon} longitude
}
}
\description{
Find the position of the highest value in a map, typically most probable value in a likelihood
map.
}
\details{
Note that this path is the most likely, considering only the observation model and ignoring the
movement model. Prefer to use \code{graph_most_likely()} for the most realistic path.

\code{interp} can be used to interpolate unrealistic position from short stationary period based on
the position of the longer ones. The interpolation assumes that the first and last stationary
period can be safely estimated from the probability map.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |>
  tag_label()
geostap <- geostap_create(tag,
  extent = c(-16, 23, 0, 50),
  scale = 2,
) |>
  geopressure_map(tag$pressure)

# Compute the path
path <- geostap2path(geostap)

}
\seealso{
\code{\link[=geopressure_map_likelihood]{geopressure_map_likelihood()}}, \code{\link[=geopressure_timeseries]{geopressure_timeseries()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#compute-altitude}{GeoPressureManual | Pressure Map}
}
