% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostap_create.R
\name{geostap_create}
\alias{geostap_create}
\title{Create a \code{geostap}}
\usage{
geostap_create(
  tag,
  extent,
  scale = 10,
  known = data.frame(stap_id = integer(), known_lat = double(), known_lon = double()),
  stap_include = tag$stap$stap_id
)
}
\arguments{
\item{tag}{Data logger list with label information. See \code{\link[=tag_label]{tag_label()}} for the required input.}

\item{extent}{Geographical extent of the map on which the likelihood and graph model will be
computed. Vector of length 4 \code{c(xmin, xmax, ymin, ymax)} or \code{c(W, E, S, N)}.}

\item{scale}{Number of pixels per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and \code{scale=4} for a resolution of 0.25° (~30km). To avoid
interpolating the ERA5 data, the scale should be smaller than 10. Read more about scale on the \href{https://developers.google.com/earth-engine/guides/scale}{Google earth Engine documentation}.}

\item{known}{Data.frame containing the known positions of the bird (e.g., equipment or retrieval
site). This information can only be attached at the level of a stationary period.}

\item{stap_include}{Vector of the stationary period to model, that is, to compute in the likelihood
map and use in the graph.}
}
\value{
List of the misfit map for each stationary period, containing:
\itemize{
\item \code{ìd}: Tag identifier, same as \code{tag$id}
\item \code{stap}: Data.frame of all stationary periods. Same as \code{tag$stap} but with three new columns:
\code{known_lat} and \code{known_lon} define the known position during these stationary periods, and \code{model}
defines whether the likelihood map of this stationary period should be computed and
later used in the graph.
\item \code{extent} same as parameter \code{extent}
\item \code{scale} same as parameter \code{scale}
}
}
\description{
This function creates a \code{geostap} list, which defines all the essential information needed to construct the
likelihood map and later create the graph.

\code{geostap} stands for \strong{geo}graphical and \strong{sta}tionary \strong{p}eriod, as these are the two
key dimensions (space-time) which are defined here for the rest of the analysis.

In addition, \code{geostap} also includes the ability to define \code{known} locations (e.g., equipment or
retrieval site). These can only be defined at the level of a stationary period (i.e., assuming
constant position during the whole stationary period) but you can define as many known stationary
periods as you wish. No likelihood map will be computed for these stationary periods, thus saving
computational time.

Furthermore, it is possible to compute (and later model) a subset of the stationary periods.
This is done with \code{stap_include}. By default, \code{stap_include} includes all stationary periods.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |>
  tag_label()

# Default geostap
geostap <- geostap_create(tag, c(-16, 23, 0, 50))
str(geostap)

# Customized geostap, with coarse grid scale, known position for the first stationary period and
# considering only the stationary periods lasting more than 20hours.
geostap <- geostap_create(tag,
  extent = c(-16, 23, 0, 50),
  scale = 1,
  known = data.frame(
    stap_id = 1,
    known_lon = 17.05,
    known_lat = 48.9
  ),
  stap_include = which(difftime(tag$stap$end, tag$stap$start, units = "hours") > 20)
)
str(geostap)
}
