% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_add_movement.R
\name{graph_add_movement}
\alias{graph_add_movement}
\title{Define the movement model}
\usage{
graph_add_movement(
  graph,
  type = ifelse("ws" \%in\% names(graph), "as", "gs"),
  ...
)
}
\arguments{
\item{graph}{graph constructed with \code{\link[=graph_create]{graph_create()}}}

\item{type}{Ground speed \code{"gs"} or airspeed \code{"as"}}

\item{...}{Arguments to be passed to \code{\link[=speed2prob]{speed2prob()}}}
}
\value{
Graph list with a new list \code{graph$movement} storing all the parameters needed to compute
the transition probability
}
\description{
Define the movement model used later by storing the parameters needed to build \code{\link[=speed2prob]{speed2prob()}}.
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \code{\link[=graph_transition]{graph_transition()}}, \code{\link[=graph_create]{graph_create()}},
\href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-2-marginal-probability-map}{GeoPressureManual | Basic graph}
}
