% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_add_wind.R
\name{graph_add_wind}
\alias{graph_add_wind}
\title{Compute windspeed and airspeed on a \code{graph}}
\usage{
graph_add_wind(graph, thr_as = Inf, variable = lifecycle::deprecated(), ...)
}
\arguments{
\item{graph}{a GeoPressureR graph object.}

\item{thr_as}{threshold of airspeed (km/h).}

\item{variable}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} No longer used.
\code{graph_add_wind()} always uses \code{c("u", "v")}.}

\item{...}{
  Arguments passed on to \code{\link[=edge_add_wind]{edge_add_wind}}
  \describe{
    \item{\code{pressure}}{pressure measurement of the associated \code{tag} data used to estimate the pressure
level (i.e., altitude) of the bird during the flights. This data.frame needs to contain \code{date} as
POSIXt and \code{value} in hPa. If not provided, uses \code{graph$pressure}, assuming that argument \code{graph}
is a GeoPressureR \code{tag} object.}
    \item{\code{rounding_interval}}{temporal resolution on which to query the variable (min). Default is to
match ERA5 native resolution (1hr).}
    \item{\code{interp_spatial_linear}}{logical to interpolate the variable linearly over space, if \code{FALSE}
takes the nearest neighbour. ERA5 native resolution is 0.25°}
    \item{\code{path_model}}{model used to interpolate the bird path during flight. \code{"gs"} assumes constant
groundspeed. \code{"as_2step"} first samples wind on the \code{"gs"} path, then approximately redistributes
along-track progress using a constant airspeed assumption.}
    \item{\code{quiet}}{logical to hide messages about the progress}
    \item{\code{file}}{absolute or relative path of the ERA5 wind data file to be downloaded. Function
taking as arguments (1) the stationary period identifier and (2) the tag_id.}
  }}
}
\value{
A \code{graph} object with windspeed and airspeed as \code{ws} and \code{as} respectively.
}
\description{
Reads the NetCDF files downloaded and interpolate the average windspeed experienced by the
bird on each possible edge, as well as the corresponding airspeed.

In addition, the graph can be further pruned based on a threshold of airspeed \code{thr_as}.

See section \href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.14082#mee314082-sec-0008-title}{2.2.4 in Nussbaumer (2023b)}
for more technical details and the \href{https://geopressure.org/GeoPressureManual/trajectory-with-wind.html}{GeoPressureManual} for an
illustration on how to use it.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    tag_set_map(extent = c(-16, 23, 0, 50), scale = 1) |>
    geopressure_map(quiet = TRUE)
})
graph <- graph_create(tag, quiet = TRUE)
graph <- graph_add_wind(graph, pressure = tag$pressure, quiet = TRUE)
\dontshow{\}) # examplesIf}
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\doi{10.1111/2041-210X.14082}.}
}
\seealso{
\href{https://geopressure.org/GeoPressureManual/trajectory-with-wind.html}{GeoPressureManual}

Other graph: 
\code{\link{graph_create}()},
\code{\link{graph_marginal}()},
\code{\link{graph_most_likely}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_simulation}()},
\code{\link{print.graph}()}
}
\concept{graph}
