% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assert.R
\name{graph_assert}
\alias{graph_assert}
\title{Assert the status of a \code{graph}}
\usage{
graph_assert(graph, condition = "graph")
}
\arguments{
\item{graph}{a GeoPressureR \code{graph} object}

\item{condition}{condition to assert \code{graph} for. One of \code{"graph"} (default), \code{"movement"},
or \code{"full"}.}
}
\value{
logical indicating whether the \code{graph} object fulfill the condition.
}
\description{
This function check the condition of a \code{graph} object.
}
