% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_marginal.R
\name{graph_marginal}
\alias{graph_marginal}
\title{Compute marginal probability map}
\usage{
graph_marginal(graph, quiet = FALSE)
}
\arguments{
\item{graph}{a GeoPressureR \code{graph} with defined movement model \code{graph_set_movement()}.}

\item{quiet}{logical to hide messages about the progress.}
}
\value{
A list of the marginal maps for each stationary period (even those not modelled). Best to
include within \code{tag}.
}
\description{
Compute the marginal probability map from a graph. The computation uses the \href{https://en.wikipedia.org/wiki/Forward\%E2\%80\%93backward_algorithm}{forward backward algorithm}. For more
details, see \href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.14082#mee314082-sec-0012-title}{section 2.3.2 of Nussbaumer et al. (2023b)}
and the \href{https://geopressure.org/GeoPressureManual/trajectory.html#product-2-marginal-probability-map}{GeoPressureManual}.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    twilight_create() |>
    twilight_label_read() |>
    tag_set_map(
      extent = c(-16, 23, 0, 50),
      known = data.frame(stap_id = 1, known_lon = 17.05, known_lat = 48.9)
    ) |>
    geopressure_map(quiet = TRUE) |>
    geolight_map(quiet = TRUE)
})

# Create graph
graph <- graph_create(tag, quiet = TRUE)

# Define movement model
graph <- graph_set_movement(graph)

# Compute marginal
marginal <- graph_marginal(graph)

plot(marginal)

}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\doi{10.1111/2041-210X.14082}.}
}
\seealso{
\href{https://geopressure.org/GeoPressureManual/trajectory.html#product-2-marginal-probability-map}{GeoPressureManual}

Other graph:
\code{\link[=graph_add_wind]{graph_add_wind()}},
\code{\link[=graph_create]{graph_create()}},
\code{\link[=graph_most_likely]{graph_most_likely()}},
\code{\link[=graph_set_movement]{graph_set_movement()}},
\code{\link[=graph_simulation]{graph_simulation()}},
\code{\link[=print.graph]{print.graph()}}
}
\concept{graph}
