% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_path2edge}
\alias{graph_path2edge}
\title{Find the edge index from a path index}
\usage{
graph_path2edge(path_id, grl)
}
\arguments{
\item{path_id}{list or matrix of node index nj x nsta}

\item{grl}{graph constructed with \code{geopressure_graph_create()}}
}
\value{
list or matrix of the edge nj x (nsta-1)
}
\description{
Very inefficient way to find the edges...
}
