% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_path2lonlat}
\alias{graph_path2lonlat}
\title{Find the lattitude and longitude from a path index}
\usage{
graph_path2lonlat(path_id, grl)
}
\arguments{
\item{path_id}{list or matrix of node index}

\item{grl}{graph constructed with \code{geopressure_graph_create()}}
}
\value{
list of the path with latitude and longitude and index fo the the path provided
}
\description{
Find the lattitude and longitude from a path index
}
