% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_trim}
\alias{graph_trim}
\title{Trim a graph}
\usage{
graph_trim(grl)
}
\arguments{
\item{grl}{Graph constructed with \code{\link[=graph_create]{graph_create()}}.}
}
\value{
Graph trimmed
}
\description{
Trimming consists in removing "dead branch" of a graph, that is removing the edges which are not
connected to both the source (i.e, equipment) or sink (i.e. retrieval site).
}
\seealso{
\code{\link[=graph_create]{graph_create()}}
}
