% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grl}
\alias{grl}
\title{Graph for 18LX}
\format{
An object of class \code{list} of length 13.
}
\usage{
grl
}
\description{
This dataset was generated with the vignette \href{articles/basic-graph.html}{Basic Graph}
}
\keyword{datasets}
