% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_create_distance.R
\name{haversine_distance}
\alias{haversine_distance}
\title{Calculate great-circle distances}
\usage{
haversine_distance(from_coords, to_coords)
}
\arguments{
\item{from_coords}{Matrix of source coordinates (longitude, latitude) in decimal degrees}

\item{to_coords}{Matrix of target coordinates (longitude, latitude) in decimal degrees}
}
\value{
A vector of distances in kilometers.
}
\description{
Vectorized implementation of the Haversine formula.
}
\keyword{internal}
