% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind2path.R
\name{ind2path}
\alias{ind2path}
\title{Create a path from indices of coordinate}
\usage{
ind2path(ind, geostap_graph)
}
\arguments{
\item{ind}{index in the 2D map (vector or matrix)}

\item{geostap_graph}{either a \code{geostap} or a \code{graph} which contains both \code{stap}, \code{extent} and
\code{scale}.}
}
\value{
A path data.frame
\itemize{
\item \code{stap_id} stationary period
\item \code{ind} indices of the coordinate in the 2D grid. Useful to retrieve map or graph information.
\item \code{lat} Latitude,
\item \code{lon} longitude
}
}
\description{
Convert the 2D coordinate index of a map and produce a path data.frame. This function is used by
\code{geostap2path}, \code{graph_most_likely}, \code{graph_simulation} and \code{geopressureviz}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |> tag_label()
geostap <- geostap_create(tag, c(-16, 23, 0, 50), scale = 1)

# ind generated by `graph_most_likely`
ind <- c(1652, 1603, 1755, 1708, 1607)
path <- ind2path(ind, geostap)
str(path)

# ind generated by `graph_simulation`
ind <- matrix(c(
  1652, 1652, 1652, 1652, 1652, 1652, 1652, 1652, 1652, 1652,
  1653, 1606, 1653, 1504, 1604, 1704, 1504, 1653, 1753, 1701,
  1805, 1457, 1408, 1657, 1609, 1903, 1506, 1757, 1856, 1804,
  1607, 1514, 1611, 1457, 1860, 1802, 1759, 1609, 1760, 1208,
  1505, 1314, 1559, 1357, 1758, 1852, 1661, 1706, 1862, 1358
), nrow = 10)
path <- ind2path(ind, geostap)
str(path)
}
