% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird.R
\name{is.bird}
\alias{is.bird}
\title{Check if an object is of class \code{bird}}
\usage{
\method{is}{bird}(x)
}
\arguments{
\item{x}{A \code{bird} object.}
}
\value{
\code{TRUE} for an object of class \code{bird}, otherwise \code{FALSE}.
}
\description{
Check if an object is of class \code{bird}
}
\seealso{
Other bird: 
\code{\link{bird_create}()},
\code{\link{print.bird}()}
}
\concept{bird}
