% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{is.graph}
\alias{is.graph}
\title{Check if an object is of class \code{graph}}
\usage{
is.graph(x)
}
\arguments{
\item{x}{A \code{graph} object.}
}
\value{
\code{TRUE} for an object of class \code{graph}, otherwise \code{FALSE}.
}
\description{
Check if an object is of class \code{graph}
}
