% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.tag.R
\name{is.tag}
\alias{is.tag}
\title{Check if an object is of class \code{tag}}
\usage{
\method{is}{tag}(x)
}
\arguments{
\item{x}{A \code{tag} object.}
}
\value{
\code{TRUE} for an object of class \code{tag}, otherwise \code{FALSE}.
}
\description{
Check if an object is of class \code{tag}
}
\seealso{
Other tag: 
\code{\link{plot.tag}()},
\code{\link{print.tag}()},
\code{\link{tag_create}()}
}
\concept{tag}
