% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light2mat.R
\name{light2mat}
\alias{light2mat}
\title{Convert light data in matrix format}
\usage{
light2mat(light, twl_offset = 0)
}
\arguments{
\item{twl_offset}{Shift of the middle of the night compared to 00:00 UTC (in seconds). If not
provided, it uses the middle of all nights.}
}
\value{
A dataframe with columns \code{date} and \code{value}.
}
\description{
Convert light data in matrix format
}
