% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{light_prob}
\alias{light_prob}
\title{Light probability map for 18LX}
\format{
An object of class \code{list} of length 29.
}
\usage{
light_prob
}
\description{
This dataset contains the data generated with the GeoPressure API generated
with the vignettes \href{articles/light-map.html}{Light Map}.
\code{light_prob} is a list of raster containing the probability map of each
stationary period.
}
\keyword{datasets}
