% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2path.R
\name{map2path}
\alias{map2path}
\title{Return the path from a likelihood map}
\usage{
map2path(tag, likelihood = NA, interp = -1, .use_known = TRUE)
}
\arguments{
\item{tag}{List of the geographical and stationary period information. See
\code{\link[=tag_create]{tag_create()}} for details.}

\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character). Default
\code{NA} is to take the product of \code{map_pressure} and \code{map_light}, or only \code{map_pressure} if
\code{map_light} is not available.}

\item{interp}{The position of the stationary period shorter than \code{interp} will be
replace by a linear average from other position (in days) .}

\item{.use_known}{If true, enforce the known position defined in \code{tag} in the path created. Known
position are not interpolated (even if shorter than \code{interp}) and used in the interpolation. In
most (all?) case, the likelihood map was computed using known, and therefore will result in the
same position (approx. to the map resolution).}
}
\value{
A path data.frame
\itemize{
\item \code{stap_id} stationary period
\item \code{ind} indices of the coordinate in the 2D grid. Useful to retrieve map or graph information.
\item \code{lat} Latitude,
\item \code{lon} longitude
}
}
\description{
Find the position of the highest value in a map, typically most probable value in a likelihood
map.
}
\details{
Note that this path is the most likely, considering only the observation model and ignoring the
movement model. Prefer to use \code{graph_most_likely()} for the most realistic path.

\code{interp} can be used to interpolate unrealistic position from short stationary period based on
the position of the longer ones. The interpolation assumes that the first and last stationary
period can be safely estimated from the probability map.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX") |>
  tag_label()
tag <- tag_create(tag,
  extent = c(-16, 23, 0, 50),
  scale = 2,
) |>
  geopressure_map(tag$pressure)

# Compute the path
path <- map2path(tag)

}
\seealso{
\code{\link[=geopressure_map_likelihood]{geopressure_map_likelihood()}}, \code{\link[=pressurepath_create]{pressurepath_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#compute-altitude}{GeoPressureManual | Pressure Map}
}
