% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_expand.R
\name{map_expand}
\alias{map_expand}
\title{Construct grid from \code{extent} and \code{scale}}
\usage{
map_expand(extent, scale)
}
\arguments{
\item{extent}{geographical extent of the map on which the likelihood and graph model will be
computed. Vector of length 4 \code{c(xmin, xmax, ymin, ymax)} or \code{c(W, E, S, N)}.}

\item{scale}{number of pixels per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and \code{scale=4} for a resolution of 0.25° (~30km). To avoid
interpolating the ERA5 single-levels data, the scale should be equal to or smaller than 4
(the default in \code{tag_set_map()}). \code{tag_set_map()} warns for higher values. Read more about
scale on the \href{https://developers.google.com/earth-engine/guides/scale}{Google earth Engine documentation }.}
}
\value{
A list containing:
\itemize{
\item \code{extent} same as input
\item \code{scale} same as input
\item \code{lat} vector of latitude
\item \code{lon} vector of longitude
\item \code{dim} vector of length 2 of the dimension of the map (number of pixel in lat and lon)
}
}
\description{
Take the two parameters defining the spatial grid of a map in GeoPressureR (\code{extent} and \code{scale})
and constructs various possible spatial variable of interest for the spatial grid (e.g., grid
dimension, latitude and longitude...)

This function is used in multiple functions of GeoPressureR to allow to only store \code{extent} and
\code{scale} in \code{tag} and \code{graph} while having access to all spatial variables anywhere anytime.
}
\examples{
str(map_expand(extent = c(0, 10, 0, 5), scale = 1))

str(map_expand(extent = c(-16, 23, 0, 50), scale = 10))
}
