% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_type.R
\name{map_type}
\alias{map_type}
\title{Get map type specifications}
\usage{
map_type()
}
\value{
A named list of map type specifications used by plotting and Shiny apps.
}
\description{
Get map type specifications
}
\keyword{internal}
