% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_classify}
\alias{pam_classify}
\title{Automatic classification of pam}
\usage{
pam_classify(pam, min_duration = 30)
}
\arguments{
\item{pam}{data list}

\item{min_duration}{duration in minutes}
}
\value{
pam
}
\description{
This function uses activity data to classify migratory flapping flight. This
fonction is inspired by the function \code{classify_flap} from
the \href{https://github.com/KiranLDA/pamLr}{pamLr package}.
}
