% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_read_check}
\alias{pam_read_check}
\title{Check that file exists and is unique}
\usage{
pam_read_check(pathname, file)
}
\arguments{
\item{pathname}{is the path where file is}

\item{file}{is the name of he file}
}
\description{
Check that file exists and is unique
}
