% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_read_file}
\alias{pam_read_file}
\title{Read pam file}
\usage{
pam_read_file(filename, crop_start, crop_end)
}
\arguments{
\item{filename}{is the path where files are stored}

\item{crop_start}{posicxt object for date that pam data should start}

\item{crop_end}{posicxt object for date that pam data should end}
}
\value{
a data.frame of the measurement
}
\description{
Read a specific pam file and return the corresponding data.frame.
}
\seealso{
\code{\link{pam_read}}
}
