% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_sta}
\alias{pam_sta}
\title{Compute stationary periods}
\usage{
pam_sta(pam)
}
\arguments{
\item{pam}{pam logger dataset list}
}
\value{
pam logger dataset list with a the dataframe of stationary periods
\code{pam$sta} as well as the new label named \code{sta_id} (\code{pam$pressure$sta_id} and
\code{pam$acceleration$sta_id})
}
\description{
This function computes the table of stationary periods from the class of
acceleration \code{pam$acceleration$class} and add it to the pam
}
