% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path2df}
\alias{path2df}
\title{Path to data.frame}
\usage{
path2df(tag, path)
}
\arguments{
\item{tag}{data logger dataset list with \code{tag$sta} computed (see \code{tag_sta})}

\item{path}{data.frame containtings the path(s) of the bird with column \code{lat}, \code{lon} and \code{stap}
at least. Path can be generated with \code{map2path}, \code{graph_simulation}, \code{geopressureviz}
.}
}
\description{
This function convert a GeoPressureR path to a data.frame which can be read in \href{https://movevis.org/index.html}{movevis} for instance.
}
\details{
The function basically duplicate location position at the start and end time of each stationary
period.
}
\examples{
tag <- tag_read(
  pathname = system.file("extdata/0_tag/18LX", package = "GeoPressureR")
)
tag <- trainset_read(tag,
  pathname = system.file("extdata/1_pressure/labels", package = "GeoPressureR")
)
tag <- tag_stap(tag)
pressure_likelihood_1 <- readRDS(system.file("extdata/1_pressure/", "18LX_pressure_likelihood_1.rda",
  package = "GeoPressureR"
))
path <- map2path(list(pressure_likelihood_1))
path2df(tag, path)
}
\seealso{
\href{https://movevis.org/reference/df2move.html}{\code{movevis::df2move}},
\code{\link{map2path}}, \code{\link{graph_simulation}}, \code{\link{geopressureviz}}
}
