% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2edge.R
\name{path2edge}
\alias{path2edge}
\title{Extract the edges of a \code{path} from a \code{graph}}
\usage{
path2edge(path, tag_graph)
}
\arguments{
\item{path}{a GeoPressureR \code{path} data.frame}

\item{tag_graph}{either a \code{tag} or a \code{graph} GeoPressureR object.}
}
\value{
Data.frame of the edge containing:
\itemize{
\item \code{s}: index in 3D (lat-lon-stap) of the origin (source).
\item \code{t}: index in 3D (lat-lon-stap) of the destination (target).
\item \code{lat_s}: latitude of the origin (source).
\item \code{lat_t}: latitude of the destination (target).
\item \code{lon_s}: longitude of the origin (source).
\item \code{lon_t}: longitude of the destination (target).
\item \code{stap_s}: stationary period of the origin (source).
\item \code{stap_t}: stationary period of the destination (target).
\item \code{distance}: Distance (in km) of the flight.
\item \code{start}: end of the flight.
\item \code{end}: start of the flight.
\item \code{duration}: duration of the flight.
\item \code{n}: number of flight.
\item \code{bearing}: Angle of the flight in degree (0° = North, 90° = East).
\item \code{gs}: groundspeed vector expressed as a complex number. You can compute the groundspeed value
(km/h) with \code{abs(gs)}, the W-E and S-N component of the flight with \code{Re(gs)} and \code{Im(gs)}, and
the angle/direction with \code{Arg(gs)}. If graph provided.
\item \code{ws}: if computed with \code{graph_add_wind()}, same value as \code{gs}. Airspeed is computed with
\code{as = gs - ws} in complex number to keep the vectorial additive properties. If graph provided.
}
}
\description{
Retrieve the edges in a \code{graph} corresponding to the flight transition defined by a \code{path}. These
edges can be useful to extract flight information specific to a path.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
  edge <- path2edge(path, tag)
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://geopressure.org/GeoPressureManual/trajectory-with-wind.html#extract-flight-information}{GeoPressureManual}

Other path:
\code{\link[=path2elevation]{path2elevation()}},
\code{\link[=path2twilight]{path2twilight()}},
\code{\link[=plot_path]{plot_path()}},
\code{\link[=tag2path]{tag2path()}}
}
\concept{path}
