% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{plot_path}
\alias{plot_path}
\title{Plot a \code{path}}
\usage{
plot_path(
  path,
  plot_leaflet = TRUE,
  map = NULL,
  provider = NULL,
  provider_options = leaflet::providerTileOptions(),
  pad = 3,
  polyline = NULL,
  circle = NULL,
  group = NULL
)
}
\arguments{
\item{path}{a GeoPressureR \code{path} data.frame.}

\item{plot_leaflet}{logical defining if the plot is an interactive \code{leaflet} map or a static
basic plot.}

\item{map}{optional \code{map} object to plot the path on top of.}

\item{provider}{optional tile provider name (see \code{leaflet::providers}). When \code{NULL}, uses all
available GeoPressureR basemap layers; see the \emph{Basemap API keys} section.}

\item{provider_options}{tile options. See leaflet::addProviderTiles() and
leaflet::providerTileOptions().}

\item{pad}{padding of the map in degree lat-lon (only for \code{plot_leaflet = FALSE}).}

\item{polyline}{list of parameters passed to \code{leaflet::addPolylines()}}

\item{circle}{list of parameters passed to \code{leaflet::addCircleMarkers()}}

\item{group}{optional Leaflet overlay group name. When supplied, all
trajectories in \code{path} can be shown or hidden together while retaining
separate line geometries for each \code{j}.}
}
\value{
A \code{leaflet} map when \code{plot_leaflet = TRUE}, otherwise a \code{ggplot2} object.
}
\description{
This function plots a \code{path} data.frame. This function is used in \code{\link[=plot.map]{plot.map()}}.
}
\section{Basemap API keys}{

Interactive maps always offer Esri basemaps (gray canvas, imagery and topographic). CARTO,
Mapbox and MapTiler basemaps are added to the layer control only when their API key is stored
in your system keyring. Get a key from each provider:
\itemize{
\item CARTO: "Get your API key" on \href{https://carto.com/basemaps}{CARTO Basemaps}.
\item Mapbox: \href{https://account.mapbox.com/access-tokens/}{Access tokens} page of your account
(see \href{https://docs.mapbox.com/help/getting-started/access-tokens/}{Mapbox documentation}).
\item MapTiler: \href{https://cloud.maptiler.com/account/keys/}{API keys} page of your MapTiler Cloud
account (see \href{https://docs.maptiler.com/cloud/api/authentication-key/}{MapTiler documentation}).
}

Then store each key once:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{keyring::key_set_with_value("CARTO_API_KEY", password = "<your CARTO API key>")
keyring::key_set_with_value("MAPBOX_ACCESS_TOKEN", password = "<your Mapbox token>")
keyring::key_set_with_value("MAPTILER_API_KEY", password = "<your MapTiler API key>")
}\if{html}{\out{</div>}}

Mapbox access tokens are visible in the browser when a map is rendered. Use a public, read-only
token rather than a secret token.
}

\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE)
})

path <- data.frame(
  stap_id = 1:4,
  lon = c(17.05, 16.2, NA, 15.4),
  lat = c(48.9, 47.8, NA, 46.5),
  known_lon = c(17.05, NA, NA, NA),
  known_lat = c(48.9, NA, NA, NA),
  start = as.POSIXct(
    c("2018-01-01", "2018-01-03", "2018-01-05", "2018-01-07"),
    tz = "UTC"
  ),
  end = as.POSIXct(
    c("2018-01-02", "2018-01-04", "2018-01-06", "2018-01-08"),
    tz = "UTC"
  )
)

plot_path(path)

plot_path(path, plot_leaflet = FALSE)

}
\seealso{
\code{\link[=plot.map]{plot.map()}}

Other path:
\code{\link[=path2edge]{path2edge()}},
\code{\link[=path2elevation]{path2elevation()}},
\code{\link[=path2twilight]{path2twilight()}},
\code{\link[=tag2path]{tag2path()}}
}
\concept{path}
