% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pressure_path.R
\name{plot_pressure_path}
\alias{plot_pressure_path}
\title{Plot pressure path_pres}
\usage{
plot_pressure_path(
  path_pres,
  type = "ts",
  pressure_err_warning = 2,
  pressure_err_danger = 5
)
}
\arguments{
\item{path_pres}{rerer}

\item{type}{\code{"ts"} or \code{"hist"}}

\item{pressure_err_warning}{Threshold of pressure difference marking as ️warning (hPa)}

\item{pressure_err_danger}{Threshold of pressure difference marking as ️danger (hPa)}
}
\description{
description of check
}
