% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pressure_tag.R
\name{plot_pressure_tag}
\alias{plot_pressure_tag}
\title{Check pressure data}
\usage{
plot_pressure_tag(
  tag,
  known = data.frame(stap_id = integer(), known_lat = double(), known_lon = double()),
  stap_include = tag$stap$stap_id,
  pressure_diff_danger = 5,
  pressure_diff_warning = 3,
  stap_length_danger = 6,
  stap_length_warning = 12
)
}
\arguments{
\item{tag}{Data logger list with label information. See \code{\link[=tag_label]{tag_label()}} for the required input.}

\item{known}{Data.frame containing the known positions of the bird (e.g., equipment or retrieval
site). This information can only be attached at the level of a stationary period.}

\item{stap_include}{Vector of the stationary period to model, that is, to compute in the likelihood
map and use in the graph.}

\item{pressure_diff_danger}{Threshold of pressure hourly difference marking as ️danger (hPa)}

\item{pressure_diff_warning}{Threshold of pressure hourly difference marking as ️warning (hPa)}

\item{stap_length_danger}{Threshold number of pressure datapoints flagged as ️danger (hourly).}

\item{stap_length_warning}{Threshold number of pressure datapoints flagged as ️warning (hourly).}
}
\value{
Return \code{TRUE} if no danger were raised (warning is ok).
}
\description{
This function perform the following basic check on the pressure data and its suitability for
\code{\link[=geopressure_map]{geopressure_map()}}:
\itemize{
\item Is there sufficient data for each stationary periods?
\item Is the hourly change of pressure within each stationary period possible for a natural variation
of pressure or does it suggest vertical movement of the bird (flight or elevation change)?
}

Beyond the hood, it uses \code{\link[=geopressure_map_preprocess]{geopressure_map_preprocess()}} to compute the cleaned
pressure to be match and perform the check on those (an not the raw pressure data).

Use \code{known} and a \code{stap_include} to only check the pressure belonging to a stationary period
which is included and not known.
}
