% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_light}
\alias{plot_tag_light}
\title{Plot light data of a \code{tag}}
\usage{
plot_tag_light(tag, transform_light = TRUE, plot_plotly = TRUE)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{transform_light}{logical to display a log transformation of light}

\item{plot_plotly}{logical to use \code{plotly}}
}
\value{
a plot or ggplotly object.
}
\description{
This function display a plot of light time series recorded by a tag
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)
})

plot_tag_light(tag)

}
\seealso{
Other plot_tag:
\code{\link[=plot_tag_acceleration]{plot_tag_acceleration()}},
\code{\link[=plot_tag_actogram]{plot_tag_actogram()}},
\code{\link[=plot_tag_pressure]{plot_tag_pressure()}},
\code{\link[=plot_tag_temperature]{plot_tag_temperature()}},
\code{\link[=plot_tag_twilight]{plot_tag_twilight()}}
}
\concept{plot_tag}
