% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_map}
\alias{plot_tag_map}
\title{Plot twilight data of a \code{tag}}
\usage{
plot_tag_map(
  tag,
  likelihood = NA,
  plot_leaflet = TRUE,
  provider = "Stamen.TerrainBackground",
  palette = "OrRd",
  opacity = 0.8,
  legend = FALSE,
  ...
)
}
\arguments{
\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character). Default
\code{NA} is to take the product of \code{map_pressure} and \code{map_light}, or only \code{map_pressure} if
\code{map_light} is not available.}

\item{plot_leaflet}{Logical defining if the plot is an interactive \code{leaflet} map or a static
\code{terra::plot} map.}

\item{provider}{the name of the provider (see
\url{https://leaflet-extras.github.io/leaflet-providers/preview/} and
\url{https://github.com/leaflet-extras/leaflet-providers})}

\item{palette}{The colors or color function that values will be mapped to}

\item{opacity}{the base opacity of the raster, expressed from 0 to 1}

\item{legend}{logical or character. If not \code{FALSE} a legend is drawn. The character value can be used to indicate where the legend is to be drawn. For example "topright" or "bottomleft". Use \code{plg} for more refined placement. Not supported for continuous legends (the default for raster data)}

\item{...}{
  Arguments passed on to \code{\link[terra:plot]{terra::plot}}
  \describe{
    \item{\code{x}}{SpatRaster or SpatVector}
    \item{\code{y}}{missing or positive integer or name indicating the layer(s) to be plotted}
  }}
}
\description{
This function display a plot of twilight timeseries recorded by a tag
}
