% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_pressure}
\alias{plot_tag_pressure}
\title{Plot pressure data of a \code{tag}}
\usage{
plot_tag_pressure(tag, plot_plotly = TRUE, warning_pressure_diff = 3)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{plot_plotly}{logical to use \code{plotly}.}

\item{warning_pressure_diff}{Threshold of pressure hourly difference marking as warning (hPa).}
}
\value{
a plot or ggplotly object.
}
\description{
This function display a plot of pressure time series recorded by a tag
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)
})

plot_tag_pressure(tag, plot_plotly = FALSE)

withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_label(tag, quiet = TRUE)
})

plot_tag_pressure(tag)
}
\seealso{
Other plot_tag:
\code{\link[=plot_tag_acceleration]{plot_tag_acceleration()}},
\code{\link[=plot_tag_actogram]{plot_tag_actogram()}},
\code{\link[=plot_tag_light]{plot_tag_light()}},
\code{\link[=plot_tag_temperature]{plot_tag_temperature()}},
\code{\link[=plot_tag_twilight]{plot_tag_twilight()}}
}
\concept{plot_tag}
