% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pressure_maps}
\alias{pressure_maps}
\title{Pressure mismatch and threshold map for 18LX}
\format{
An object of class \code{list} of length 29.
}
\usage{
pressure_maps
}
\description{
This dataset contains the data generated with the GeoPressure API generated
with the vignettes \href{articles/pressure-map.html}{Pressure Map}.
\code{pressure_maps} is a list of brick containing the mismatch and threshold
map of each stationary period.
}
\keyword{datasets}
