% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pressure_prob}
\alias{pressure_prob}
\title{Pressure probability map for 18LX}
\format{
An object of class \code{list} of length 27.
}
\usage{
pressure_prob
}
\description{
This dataset contains the data generated with the GeoPressure API generated
with the vignettes \href{articles/pressure-map.html}{Pressure Map}.
\code{pressure_prob} is a list of raster containing the probability map of each
stationary period.
}
\keyword{datasets}
