% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pressure_timeserie}
\alias{pressure_timeserie}
\title{Pressure timeseries for 18LX}
\format{
An object of class \code{list} of length 27.
}
\usage{
pressure_timeserie
}
\description{
This dataset contains the data generated with the GeoPressure API generated
with the vignettes \href{articles/pressure-map.html}{Pressure Map}.
\code{pressure_timeserie} is a list for each stationary period
}
\keyword{datasets}
