% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressurepath2altitude.R
\name{pressurepath2altitude}
\alias{pressurepath2altitude}
\title{Extract altitude from a \code{pressurepath}}
\usage{
pressurepath2altitude(pressurepath)
}
\arguments{
\item{pressurepath}{a GeoPressureR \code{pressurepath} data.frame}
}
\value{
A data.frame with columns:
\itemize{
\item \code{date} same as \code{pressure$date}
\item \code{altitude}
}
}
\description{
This function compute the timeseries of altitude from a pressurepath.

As the altitude is estimated both at the arrival and departure stationary period location, we
need to find a way to correct for the altitude in between.
}
\examples{
owd <- setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)
setwd(owd)

path <- data.frame(
  stap_id = tag$stap$stap_id,
  lat = c(48.5, 32.5, 30.5, 49.5, 41.6),
  lon = c(17.5, 13.5, 16.5, 21.5, 12.7)
)

pressurepath <- pressurepath_create(tag, path = path, quiet = TRUE)

altitude <- pressurepath2altitude(pressurepath)

plot(altitude$date, altitude$altitude, type = "l")

knitr::kable(head(altitude))

}
\seealso{
Other pressurepath: 
\code{\link{geopressure_timeseries}()},
\code{\link{plot_pressurepath}()},
\code{\link{pressurepath_create}()},
\code{\link{pressurepath_update}()}
}
\concept{pressurepath}
