% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressurepath_update.R
\name{pressurepath_update}
\alias{pressurepath_update}
\title{Update tag label and recompute tag}
\usage{
pressurepath_update(pressurepath, tag_new)
}
\arguments{
\item{path_pres}{pressurepath_create}
}
\value{
a list containing the new \code{tag} and \code{path_pres}.
}
\description{
When updating the labelling file of a tag, often, only a few stationary periods are changing. To
avoid recomputing the entire workflow, this function figure out which stationary period have
been changing on only update those in \code{tag$map_pressure} and \code{path_pres}.
}
\examples{
tag <- tag_label(tag)

tag <- tag |>
  tag_create(extent = c(-16, 23, 0, 50), scale = 1) |>
  geopressure_map(tag$pressure)

path_pres <- tag |>
  map2path() |>
  pressurepath_create(tag$pressure)

update <- tag_update(
  file = "data/1-tag_label/18LX-labeled-modif.csv",
  tag,
  tag,
  path_pres
)

list2env(update, env = .GlobalEnv)
}
