% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_stap.R
\name{pretty_dt}
\alias{pretty_dt}
\title{Format difftime in d, h, m and s}
\usage{
pretty_dt(tim)
}
\arguments{
\item{tim}{character string or numeric value specifying a time interval.
noRd}
}
\description{
Format difftime in d, h, m and s
}
