% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.param.R
\name{print.param}
\alias{print.param}
\title{Print a \code{param} list}
\usage{
\method{print}{param}(x, ...)
}
\arguments{
\item{x}{a GeoPressureR \code{param} list.}

\item{...}{arguments passed to other methods}
}
\value{
\code{param} is returned invisibly and unchanged
}
\description{
This function displays the information of a \code{param} list.
}
\examples{
# Display the default parameters used in the package.
param <- param_create("18LX", default = TRUE)

print(param)

}
\seealso{
Other param:
\code{\link[=param_create]{param_create()}}
}
\concept{param}
