% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prob_map_list}
\alias{prob_map_list}
\title{Probability map for 18LX}
\format{
An object of class \code{list} of length 27.
}
\usage{
prob_map_list
}
\description{
This dataset contains the data generated with the GeoPressure API for the
vignettes
}
\details{
\code{prob_map_list} was created with
}
\examples{
\dontrun{
prob_map_list = geopressure_prob_map(raster_list)
# Save the data for vignette
usethis::use_data(prob_map_list, overwrite=T)
}
}
\keyword{datasets}
