% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast.map.R
\name{rast.map}
\alias{rast.map}
\title{Construct a SpatRaster from a \code{map}}
\usage{
rast.map(x, names = glue::glue("#{map$stap$stap_id}"), crs = "epsg:4326", ...)
}
\arguments{
\item{x}{a GeoPressureR \code{map} object}

\item{names}{names of the SpatRaster layers created. See \code{\link[terra:names]{terra::names}}.}

\item{crs}{character. Description of the Coordinate Reference System (map projection) in \code{PROJ.4}, \code{WKT} or \code{authority:code} notation. See \code{\link[terra]{crs}}. If this argument is missing, and the x coordinates are within -360 .. 360 and the y coordinates are within -90 .. 90, longitude/latitude is assigned}

\item{...}{additional parameters for \code{terra::rast}}
}
\value{
A \link[terra:SpatRaster]{terra::SpatRaster} object.
}
\description{
This function convert a GeoPressureR \code{map} object into a \code{\link[terra:SpatRaster]{terra::SpatRaster}} with the data of
each stationary periods stored in a different layer.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    tag_set_map(
      extent = c(-16, 23, 0, 50),
      scale = 4
    ) |>
    geopressure_map(quiet = TRUE)
})

rast.map(tag$map_pressure)

}
\seealso{
Other map:
\code{\link[=map_add_mask_water]{map_add_mask_water()}},
\code{\link[=map_create]{map_create()}},
\code{\link[=print.map]{print.map()}}
}
\concept{map}
