% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raster_light_list}
\alias{raster_light_list}
\title{Probability map of light for 18LX}
\format{
An object of class \code{list} of length 29.
}
\usage{
raster_light_list
}
\description{
This dataset was generated with the vignette \verb{Light-based geopositiong}
}
\keyword{datasets}
