% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raster_list}
\alias{raster_list}
\title{Mismatch map for 18LX}
\format{
An object of class \code{list} of length 27.
}
\usage{
raster_list
}
\description{
This dataset contains the data generated with the GeoPressure API for the
vignettes
}
\details{
\code{raster_list} was created with
\dontrun{
pam_data = pam_read(system.file("extdata", package = "GeoPressureR"),
crop_start = "2017-06-20", crop_end = "2018-05-02")
pam_data = trainset_read(pam_data, system.file("extdata", package = "GeoPressureR"))
pam_data = pam_sta(pam_data)

sta_id_keep = pam_data$sta$sta_id[difftime(pam_data$sta$end,pam_data$sta$start, units = "hours")>12]
pam_data$pressure$sta_id[!(pam_data$pressure$sta_id %in% sta_id_keep)] = NA

raster_list = geopressure_map(pam_data$pressure, extent=c(-16,20,0,50), scale=10, max_sample=100)

# Save the data for vignette
usethis::use_data(raster_list, overwrite=T)
}
}
\keyword{datasets}
