% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{read_pam}
\alias{read_pam}
\title{Read pam data}
\usage{
read_pam(
  pathname = pathname,
  extension = c("pressure", "glf", "acceleration", "temperature", "magnetic")
)
}
\arguments{
\item{pathname}{path where files are stored}

\item{extension}{list of file extentions to read (e.g., ".pressure",
".glf", ".gle",".acceleration", ".temperature" and ".magnetic")}
}
\value{
a list of all measurements
}
\description{
Imports multi-sensor logger data.
}
