% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{read_pam_file}
\alias{read_pam_file}
\title{Read pam file}
\usage{
read_pam_file(filename)
}
\arguments{
\item{filename}{is the path where files are stored}
}
\value{
a data.frame of the measurement
}
\description{
Read any pam file and return the corresponding data.frame.
}
