% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_path_diagnostic.R
\name{sampling_path_diagnostic}
\alias{sampling_path_diagnostic}
\title{Diagnostics for Path sampling output}
\usage{
sampling_path_diagnostic(paths, quiet = FALSE)
}
\arguments{
\item{paths}{data.frame returned by \code{sampling_path()}.}

\item{quiet}{logical to hide printed diagnostics.}
}
\value{
A data.frame with columns \code{stap_id}, \code{rhat_lat}, and \code{rhat_lon}.
}
\description{
Compute basic diagnostics for paths produced by \code{sampling_path()}. Currently supports
the Gelman-Rubin R-hat statistic for latitude and longitude per stationary period.
}
\seealso{
Other sampling_path: 
\code{\link{sampling_path}()}
}
\concept{sampling_path}
