% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed2bearing.R
\name{speed2bearing}
\alias{speed2bearing}
\title{Compute the bearing of a speed vector}
\usage{
speed2bearing(
  speed,
  speed_ref = complex(real = 0, imaginary = 1),
  positive = TRUE
)
}
\arguments{
\item{speed}{speed as complex value}

\item{speed_ref}{reference vector of the angle. Default is the North in order to return bearing.
Use \code{speed_ref = 1} for trigonometric orientation (clockwise from East).}

\item{positive}{logical to ensure the bearing is positive between 0-360 degree.}
}
\value{
bearing angle in degree
}
\description{
This function convert a speed vector represented in complex number into a bearing angle (0°=N ;
90°=E...). This conversion is needed because trigonometric angle are different than bearing.
}
