% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap2duration.R
\name{stap2duration}
\alias{stap2duration}
\title{Compute duration of stationary periods}
\usage{
stap2duration(stap_path, units = "days", return_numeric = TRUE)
}
\arguments{
\item{stap_path}{A \code{stap} data.frame.}

\item{units}{character string. Units in which the results are desired. Can be abbreviated.
See \code{\link{difftime}}}

\item{return_numeric}{logical to return the duration as a numeric rather than with a duration
format.}
}
\value{
vector of duration
}
\description{
This function returns the duration between \code{stap$start} and \code{stap$end}.

The function can be used with any data.frame containing \code{start} and \code{end} as POSIXct (e.g.,
\code{flight} or \code{path}).
}
\examples{
# Create fake stap
stap <- data.frame(
  start = seq(as.POSIXct("1990-01-01"), as.POSIXct("1991-01-01"), length.out = 13)
)
stap$end <- stap$start + 60 * 60 * 24 * 30
stap$stap_id <- seq_len(nrow(stap))

stap2duration(stap)

stap2duration(stap, units = "mins", return_numeric = FALSE)

}
