% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stap2flight.R
\name{stap2flight}
\alias{stap2flight}
\title{Flight from stationary period}
\usage{
stap2flight(stap, stap_include = NA, format = "df", units = "hours")
}
\arguments{
\item{stap}{Stationary period structure (see \code{\link[=tag_label_stap]{tag_label_stap()}}).}

\item{stap_include}{Vecto of the stationary period \code{stap_id} to consider in the flight. Default
is to use \code{stap$stap_id[stap$include]} or \code{stap$stap_id} if \code{model} is not available in \code{stap}.}

\item{format}{Character to return a list \code{"list"} or a data.frame \code{"df"} (see description)}

\item{units}{Units in which the results are desired. see \code{\link[=difftime]{difftime()}}}
}
\value{
A list or a data.frame (see description) containing
\itemize{
\item \code{start}: Start time of the (first) flight
\item \code{end}: End time of the (last) flight
\item \code{stap_id_s}: Source stap_id (i.e, start)
\item \code{stap_id_t}: Target stap_id (i.e, end)
\item \code{duration}: (Sum of the) duration of flight(s)
\item (\code{n}: Numer of flights)
The value in brackets are only for the data.frame
}
}
\description{
Convert a stationary period data.frame \code{stap} into a flight data.frame or list. Flight are
computed as the difference between the end of a stationary period to the start of the next one.

You can compute the flight between specific stationary periods using \code{stap_include}. In this
case, the flight duration is computed as the sum of individual flights in between.

You can return the flight as a data.frame or as a list if you want to retrieve the information
of all individual flight between the \code{stap_include}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |>
  tag_label()

str(stap2flight(tag$stap))

str(stap2flight(tag$stap, stap_include = c(1, 3, 5)))

str(stap2flight(tag$stap, stap_include = c(2, 3, 5), format = "list", units = "secs"))

}
