% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{static_prob}
\alias{static_prob}
\title{Static probability map for 18LX}
\format{
An object of class \code{list} of length 29.
}
\usage{
static_prob
}
\description{
This dataset was generated with the vignette \href{articles/prepare-data.html}{Prepare Data}.
\code{static_prob} is a list of raster containing the probability map of each stationary period.
}
\keyword{datasets}
