% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{static_timeserie}
\alias{static_timeserie}
\title{Static timeseries for 18LX}
\format{
An object of class \code{list} of length 29.
}
\usage{
static_timeserie
}
\description{
This dataset contains the data generated with the GeoPressure API generated
with the vignettes \href{articles/prepare-data.html}{Prepare Data}.
\code{static_timeserie} is a list for each stationary period
}
\keyword{datasets}
