% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2map.R
\name{tag2likelihood}
\alias{tag2likelihood}
\title{Return a valid likelihood map name}
\usage{
tag2likelihood(tag, likelihood = NULL)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{likelihood}{field of the \code{tag} list containing the likelihood map(s), or a map type name.}
}
\value{
Likelihood map name
}
\description{
Return a valid likelihood map name
}
\keyword{internal}
