% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2map.R
\name{tag2map}
\alias{tag2map}
\title{Extract a \code{map} from a \code{tag}}
\usage{
tag2map(tag, likelihood = NULL)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character). Possible
value are \code{map_pressure}, \code{map_light}, \code{map_pressure_mse}, \code{map_pressure_mse},
\code{map_pressure_mse}, \code{mask_water}. Default \code{NA} is to take the product of \code{map_pressure} and
\code{map_light}, or if not available, taking the first of the possible values.}
}
\value{
likelihood map
}
\description{
Extract a \code{map} from a \code{tag}
}
