% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2path.R
\name{tag2path}
\alias{tag2path}
\title{Build a \code{path} from the likelihood maps of a \code{tag}}
\usage{
tag2path(tag, likelihood = NULL, interp = FALSE, use_known = TRUE)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character). By default,
uses the product of \code{map_pressure} and \code{map_light} if available, otherwise the first available
likelihood map in the tag.}

\item{interp}{the position of the stationary period shorter than \code{interp} will be replace by a
linear average from other position accounting for flight duration (in days).}

\item{use_known}{If \code{TRUE}, enforce known positions from \code{tag$stap} when building the path. Known
positions are approximated to the map resolution so they correspond to integer indices.}
}
\value{
A path data.frame
\itemize{
\item \code{stap_id} stationary period
\item \code{ind} indices of the coordinate in the 2D grid. Useful to retrieve map or graph information.
\item \code{lat} Latitude,
\item \code{lon} longitude
}
}
\description{
Find the position of the highest value in a map, typically most probable value in a likelihood
map.

Note that this path is the most likely, considering only the observation model and ignoring the
movement model. Prefer to use \code{graph_most_likely()} for the most realistic path accounting for
flight duration.

\code{interp} can be used to interpolate unrealistic position from short stationary periods based on
the position of the longer ones. In order to preserve a reference to the grid, the interpolation
is only performed at the centre of the grid cell (defined by the \code{likelihood} map). Note also
that it is not possible to interpolate the first and last stationary period.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    tag_set_map(
      extent = c(-16, 23, 0, 50),
      scale = 2
    ) |>
    geopressure_map(quiet = TRUE)
})

# Extract a path from pressure map
path <- tag2path(tag)
plot_path(path)

# Short stationary periods (e.g. 1 day) can be unreliably
# estimated, so interpolating them is often better
path <- tag2path(tag, interp = 1)
plot_path(path)

}
\seealso{
Other path:
\code{\link[=path2edge]{path2edge()}},
\code{\link[=path2elevation]{path2elevation()}},
\code{\link[=path2twilight]{path2twilight()}},
\code{\link[=plot_path]{plot_path()}}
}
\concept{path}
