% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2rast.R
\name{tag2rast}
\alias{tag2rast}
\title{Return a \code{terra:rast} for the likelihood map of \code{tag}}
\usage{
tag2rast(
  tag,
  time = tag$stap$start,
  names = tag$stap$stap_id,
  likelihood = NA,
  ...
)
}
\arguments{
\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character). Default
\code{NA} is to take the product of \code{map_pressure} and \code{map_light}, or only \code{map_pressure} if
\code{map_light} is not available.}

\item{...}{
  Arguments passed on to \code{\link[terra:rast]{terra::rast}}
  \describe{
    \item{\code{x}}{filename (character), missing, SpatRaster, SpatRasterDataset, SpatExtent, SpatVector, matrix, array, list of SpatRaster objects. For other types it will be attempted to create a SpatRaster via (`as(x, "SpatRaster")`}
  }}
}
\description{
This function display a plot of twilight timeseries recorded by a tag
}
