% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_assert.R
\name{tag_assert}
\alias{tag_assert}
\title{Assert the status of a \code{tag}}
\usage{
tag_assert(tag, condition = "tag", type = "abort")
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{condition}{condition to assert \code{tag} for. One of \code{"tag"} (default), \code{"read"},
\code{"pressure"}, \code{"light"}, \code{"acceleration"}, \code{"temperature_internal"},
\code{"temperature_external"}, \code{"magnetic"}, \code{"label"}, \code{"stap"}, \code{"setmap"},
\code{"map_pressure"}, \code{"map_pressure_mse"}, \code{"map_pressure_mask"}, \code{"map_light"},
\code{"map_light_twl"}, \code{"mask_water"}, \code{"map_magnetic"}, \code{"map_magnetic_intensity"},
\code{"map_magnetic_inclination"}, \code{"twl_calib"}, and \code{"twilight"}.}

\item{type}{Message type to display. One of \code{"abort"} (default), \code{"warn"}, \code{"inform"},
or \code{"logical"} (return TRUE/FALSE without messaging).}
}
\value{
logical indicating the \code{tag} object has the relevant element
}
\description{
This function check the condition of a \code{tag} object.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)
})

tag_assert(tag)

tag_assert(tag, "stap")

tag_assert(tag, "setmap", type = "warn")

tag_assert(tag, "map_pressure", type = "inform")

tag_assert(tag, "setmap", type = "logical")

}
