% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag_classify}
\alias{tag_classify}
\title{Automatic classification of tag}
\usage{
tag_classify(tag, min_duration = 30)
}
\arguments{
\item{min_duration}{duration in minutes}

\item{data}{logger dataset list. See \code{\link[=tag_read]{tag_read()}}.}
}
\value{
tag
}
\description{
This function uses activity data to classify migratory flapping flight. It returns the same data
list \code{tag} adding a column \code{ismig} to the data.frame \code{acceleration}. This function is inspired by
the function \code{classify_flap} from the \href{https://github.com/KiranLDA/taglr}{tagLr package}.
}
\examples{
tag <- tag_read(
  pathname = system.file("extdata/0_tag/18LX", package = "GeoPressureR")
)
tag <- tag_classify(tag, min_duration = 15)
head(tag$acceleration)
}
\seealso{
\code{\link[=tag_read]{tag_read()}}, \href{https://kiranlda.github.io/tagLrManual/flapping.html}{flapping chapter of the tagLr manual}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#automatic-classification-of-activity}{GeoPressureManual | Pressure Map }
}
