% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_geostap.R
\name{tag_geostap}
\alias{tag_geostap}
\title{Define grid for the trajectory}
\usage{
tag_geostap(
  tag,
  extent,
  scale = 10,
  include_stap_id = NA,
  include_min_duration = NA,
  known = data.frame(stap_id = integer(), known_lat = double(), known_lon = double())
)
}
\arguments{
\item{tag}{Data logger list with label information. See \code{\link[=tag_label]{tag_label()}} for the required input.}

\item{extent}{Geographical extent of the map on which the likelihood and graph model will be
computed. Vector of length 4 \code{c(xmin, xmax, ymin, ymax)} or \code{c(W, E, S, N)}.}

\item{scale}{Number of pixels per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and \code{scale=4} for a resolution of 0.25° (~30km). To avoid
interpolating the ERA5 data, the scale should be smaller than 10. Read more about scale on the \href{https://developers.google.com/earth-engine/guides/scale}{Google earth Engine documentation}.}

\item{include_stap_id}{Vector of \code{stap_id} defining which stationary period to model, that is,
to compute in the likelihood map and use in the graph.}

\item{include_min_duration}{Numeric defining the minimum threshold of stationary periods duration
(in hours) to includes.}

\item{known}{Data.frame containing the known positions of the bird (e.g., equipment or retrieval
site).}
}
\value{
A \code{tag} object with:
\itemize{
\item \code{stap}: Data.frame of all stationary periods with three new columns: \code{known_lat} and
\code{known_lon} define the known position during these stationary periods, and \code{model} defines
whether the likelihood map of this stationary period should be computed and later used in the
graph.
\item \code{extent} same as input parameter \code{extent}
\item \code{scale} same as input parameter \code{scale}
}
}
\description{
This function adds the parameters defining the 3D grid of the maps. The spatial parameters
(\code{extent} and \code{scale}) defines the GEOgraphical dimension. The temporal dimension is
defined based on the stationary periods build from the label. \code{include_stap_id} and
\code{include_min_duration} can be used to limit which stationary periods are computed and model in
the rest of the analysis. By default, all stationary periods are included.

In addition, \code{tag} also includes the ability to define \code{known} locations (e.g., equipment or
retrieval site). These can only be defined at the level of a stationary period (i.e., assuming
constant position during the whole stationary period) but you can define as many known stationary
periods as you wish. No likelihood map will be computed for these stationary periods and the
trajectory model will be more constrain, thus saving significant computational time.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX") |>
  tag_label()

# Default tag
tag <- tag_geostap(tag, c(-16, 23, 0, 50))
str(tag)

# Customized tag, with coarse grid scale, known position for the first stationary period and
# considering only the stationary periods lasting more than 20hours.
tag <- tag_geostap(tag,
  extent = c(-16, 23, 0, 50),
  scale = 1,
  include_min_duration = 20,
  known = data.frame(
    stap_id = 1,
    known_lon = 17.05,
    known_lat = 48.9
  )
)
str(tag)
}
