% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_auto.R
\name{tag_label_auto}
\alias{tag_label_auto}
\title{Automatic labelling acceleration data of a \code{tag}}
\usage{
tag_label_auto(
  tag,
  min_duration = 30,
  thr_reclassify = 0.1,
  post_proc_window = 2
)
}
\arguments{
\item{tag}{a GeoPressure \code{tag} object.}

\item{min_duration}{Minimal duration (in minutes) to consider a high activity as migratory
flight.}

\item{thr_reclassify}{Post-processing threshold of activity considered for re-classification.
Typically between \code{0} and \code{0.5} (no effect). See post-processing for details.}

\item{post_proc_window}{Post-processing windows considered for re-classification. Typically \code{0}
(no effect) to \code{5} in unit of the temporal resolution. See post-processing for details.}
}
\value{
Same data logger list than input \code{tag}, but with the column \code{label} filled with
\code{"flight"} in the acceleration data.frame when a sustained high-activity period is detected.
}
\description{
This function uses acceleration data to classify migratory flights. The function uses a
\code{k=2} mean clustering (\code{\link[=kmeans]{kmeans()}}) to identify high activity periods. Periods of high activity
lasting more than \code{min_duration} are then considered to be migratory flight.

Additionally, we perform a post-processing step to improve the classification of low-activity
period happening during a migration flight (e.g. gliding phase):
\enumerate{
\item Mid-activity surrounded by flights. The kmeans classification classify high-activity all
values above 50\% of the distance between the two clusters. Here, we classify as mid-activity
non-flight activity data which value are greater than \code{thr_reclassify} of the distance between
the two kmeans clusters. Any periods of mid-activity surrounded by flight are also considered
as flights.
\item To avoid stap lasting only a few datapoints, we also classify any datapoint
which is surrounded by a flight before and after, within a window of +/- \code{post_proc_window}
points. As a result, no stap will be shorter than \code{(2 * post_proc_window + 1) * dt}.
}

This function is inspired by the function \code{classify_flap} from the
\href{https://github.com/KiranLDA/PAMlr}{PAMlr package}.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)

  tag <- tag_label_auto(tag, min_duration = 15)
  str(tag$acceleration)
})

}
\seealso{
\href{https://geopressure.org/GeoPressureManual/labelling-tracks.html#with-or-without-acceleration-data}{GeoPressureManual}

Other tag_label:
\code{\link[=tag_label]{tag_label()}},
\code{\link[=tag_label_read]{tag_label_read()}},
\code{\link[=tag_label_stap]{tag_label_stap()}},
\code{\link[=tag_label_write]{tag_label_write()}}
}
\concept{tag_label}
