% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_read.R
\name{tag_label_read}
\alias{tag_label_read}
\title{Read a tag label file}
\usage{
tag_label_read(
  tag,
  file = glue::glue("./data/tag-label/{tag$param$id}-labeled.csv")
)
}
\arguments{
\item{tag}{a GeoPressure \code{tag} object.}

\item{file}{Absolute or relative path of the label file.}
}
\value{
Same data logger list as input, updated with the labels \code{tag$pressure$label} and
optionally \code{tag$acceleration$label}.
}
\description{
This function reads an exported csv file from \href{https://trainset.geocene.com/}{TRAINSET} and
updates the data logger dataset \code{tag}.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)

  tag <- tag_label_read(tag)

  str(tag)
})
}
\seealso{
\href{https://geopressure.org/GeoPressureManual/tag-object.html#read-labelled-file}{GeoPressureManual}

Other tag_label:
\code{\link[=tag_label]{tag_label()}},
\code{\link[=tag_label_auto]{tag_label_auto()}},
\code{\link[=tag_label_stap]{tag_label_stap()}},
\code{\link[=tag_label_write]{tag_label_write()}}
}
\concept{tag_label}
