% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_stap.R
\name{tag_label_stap}
\alias{tag_label_stap}
\title{Create stationary periods from a tag label}
\usage{
tag_label_stap(
  tag,
  quiet = FALSE,
  warning_flight_duration = lifecycle::deprecated(),
  warning_stap_duration = lifecycle::deprecated()
)
}
\arguments{
\item{tag}{a GeoPressure \code{tag} object.}

\item{quiet}{logical to display warning message.}

\item{warning_flight_duration}{lifecycle::deprecated()}

\item{warning_stap_duration}{lifecycle::deprecated()}
}
\value{
\code{tag} is return with (1) a new data.frame of stationary periods \code{tag$stap} and (2) a new
column \code{stap_id} for each sensor data.
}
\description{
This function computes the stationary periods from the pressure and/or acceleration label data.
In most case, this function should be run directly after \code{tag_label_read()} in order to update
\code{tag} to correspond to the new label file.

A stationary period is a period during which the bird is considered static relative to the
spatial resolution of interest (~10-50km). They are defined by being separated by a flight of any
duration (label \code{"flight"}). The \code{stap_id} is an integer value for stationary periods and decimal
value for flight. The \code{stap_id} is added as a new column to each sensor data.

If an acceleration data.frame is present and contains a column \code{label}, the stationary period
will be computed from it, otherwise, it uses the pressure data.frame.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label_read()

  tag <- tag_label_stap(tag)

  str(tag)

  str(tag$stap)
})
}
\seealso{
\href{https://geopressure.org/GeoPressureManual/tag-object.html#compute-stationary-periods}{GeoPressureManual}

Other tag_label:
\code{\link[=tag_label]{tag_label()}},
\code{\link[=tag_label_auto]{tag_label_auto()}},
\code{\link[=tag_label_read]{tag_label_read()}},
\code{\link[=tag_label_write]{tag_label_write()}}
}
\concept{tag_label}
