% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_update.R
\name{tag_label_update}
\alias{tag_label_update}
\title{Update tag label and recompute geostap}
\usage{
tag_label_update(
  tag,
  file = glue::glue("data/1_labels/{tag$id}-labeled.csv"),
  geostap = NA,
  path_pres = NA
)
}
\arguments{
\item{tag}{List containing the data logger dataset, this needs to contain at least a \code{pressure}
data.frame, but can also have a \code{light} and \code{acceleration} data.frame (see \code{\link[=tag_read]{tag_read()}}).}

\item{file}{Absolute or relative path of the label file.}

\item{geostap}{List of the geographical and stationary period information. See
\code{\link[=geostap_create]{geostap_create()}} for details.}

\item{path_pres}{geopressure_timeseries}
}
\value{
a list containing the new \code{tag}, \code{geostap} and \code{path_pres}.
}
\description{
When updating the labelling file of a tag, often, only a few stationary periods are changing. To
avoid recomputing the entire workflow, this function figure out which stationary period have
been changing on only update those in \code{geostap$map_pressure} and \code{path_pres}.
}
\examples{
tag <- tag_label(tag)

geostap <- tag |>
  geostap_create(extent = c(-16, 23, 0, 50), scale = 1) |>
  geopressure_map(tag$pressure)

path_pres <- geostap |>
  geostap2path() |>
  geopressure_timeseries(tag$pressure)

update <- tag_label_update(
  file = "data/1_labels/18LX-labeled-modif.csv",
  tag,
  geostap,
  path_pres
)

list2env(update, env = .GlobalEnv)
}
