% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_write.R
\name{tag_label_write}
\alias{tag_label_write}
\title{Write a tag label file}
\usage{
tag_label_write(
  tag,
  file = glue::glue("./data/tag-label/{tag$param$id}.csv"),
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressure \code{tag} object.}

\item{file}{Absolute or relative path of the label file to be saved.}

\item{quiet}{logical to display message.}
}
\value{
The file pathname is return invisibly
}
\description{
This function writes the csv file of labelled activity and pressure which can
be read with \href{https://trainset.geocene.com/}{TRAINSET}. If no label data exist, it will first
initialize the label data.

Optionally, it can also export a reference dataset for pressure \code{tag$pressure$ref} as another
series to be visualized on TRAINSET, but without impacting the labelling process.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)

  # Writing unlabelled tag will initialize the labelling for trainset
  tag_label_write(tag)

  # Writing unlabelled tag will initialize the labelling for trainset
  tag <- tag_label_auto(tag)
  tag_label_write(tag)

  # Writing labelled tag will use the existing labels
  tag <- tag_label(tag)
  tag_label_write(tag)
})

}
\seealso{
\href{https://geopressure.org/GeoPressureManual/tag-object.html#initialize-and-create-the-label-.csv-file}{GeoPressureManual}

Other tag_label:
\code{\link[=tag_label]{tag_label()}},
\code{\link[=tag_label_auto]{tag_label_auto()}},
\code{\link[=tag_label_read]{tag_label_read()}},
\code{\link[=tag_label_stap]{tag_label_stap()}}
}
\concept{tag_label}
