% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_prepare_likelihood.R
\name{tag_prepare_likelihood}
\alias{tag_prepare_likelihood}
\title{Prepare likelihood maps for modeling}
\usage{
tag_prepare_likelihood(
  tag,
  likelihood = NULL,
  thr_likelihood = 0.99,
  thr_gs = 150,
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character).}

\item{thr_likelihood}{threshold of percentile to keep likely locations in each stationary
period.}

\item{quiet}{logical to hide progress messages.}
}
\value{
A likelihood \code{map} object with:
\itemize{
\item \code{lk_norm}: list of normalized likelihood maps for included staps.
\item \code{lk_mask}: list of logical masks for included staps.
\item \code{lk_sparse}: list of sparse likelihood entries (\code{idx}, \code{prob}) for included staps.
\item \code{likelihood}: resolved likelihood field(s).
}
}
\description{
Prepare likelihood maps for modeling
}
\keyword{internal}
