% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_read.R
\name{tag_read}
\alias{tag_read}
\title{Read tag data}
\usage{
tag_read(
  id,
  directory = glue::glue("data/0_tag/{id}/"),
  pressure_file = "*.pressure",
  light_file = "*.glf",
  acceleration_file = "*.acceleration",
  crop_start = "1900-01-01",
  crop_end = "2100-01-01"
)
}
\arguments{
\item{id}{Unique identifier of the tag (character)}

\item{directory}{Path of the directory where the tag files can be read}

\item{pressure_file}{Name of the file with pressure data. Extension must be \code{.pressure} or
\code{.deg} (required).}

\item{light_file}{Name of the file with light data. Extension must be \code{.glf}, \code{.lux} (or \code{NA}
to ignore).}

\item{acceleration_file}{Name of the file with acceleration data. Extension must be
\code{.acceleration}, \code{.deg} (or \code{NA} to ignore).}

\item{crop_start}{Remove all data before this date (POSIXct or character in UTC).}

\item{crop_end}{Remove all data after this date (POSIXct or character in UTC).}
}
\value{
A list containing
\itemize{
\item \code{id} character of the unique identifier of the tag
\item \code{pressure} data.frame with column \code{date} and \code{value}
\item \code{light} (optional) same structure as pressure
\item \code{acceleration} (optional) same structure as pressure
}
}
\description{
Imports multi-sensor logger data from a folder (\code{directory}) and optionally crops at specific
date. The \verb{*_file} arguments are matched using a regex expression (e.g., \code{"*.pressure"}
matches any file with the extension \code{pressure}).

The current implementation can read the files from:
\itemize{
\item \href{https://www.vogelwarte.ch/en/projects/bird-migration/tracking-devices-miniaturized-geolocators}{Swiss Ornithological Institute (SOI) }
(default) with \code{pressure_file = "*.pressure"}, \code{light_file = "*.glf"} and
\code{acceleration_file = "*.acceleration"}
\item \href{http://www.migratetech.co.uk/}{Migrate Technology} with \code{pressure_file = "*.deg"},
\code{light_file = "*.lux"} and \code{acceleration_file = "*.deg"}
}

Please create \href{https://github.com/Rafnuss/GeoPressureR/issues/new}{an issue on Github} if you have data
in a format not supported yet.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))

tag <- tag_read("18LX")
str(tag)

tag <- tag_read("18LX", crop_start = "2017-08-01", crop_end = "2017-08-05")
str(tag)

# For Migrate Technology file, use
tag <- tag_read("CB621",
  pressure_file = "*.deg",
  light_file = "*.lux",
  acceleration_file = NA
)
str(tag)

# You can also specify exactly the file in case multiple file with the same
# extension exist in your directory
tag <- tag_read("CB621",
  pressure_file = "CB621_BAR.deg",
  light_file = NA,
  acceleration_file = NA
)
str(tag)

}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#read-geolocator-data}{GeoPressureManual | Pressure Map }
}
