% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag_read_check}
\alias{tag_read_check}
\title{Check that a given \code{pathname} and \code{file} exists and is unique.}
\usage{
tag_read_check(pathname, file)
}
\arguments{
\item{pathname}{is the path where file is}

\item{file}{is the name of he file}
}
\description{
Check that a given \code{pathname} and \code{file} exists and is unique.
}
\seealso{
\code{\link[=tag_read]{tag_read()}}
}
