% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag_read_delim_dto}
\alias{tag_read_delim_dto}
\title{Read data file with a DTO format (Date Time Observation)}
\usage{
tag_read_delim_dto(
  path,
  skip = 6,
  col = 3,
  date_format = "\%d.\%m.\%Y \%H:\%M"
)
}
\arguments{
\item{path}{is the full path of the file (pathname + filename)}

\item{skip}{is the number of lines of the data file to skip before beginning to read data.}

\item{col}{is the the index of the column of the data to take as observation}

\item{date_format}{format of the date}
}
\description{
Read data file with a DTO format (Date Time Observation)
}
\seealso{
\code{\link[=tag_read]{tag_read()}}
}
