% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag_stap}
\alias{tag_stap}
\title{Compute stationary periods}
\usage{
tag_stap(tag)
}
\arguments{
\item{tag}{data logger dataset list. See \code{\link[=tag_read]{tag_read()}}.}
}
\value{
Same as input \code{tag} but with (1) a new data.frame of stationary periods \code{tag$sta} and (2)
a new column \code{stap} for pressure and light data.
}
\description{
This function computes the stationary periods from classified acceleration data
(\code{tag$acceleration$ismig}).
}
\examples{
tag <- tag_read(
  pathname = system.file("extdata/0_tag/18LX", package = "GeoPressureR")
)
tag <- trainset_read(tag,
  pathname = system.file("extdata/1_pressure/labels", package = "GeoPressureR")
)
tag <- tag_stap(tag)
head(tag$pressure)
head(tag$light)
}
\seealso{
\code{\link{tag_read}}, \code{\link{tag_classify}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#identify-stationary-periods}{GeoPressureManual | Pressure Map }
}
