% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_assert.R
\name{tag_status}
\alias{tag_status}
\title{Get tag processing status}
\usage{
tag_status(tag)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}
}
\value{
Character vector of status labels.
}
\description{
Internal helper used by GeoPressure companion packages.
}
\keyword{internal}
