% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_update.R
\name{tag_update}
\alias{tag_update}
\title{Update a \code{tag} object}
\usage{
tag_update(
  tag,
  file = glue::glue("./data/tag-label/{tag$param$id}-labeled.csv"),
  known = NULL,
  include_stap_id = NULL,
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{file}{absolute or relative path of the label file.}

\item{known}{data.frame containing the known positions of the bird (e.g., equipment or retrieval
site). The default is to use \code{tag$stap}, which assumes that the \code{stap_id} has not changed for the
known stationary periods.}

\item{include_stap_id}{vector of \code{stap_id} defining which stationary period to model, that is,
to compute in the likelihood map and use in the graph. The default is to use the same original
value.}

\item{quiet}{logical to hide messages about the progress}
}
\value{
The updated \code{tag} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because (1) few people use it, (2) hard to maintain and (3) not
so slow to use the alternative of creating of new \code{tag} altogether.

When updating the labelling file of a \code{tag}, we often change only a few stationary periods. To
avoid recomputing the entire workflow, this function identifies which stationary periods have
been modified and updates only these ones in \code{tag$map_pressure} and \code{tag$map_light}.
}
\keyword{internal}
