% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolightviz.R
\name{time2plottime}
\alias{time2plottime}
\title{Convert datetimes to a plotting time-of-day scale}
\usage{
time2plottime(x, ref = x[1])
}
\arguments{
\item{x}{a POSIXct vector of datetimes.}

\item{ref}{reference datetime used to keep ordering around midnight.}
}
\value{
POSIXct vector on an arbitrary day encoding plotted time-of-day.
}
\description{
Convert datetimes to a plotting time-of-day scale
}
\keyword{internal}
