% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainset.R
\name{trainset.edit}
\alias{trainset.edit}
\title{Edit classification of activity and pressure}
\usage{
trainset.edit(
  PAM_data,
  file.location,
  file.name = paste0(PAM_data$id, "_act_pres")
)
}
\arguments{
\item{PAM_data}{PAM logger dataset list (see https://github.com/KiranLDA/PAMLr for details)}

\item{file.location}{Path to the folder where the labeled files should be saved}

\item{file.name}{Name for the file.}
}
\value{
PAM logger dataset list updated with the new label named \code{class} (\code{PAM_data$pressure$class} and \code{PAM_data$acceleration$class})
}
\description{
This function perform three steps: (1) write the csv file of the
automatically labeled activity and pressure with \code{trainset.write}, (2) open
trainset in your broweser (https://trainset.geocene.com/) so that you can
edit the labels and (3) read the exported csv file from trainset with
\code{trainset.read}.
}
