% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainset.R
\name{trainset.read}
\alias{trainset.read}
\title{Read classification of activity and pressure}
\usage{
trainset.read(
  PAM_data,
  file.location,
  file.name = paste0(PAM_data$id, "_act_pres")
)
}
\arguments{
\item{PAM_data}{PAM logger dataset list (see https://github.com/KiranLDA/PAMLr for details)}

\item{file.location}{Path to the folder where the labeled file is.}

\item{file.name}{Name of the file.}
}
\value{
PAM logger dataset list updated with the new label named \code{class} (\code{PAM_data$pressure$class} and \code{PAM_data$acceleration$class})
}
\description{
This function read an exported csv file from trainset (https://trainset.geocene.com/) and update the PAM logger dataset
}
