% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainset.R
\name{trainset.write}
\alias{trainset.write}
\title{Write classification of activity and pressure}
\usage{
trainset.write(
  PAM_data,
  file.location,
  file.name = paste0(PAM_data$id, "_act_pres")
)
}
\arguments{
\item{PAM_data}{PAM logger dataset list (see https://github.com/KiranLDA/PAMLr for details)}

\item{file.location}{Path to the folder where the labeled files should be saved}

\item{file.name}{Name for the file.}
}
\description{
This function writes the csv file of the automatically labeled activity and pressure which can be read with TRAINSET (https://trainset.geocene.com/).
}
