% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainset.R
\name{trainset_read}
\alias{trainset_read}
\title{Read classification of activity and pressure}
\usage{
trainset_read(
  tag,
  pathname = "data/1_pressure/labels/",
  filename = paste0(tag$id, "_act_pres-labeled.csv")
)
}
\arguments{
\item{tag}{data logger dataset list}

\item{pathname}{Path to the folder where the labeled file is.}

\item{filename}{Name of the file.}
}
\value{
data logger dataset list updated with the labels (\code{tag$pressure$label} and
\code{tag$acceleration$label})
}
\description{
This function read an exported csv file from \href{https://trainset.geocene.com/}{trainset} and update
the data logger dataset
}
\examples{
tag <- tag_read(pathname = system.file("extdata/0_tag/18LX", package = "GeoPressureR"))
tag <- tag_classify(tag)
tag <- trainset_read(tag, pathname = system.file("extdata/1_pressure/labels/",
  package = "GeoPressureR"
))
head(tag$pressure)
head(tag$acceleration)
}
\seealso{
\code{\link{tag_read}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#edit-activity-on-trainset}{GeoPressureManual | Pressure Map }
}
