% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{trainset_write}
\alias{trainset_write}
\title{Write classification of activity and pressure}
\usage{
trainset_write(pam, pathname, filename = paste0(pam$id, "_act_pres"))
}
\arguments{
\item{pam}{pam logger dataset list}

\item{pathname}{Path to the folder where the labeled files should be
saved}

\item{filename}{Name for the file.}
}
\description{
This function writes the csv file of the automatically labeled activity and
pressure which can be read with TRAINSET (https://trainset.geocene.com/).
}
