% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_label_write.R
\name{twilight_label_write}
\alias{twilight_label_write}
\title{Write a twilight label file}
\usage{
twilight_label_write(
  tag,
  file = glue::glue("./data/twilight-label/{tag$param$id}.csv"),
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{file}{Name of the twilight label file to be saved.}

\item{quiet}{logical to hide messages}
}
\value{
None
}
\description{
This function writes the csv file of the labelled twilight which can be read with
TRAINSET \url{https://trainset.raphaelnussbaumer.com/}.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    twilight_create()

  label_file <- twilight_label_write(tag)

  str(utils::read.csv(label_file))
})
}
\seealso{
Other geolight:
\code{\link[=geolight_fit_location]{geolight_fit_location()}},
\code{\link[=geolight_map]{geolight_map()}},
\code{\link[=geolight_solar]{geolight_solar()}},
\code{\link[=plot_twl_calib]{plot_twl_calib()}},
\code{\link[=twilight_create]{twilight_create()}},
\code{\link[=twilight_label_read]{twilight_label_read()}}
}
\concept{geolight}
