% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_add_wind.R
\name{edge_add_wind}
\alias{edge_add_wind}
\title{Retrieve ERA5 variable along edge}
\usage{
edge_add_wind(
  tag_graph,
  edge_s,
  edge_t,
  pressure = NULL,
  variable = c("u", "v"),
  rounding_interval = 60,
  interp_spatial_linear = FALSE,
  return_average_variable = FALSE,
  file = function(stap_id)
    glue::glue("./data/wind/{tag_graph$param$id}/{tag_graph$param$id}_{stap_id}.nc"),
  quiet = FALSE
)
}
\arguments{
\item{tag_graph}{either a \code{tag} or a \code{graph} GeoPressureR object.}

\item{edge_s}{a matrix of the index of the source node of the edge}

\item{edge_t}{a matrix of the index of the target node of the edge}

\item{pressure}{pressure measurement of the associated \code{tag} data used to estimate the pressure
level (i.e., altitude) of the bird during the flights. This data.frame needs to contain \code{date} as
POSIXt and \code{value} in hPa.}

\item{variable}{variable the the}

\item{interp_spatial_linear}{logical}

\item{quiet}{logical to hide messages about the progress}

\item{return_average_var}{logical}
}
\value{
a \code{graph} object with windspeed and airspeed as \code{ws} and \code{as} respectively.
}
\description{
Reads the NetCDF files downloaded and interpolate the average windspeed experienced by the
bird on each possible edge, as well as the corresponding airspeed.
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/trajectory-with-wind.html}{GeoPressureManual}

Other graph: 
\code{\link{graph_add_wind}()},
\code{\link{graph_create}()},
\code{\link{graph_marginal}()},
\code{\link{graph_most_likely}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_simulation}()},
\code{\link{print.graph}()}

Other movement: 
\code{\link{bird_create}()},
\code{\link{graph_add_wind}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_transition}()},
\code{\link{plot_graph_movement}()},
\code{\link{speed2prob}()},
\code{\link{tag_download_wind}()}
}
\concept{graph}
\concept{movement}
