% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight.R
\name{find_twilights}
\alias{find_twilights}
\title{Returns twilights for each day based on a threshold of light}
\usage{
find_twilights(light, threshold = NA, shift_k = NA)
}
\arguments{
\item{light}{a dataframe with columns \code{date} and \code{obs} that are the sequence of sample
times (as POSIXct) and light levels recorded by the tag.}

\item{threshold}{the light threshold that defines twilight.}

\item{shift_k}{shift of the middle of the night compared to 00:00 UTC (in seconds). If not
provided, will try to figure it out from the data}
}
\value{
A dataframe with columns
\item{\code{twilight}}{times of twilight}
\item{\code{rise}}{logical indicating sunrise}
where each row corresponds to a single twilight.
}
\description{
Search for sunset, sunrise pairs that correspond to a given light threshold. Function inspired
from \code{findTwilights} of the package Geolight
}
