% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flight_bird}
\alias{flight_bird}
\title{Construct bird morphology}
\usage{
flight_bird(
  species_name,
  mass = NA,
  wing_span = NA,
  wing_aspect = NA,
  wing_area = NA,
  body_frontal_area = NA
)
}
\arguments{
\item{species_name}{Scientific name of the species}

\item{mass}{Mass of the bird in kilogram.}

\item{wing_span}{Wing span in meter.}

\item{wing_aspect}{Wing aspect ratio (no unit).}

\item{wing_area}{Wing area in meter square.}

\item{body_frontal_area}{Body frontal area in meter square.}
}
\value{
List containing mass, wing span, wing aspect ratio and body frontal area.
}
\description{
This function return a list with the four morphological information necessary to construct the
power curve: mass, wing span, wing aspect ratio and body frontal area.
}
\details{
When any of these variables are missing, we query the
\href{https://doi.org/10.6084/m9.figshare.16586228.v5/}{AVONET database} using the scientific
name from \href{https://www.birds.cornell.edu/clementschecklist/}{the Clements Checklist}.
}
\examples{
# Using AVONET dataset
flight_bird("Acrocephalus arundinaceus")
# Using AVONET dataset + custom values
flight_bird("Acrocephalus arundinaceus", wing_aspect = 8)
# Import your own bird. You will need mass, and at least two of wing_span,
# wing_aspect or wing_area.
flight_bird("custum_bird", mass = 1, wing_span = 1, wing_aspect = 4)
}
