% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight.R
\name{flight_power}
\alias{flight_power}
\title{Power curve}
\usage{
flight_power(as, bird)
}
\arguments{
\item{as}{airspeed in m/s}

\item{bird}{list of basic morphological trait necessary: mass, wing span,
wing aspect ratio and body frontal area. It is best practice to create bird
with \code{flight_bird()}.}
}
\value{
mechanical power in Watt (or Joule/seconds) corresponding to the
airspeed
}
\description{
Compute the mechanical power required for a specific bird flying as at a
given airspeed in m/s. \code{bird} (created with \code{flight_bird()})
}
\examples{
bird <- flight_bird("Acrocephalus arundinaceus")
airspeed <- seq(0, 30)
power <- flight_power(airspeed, bird)
plot(airspeed, power, xlab = "Airspeed [m/s]", ylab = "Mechanical Power [W]")
}
