% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight_map_calibrate.R
\name{geolight_calibrate}
\alias{geolight_calibrate}
\title{Calibrate twilight to zenith angle using calibration locations}
\usage{
geolight_calibrate(twl, calib_stap, twl_calib_adjust = 1.4)
}
\arguments{
\item{twl}{a twilight data.frame}

\item{calib_stap}{a stationary period data.frame with calibration locations. Must contain
the columns \code{known_lat}, \code{known_lon}, \code{start}, and \code{end}. This can include both true known
locations and estimated (fitted) locations.}

\item{twl_calib_adjust}{smoothing parameter for the kernel density (see \code{\link[stats:density]{stats::density()}}).}
}
\value{
a \code{twl_calib} object with components:
\itemize{
\item \code{x}: the zenith angle sequence
\item \code{y}: the estimated density values
\item \code{hist_breaks}: common histogram breaks used for all calibration staps
\item \code{hist_counts}: list of histogram counts per calibration stationary period
\item \code{calib_stap}: calibration stationary periods used for calibration
}
}
\description{
Calibrate twilight to zenith angle using calibration locations
}
\keyword{internal}
