% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight.R
\name{geolight_light2mat}
\alias{geolight_light2mat}
\title{Convert light data in matrix format}
\usage{
geolight_light2mat(light, shift_k = 0)
}
\arguments{
\item{light}{a dataframe with columns \code{date} and \code{value} that are the sequence of sample
times (as POSIXct) and light levels recorded by the tag.}

\item{shift_k}{shift of the middle of the night compared to 00:00 UTC (in seconds). If not
provided, will try to figure it out from the data}
}
\value{
A dataframe with columns value and date
}
\description{
Convert light data in matrix format
}
