% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight_map.R
\name{geolight_map}
\alias{geolight_map}
\title{Compute likelihood map from twilight}
\usage{
geolight_map(
  geostap,
  twilight,
  twl_calib_adjust = 1.4,
  twl_llp = function(n) 0.1
)
}
\arguments{
\item{geostap}{Data}

\item{twilight}{A data.frame with columns \code{twilight} (date-time of twilights) and \code{discard}
(see \code{\link[=twilight_create]{twilight_create()}})}

\item{twl_calib_adjust}{Smoothing parameter for the kernel density (see \code{\link[stats:kernel]{stats::kernel()}}).}

\item{twl_llp}{Log-linear pooling aggregation weight.}
}
\value{
a \code{geostap} with the likelihood of light as \code{map_light}
}
\description{
This functions uses a \code{geostap} and a data.frame of twilight to estimate a likelihood map
for each stationary period. The functions performs the following steps:
\enumerate{
\item Perform a calibration on the known stationary period \verb{geostap$stap$known_l**}.
\item Compute a likelihood map for each twilight using the calibration
\item Combine all likelihood maps of the same calibration period using the stationary period dates
}
}
\section{Calibration}{
Instead of calibrating the twilight errors in terms of duration, we directly model the zenith
angle error.
The \code{twl_calib_adjust} parameter allows to manually set how smooth you want the fit to be.
Because the zenith angle error model is fitted with data from the calibration site only, and
we are using it for all locations of the bird’s journey, it is safer to assume a broader/smoother
distribution.
}

\section{Log-linear pooling of the twilight likelihood map}{
The twilight maps are aggregated by stationary period according to the date and time defined in
\code{tag$stap} and twilight. See \href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html#probability-aggregation-1}{GeoPressureManual | Probability aggregation}
for more information on probability aggregation using log-linear pooling.
}

\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
# Read geolocator data and build twilight
tag <- tag_read("18LX") |> tag_label()
tag <- twilight_create(tag) |> twilight_label_read()
# Initiate the geostap structure
geostap <- geostap_create(tag,
  extent = c(-16, 23, 0, 50),
  scale = 10,
  known = data.frame(
    stap_id = 1,
    known_lon = 17.05,
    known_lat = 48.9
  )
)

# Compute likelihood map
geostap <- geolight_map(
  geostap,
  tag$twilight
)

terra::plot(
  terra::rast(geostap$map_light[[1]],
    extent = geostap$extent
  ),
  main = "Likelihood"
)
}
