% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geolight_map.R
\name{geolight_map}
\alias{geolight_map}
\title{Compute likelihood map from twilight}
\usage{
geolight_map(
  tag,
  twl_calib_adjust = 1.4,
  twl_llp = function(n) log(n)/n,
  compute_known = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{twl_calib_adjust}{smoothing parameter for the kernel density (see \code{\link[stats:kernel]{stats::kernel()}}).}

\item{twl_llp}{log-linear pooling aggregation weight.}

\item{compute_known}{logical defining if the map(s) for known stationary period should be
estimated based on twilight or hard defined by the known location \verb{stap$known_l**}}

\item{quiet}{logical to hide messages about the progress}
}
\value{
a \code{tag} with the likelihood of light as \code{tag$map_light}
}
\description{
This functions estimate a likelihood map for each stationary period based on twilight data.
The functions performs the following steps:
\enumerate{
\item Perform a calibration on the known stationary period. See below for details
\item Compute a likelihood map for each twilight using the calibration.
\item Combine the likelihood maps of all twilights belonging to the same stationary periods with a
log-linear pooling. See \href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation}
for more information on probability aggregation using log-linear pooling.
}
}
\section{Calibration}{
Calibration requires to have a known position for a least one stationary periods. Use
\code{tag_set_map()} to define the known position.

Instead of calibrating the twilight errors in terms of duration, we directly model the zenith
angle error. We use a kernel distribution to fit the zenith angle during the known stationary
period(s). The \code{twl_calib_adjust} parameter allows to manually adjust how smooth you want the
fit of the zenith angle to be. Because the zenith angle error model is fitted with data from the
calibration site only, and we are using it for all locations of the bird’s journey, it is safer
to assume a broader/smoother distribution.
}

\examples{
setwd(system.file("extdata", package = "GeoPressureR"))
# Read geolocator data and build twilight
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  tag_set_map(
    extent = c(-16, 23, 0, 50),
    scale = 10,
    known = data.frame(
      stap_id = 1,
      known_lon = 17.05,
      known_lat = 48.9
    )
  )

# Compute the twilight
tag <- twilight_create(tag) |> twilight_label_read()

# Compute likelihood map
tag <- geolight_map(tag, quiet = TRUE)

plot(tag, type = "map_light")
}
\seealso{
Other geolight: 
\code{\link{path2twilight}()},
\code{\link{twilight_create}()},
\code{\link{twilight_label_read}()},
\code{\link{twilight_label_write}()}
}
\concept{geolight}
