% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_map.R,
%   R/geopressure_map_likelihood.R, R/geopressure_map_mismatch.R
\name{geopressure_map}
\alias{geopressure_map}
\alias{geopressure_map_likelihood}
\alias{geopressure_map_mismatch}
\title{Compute likelihood map from pressure data}
\usage{
geopressure_map(
  tag,
  max_sample = 250,
  margin = 30,
  sd = 1,
  thr_mask = 0.9,
  log_linear_pooling_weight = function(n) log(n)/n,
  timeout = 300,
  workers = "auto",
  era5_dataset = "land",
  keep_mask = FALSE,
  keep_mse = FALSE,
  compute_known = FALSE,
  debug = FALSE,
  quiet = FALSE
)

geopressure_map_likelihood(
  tag,
  sd = 1,
  log_linear_pooling_weight = function(n) log(n)/n,
  keep_mse = TRUE
)

geopressure_map_mismatch(
  tag,
  max_sample = 250,
  margin = 30,
  thr_mask = 0.9,
  timeout = 300,
  workers = "auto",
  era5_dataset = "land",
  compute_known = FALSE,
  keep_mask = TRUE,
  debug = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{max_sample}{the computation of the maps is only performed on \code{max_sample} datapoints of
pressure to reduce computational time. The samples are randomly (uniformly) selected on the
time series.}

\item{margin}{the margin is used in the mask map to accept measurement errors, small-scale
topography, and vertical movements of the bird (unit in meters, 1hPa~10m).}

\item{sd}{standard deviation of the pressure error. Numeric of length 1 or number of stationary
periods.}

\item{thr_mask}{threshold of the percentage of data points outside the elevation range to be
considered not possible.}

\item{log_linear_pooling_weight}{weighting function of the log-linear pooling, taking the number
of samples of the stationary periods used and returning the weight of the aggregation. See
\href{https://geopressure.org/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation } for more details.}

\item{timeout}{duration before the code is interrupted both for the request on
GeoPressureAPI and on GEE (in seconds, see \code{httr2::req_timeout()}).}

\item{workers}{number of parallel requests on GEE. Integer between 1 and 99. \code{"auto"} adjust the
number of workers to the number of \code{stapelev} to query.}

\item{era5_dataset}{ERA5 product to use: \code{"land"} for ERA5-Land, \code{"single-levels"} for global
ERA5, or \code{"both"} to use ERA5-Land over land and global ERA5 elsewhere.}

\item{keep_mask}{logical defining if the mask map is returned in \code{tag}.}

\item{keep_mse}{logical defining if the MSE map is returned in \code{tag}.}

\item{compute_known}{logical defining if the map(s) for known stationary period should be
estimated based on twilight or hard defined by the known location \verb{stap$known_l**}}

\item{debug}{logical to display additional information to debug a request}

\item{quiet}{logical to hide messages about the progress}
}
\value{
Returns the same GeoPressureR \code{tag} object including the GeoPressureR \code{map} object
\code{tag$map_pressure} containing the likelihood map of each stationary period. See \code{map_create()}
for details.
If \code{keep_mask} and \code{keep_mse} are each true, \code{tag} also includes the \code{tag$map_pressure_mse} and
\code{tag$map_pressure_mask} maps, as well as \code{tag$stap$nb_sample}, indicating the number of
datapoints used to compute the MSE.
}
\description{
The \code{geopressure_map()} function computes a likelihood map for each stationary period based on
pressure measurements. It is a wrapper of the following two child functions:
\enumerate{
\item \code{geopressure_map_mismatch()} computes the mismatch maps between the pressure sensor
measurements and the ERA5 reanalysis database.
\item \code{geopressure_map_likelihood()} converts the mismatch maps into a likelihood map.
}

See below for details. For more background on the method behind these functions, please refer to
Nussbaumer et al. (2023a; \doi{10.1111/2041-210X.14043}).
}
\details{
A map will only be computed for the stationary periods included in \code{tag$stap$include} and without
a known position \verb{tag$stap$known_l**} as defined by \code{tag_set_map()}. At known stationary
periods, the likelihood map is a map of \code{0}s with a single \code{1} at the pixel closest to the known
position.
}
\section{Mismatch map with GeoPressureAPI}{


\code{geopressure_map_mismatch()} computes the mismatch maps on Google Earth Engine via the map
entry point of the \href{https://github.com/GeoPressure/GeoPressureAPI}{GeoPressure API}.
This consists of the following steps:
\enumerate{
\item \strong{Pre-process pressure}: the pressure measurements are first smoothed and downscaled to a
1-hour resolution in order to match ERA-5 resolution (see \code{geopressure_map_preprocess()}).
\item \strong{Generate requests}: Send a single request to the GeoPressureAPI to generate the Google
Earth Engine (GEE) URLs, one for each elevation/stationary period which can be used to compute
the maps on the GEE server.
\item \strong{Send the requests}: Call the URLs, which will start the computation on the GEE server.
At this step, it does not wait for an answer, but sends the requests in parallel to be faster.
\item \strong{Compute and download the maps}: When all requests are sent, we wait for the GEE server to
return a geotiff file (map) for each elevation/ stationary period.
\item \strong{Post-process maps}: Read these geotiff maps as matrix, create the corresponding
GeoPressureR \code{map} object with \code{create_map()}.
}

The following two maps are returned for each stationary period:
\enumerate{
\item \strong{map_pressure_mse}: The Mean Square Error (MSE) between the data logger pressure time series
and the reanalysis. The mean error is removed because we assume no specific altitude of the
geolocator, thus allowing an altitudinal shift of the pressure time series.
\item \strong{map_pressure_mask} (optionally): The mask of the proportion of pressure measurements
corresponding to altitude values found within the min and max ground elevation at each location.
The altitude value of the geolocator pressure time series is computed with the barometric formula
accounting for the temporal variation of pressure (surface-pressure) and temperature
(2m-temperature) based on ERA5 data. The min and max ground elevation of each pixel is computed
from SRTM-90. This map is only returned if \code{keep_mask} is \code{TRUE}.
}

For more details, read the \href{https://github.com/GeoPressure/GeoPressureAPI}{GeoPressure API documentation }.
}

\section{Elevation levels}{


It is possible to indicate different elevation levels when the bird was spending time at
locations with different elevations within a general area (~10km), and thus within the same
stationary period. This can be done by using \code{tag$label="elev_x"} for all measurements of the
same elevation level \emph{x}. See more information on the labelling of elevation levels in \href{https://geopressure.org/GeoPressureManual/labelling-tracks.html#elevation-period}{the corresponding section in the GeoPressureManual}.

Behind the scene, each of these elevation levels produces a new request on the GeoPressureAPI.
The mismatch maps of all elevation levels belonging to the same stationary periods are combined
(as a weighted average) to result in a single mismatch map per stationary period.
}

\section{Convert mismatch map into likelihood map}{


We convert the map of the mean square error \eqn{MSE} and altitude mask \eqn{z_{mask}} computed
by \code{\link[=geopressure_map_mismatch]{geopressure_map_mismatch()}} into a likelihood map with,

\deqn{L = \left( \frac{1}{2 \pi \sigma^2}\right)^{\frac{nw}{2}}
\exp \left(-w n \frac{MSE}{2\sigma^2} \right) \left[z_{mask}>T \right],}

where \eqn{\sigma} is the standard deviation (\code{sd}) of pressure and \eqn{T} is the mask threshold
(\code{thr_mask}).

Because the auto-correlation of the time series is not accounted for in this equation, we use a
log-linear pooling weight of \eqn{w=\log(n)/n} by default, where \eqn{n} is the number of samples
in the time series (i.e., data points used to compute the MSE). See \href{https://geopressure.org/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation } for details.

\strong{Important Note}: Since GeoPressure v3.1.0, the threshold of the mask is happening directly
on the GEE server (i.e., during \code{geopressure_map_mismatch()}). This allows to compute the MSE
only for pixels which are within the threshold, thus reducing the computational cost
significantly.
}

\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    tag_set_map(
      extent = c(-16, 23, 0, 50),
      scale = 4
    )
})

tag <- geopressure_map_mismatch(tag,
  max_sample = 50,
  margin = 20,
  thr_mask = 0.95,
  keep_mask = TRUE,
  quiet = TRUE
)

plot(tag, type = "map_pressure_mse", plot_leaflet = FALSE)

plot(tag, type = "map_pressure_mask", plot_leaflet = FALSE)

tag <- geopressure_map_likelihood(tag,
  sd = 1,
  log_linear_pooling_weight = function(n) log(n) / n
)

plot(tag, type = "map_pressure")

}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, and Felix Liechti. 2023.
Global Positioning with Animal‐borne Pressure Sensors. \emph{Methods in Ecology and Evolution}, 14,
1118–1129 \doi{10.1111/2041-210X.14043}.}
}
\seealso{
\href{https://geopressure.org/GeoPressureManual/pressure-map.html#compute-pressure-maps}{GeoPressureManual}
}
\concept{geopressure_map}
