% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_map.R
\name{geopressure_map}
\alias{geopressure_map}
\title{Compute likelihood map from pressure data}
\usage{
geopressure_map(
  tag,
  max_sample = 250,
  margin = 30,
  timeout = 60 * 5,
  workers = "auto",
  sd = 1,
  thr_mask = 0.9,
  log_linear_pooling_weight = function(n) log(n)/n,
  keep_mse_mask = FALSE,
  compute_known = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{max_sample}{The computation of the maps is only performed on \code{max_sample} datapoints of
pressure to reduce computational time. The samples are randomly (uniformly) selected on the
timeseries.}

\item{margin}{The margin is used in the mask map to accept measurement errors, small-scale
topography, and vertical movements of the bird (unit in meters, 1hPa~10m).}

\item{timeout}{Duration before the code is interrupted both for the request on
GeoPressureAPI and GEE (in seconds, see \code{\link[httr:timeout]{httr::timeout()}}).}

\item{workers}{Number of parallel requests on GEE. Integer between 1 and 99. \code{"auto"} adjust the
number of workers to the number of \code{stap_elev} to query.}

\item{sd}{Standard deviation of the pressure error . numeric of lenght 1 or number of stationary periods.}

\item{thr_mask}{Threshold of the percentage of data points outside the elevation range to be considered
not possible.}

\item{log_linear_pooling_weight}{Weighting function of the log-linear pooling, taking the number of samples of the
stationary periods used and returning the weight of the aggregation. See
\href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation } for more details.}

\item{keep_mse_mask}{logical defining if the matrix of the MSE and mask are returned in the list.}

\item{compute_known}{logical defining if the map(s) for known stationary period should be
estimated based on twilight or hard defined by the known location \verb{stap$known_l**}}

\item{quiet}{logical to hide messages about the progress}
}
\value{
Same list as parameter \code{tag} but with a \code{likelihood} field containing a list of maps
for each stationary period stored as matrix. Note that stationary periods not marked as model in
\code{tag$stap$include} will be included with a \code{NULL}value in \code{likelihood}.
\code{tag$param} is a list of all the parameters used to compute the likelihood map.
If \code{keep_mse_mask} is
true, the \code{mse} and \code{mask} maps computed with \code{geopressure_map_mismatch} will also be kept, as
well as \code{tag$stap$nb_sample}, indicating the number of datapoints used to compute the MSE.
}
\description{
This function computes a likelihood map for each stationary period based on pressure measurements.
It performs the following operations:
\enumerate{
\item Compute the mismatch map between the pressure sensor measurements and the ERA5 reanalysis
database. See \code{\link[=geopressure_map_mismatch]{geopressure_map_mismatch()}} for details.
\item Convert the mismatch map into a likelihood map with \code{\link[=geopressure_map_likelihood]{geopressure_map_likelihood()}}.
}

A map will only be computed for the stationary periods included in \code{tag$stap$include} and without a
known position \verb{tag$stap$known_l**}. If the position is known, the function will
create a likelihood map with a single 1 value at the grid cell closest to the known position.

For more background and details on the method behind these functions, please refer to the \href{https://doi.org/10.1111/2041-210X.14043}{associated scientific publication}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  tag_set_map(
    extent = c(-16, 23, 0, 50),
    scale = 4
  )

tag <- geopressure_map(tag,
  max_sample = 50,
  sd = 0.7,
  keep_mse_mask = TRUE
)

plot(tag, type = "map_pressure_mse", plot_leaflet = FALSE)
plot(tag, type = "map_pressure_mask", plot_leaflet = FALSE)

}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, and Felix Liechti. 2023.
Global Positioning with Animal‐borne Pressure Sensors. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\url{https://doi.org/10.1111/2041-210X.14043}.}
}
\seealso{
Other geopressure_map: 
\code{\link{geopressure_map_likelihood}()},
\code{\link{geopressure_map_mismatch}()}
}
\concept{geopressure_map}
