% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_map}
\alias{geopressure_map}
\title{Request and download mismatch maps of pressure}
\usage{
geopressure_map(pressure, extent, scale = 10, max_sample = 250, margin = 30)
}
\arguments{
\item{pressure}{Pressure data.frame from a PAM logger. This data.frame needs to contains \code{date}
as POSIXt, \code{obs} in hPa, \code{sta_id} grouping observation measured during the same stationary period
and \code{isoutliar} as logical to label observation which need to be ignorede. It is best practice to
use \code{pam_read()} and \code{pam_sta()} to build this data.frame.}

\item{extent}{Geographical extent of the map to query as a list ordered by North, West, South,
East  (e.g. \code{c(50,-16,0,20)}).}

\item{scale}{Number of pixel per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and 4 for a resolution of 0.25° (~30km). To avoid interpolating the
ERA5 data, scale should be smaller than 10. Read more about \href{https://developers.google.com/earth-engine/guides/scale}{scale on Google earth Engine documentation}.}

\item{max_sample}{The computation of the mismatch is only performed on \code{max_sample} datapoints of
pressure to reduce computational time. The samples are randomly (uniformly) selected on the
timeserie.}

\item{margin}{The margin is used in the threshold map to accept some measurement error. unit in
meter. (1hPa~10m)}
}
\value{
List of raster map
}
\description{
This function return a maps of mismatch of atmospheric pressure measured by a geolocator. It
performs the following actions:
\enumerate{
\item Send a query to produce the Google Earth Engine (GEE) url of the code producing the maps for
each stationary periods separately
\item Read these map (geotiff) in a raster
3 compute the likelihood map from the mismatch. See
\href{https://raphaelnussbaumer.com/GeoPressureServer/#description}{the GeoPressure API documentation}
}
}
\examples{
# See pam_sta() for generating pam
\dontrun{
pressure_maps <- geopressure_map(
  pam$pressure,
  extent = c(50, -16, 0, 23),
  scale = 4,
  max_sample = 250,
  margin = 30
)
pressure_maps_1 <- pressure_maps[[1]]
}
pressure_maps_1 <- readRDS(system.file("extdata/1_pressure/", "18LX_pressure_maps_1.rda",
  package = "GeoPressureR"
))
raster::metadata(pressure_maps_1)
raster::plot(pressure_maps_1,
  main = c("Mean Square Error", "Mask of pressure")
)
}
\seealso{
\code{\link[=geopressure_prob_map]{geopressure_prob_map()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map }
}
