% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_map}
\alias{geopressure_map}
\title{Request and download mismatch and threshold maps of pressure}
\usage{
geopressure_map(pressure, extent, scale = 10, max_sample = 250, margin = 30)
}
\arguments{
\item{pressure}{Pressure data.frame from a PAM logger. This data.frame needs to contains \code{date}
as POSIXt, \code{obs} in hPa, \code{sta_id} grouping observation measured during the same stationary period
and \code{isoutlier} as logical to label observation which need to be ignored. It is best practice to
use \code{pam_read()} and \code{pam_sta()} to build this data.frame.}

\item{extent}{Geographical extent of the map to query as a list ordered by North, West, South,
East  (e.g. \code{c(50,-16,0,20)}).}

\item{scale}{Number of pixel per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and 4 for a resolution of 0.25° (~30km). To avoid interpolating the
ERA5 data, scale should be smaller than 10. Read more about \href{https://developers.google.com/earth-engine/guides/scale}{scale on Google earth Engine documentation}.}

\item{max_sample}{The computation of the maps is only performed on \code{max_sample} datapoints of
pressure to reduce computational time. The samples are randomly (uniformly) selected on the
timeserie.}

\item{margin}{The margin is used in the threshold map to accept some measurement error. unit in
meter. (1hPa~10m)}
}
\value{
List of raster map
}
\description{
This function returns for each stationary period (1) a map of mismatch between the pressure
measured by a geolocator and the ERA5 pressure database and (2) a map altitude threshold using
a fine scale digital elevation model.
}
\details{
These maps are generated on Google Earth Engine via the Pressure map entry point of the
\href{https://raphaelnussbaumer.com/GeoPressureAPI/#description}{GeoPressure API}. The computation
performed by this function consists of the following:
\enumerate{
\item Send a query to generate the Google Earth Engine (GEE) url of the code producing the maps for
each stationary periods separately.
\item Download and read these geotiff maps as raster.
}

The maps of each stationary period are returned in two layers:
\enumerate{
\item The mismatch between the input pressure timeseries and the reanalysis one at each location.
This is computed with a mean square error (MSE) where the mean error is removed. The mean error
is removed because we assume no specific altitude of the geolocator, thus allowing an
altitudinal shift of the pressure timeserie.
\item The proportion of datapoint of the input pressure timeseries corresponding to altitude value
which fall within the min and max ground elevation found at each location. The altitude value
of the geolocator pressure timeseries is computed with the barometric formula accounting for the
temporal variation of pressure (surface-pressure) and temperature (2m-temperature) based on
ERA5 data. The min and max ground elevation of each pixel is computed from SRTM-90.
}
}
\examples{
# See `pam_sta()` for generating pam
\dontrun{
pressure_maps <- geopressure_map(
  pam$pressure,
  extent = c(50, -16, 0, 23),
  scale = 4,
  max_sample = 250,
  margin = 30
)
pressure_maps_1 <- pressure_maps[[1]]
}
pressure_maps_1 <- readRDS(system.file("extdata/1_pressure/", "18LX_pressure_maps_1.rda",
  package = "GeoPressureR"
))
raster::metadata(pressure_maps_1)
raster::plot(pressure_maps_1,
  main = c("Mean Square Error", "Mask of pressure")
)
}
\seealso{
\code{\link[=geopressure_prob_map]{geopressure_prob_map()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map }
}
