% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_map_preprocess.R
\name{geopressure_map_preprocess}
\alias{geopressure_map_preprocess}
\title{Preprocess pressure data}
\usage{
geopressure_map_preprocess(pressure, stap)
}
\arguments{
\item{pressure}{Data.frame of the pressure measurements, usually \code{tag$pressure}.}

\item{stap}{Stationary period information (typically \code{geostap$stap} or \code{tag$stap}).}
}
\value{
Pressure data.frame without flight and discarded values, on a 1hr resolution.
}
\description{
Performs the follow pre-processing for the pressure data.frame:
\enumerate{
\item Remove discarded and flight pressure measurements.
\item Group pressure measurements by stationary period and elevation level.
\item Smooth and downscale pressure to a 1 hour resolution to match ERA5 data.
}

This function is used within \code{\link[=geopressure_map_mismatch]{geopressure_map_mismatch()}} but can be useful to check the
pressure data sent to the GeoPressureAPI.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |>
  tag_label()

pressure_processed <- geopressure_map_preprocess(tag$pressure, tag$stap)
str(pressure_processed)
}
