% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_map_preprocess.R
\name{geopressure_map_preprocess}
\alias{geopressure_map_preprocess}
\title{Prepare pressure data for geopositioning}
\usage{
geopressure_map_preprocess(tag, compute_known = FALSE)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{compute_known}{logical defining if the map(s) for known stationary period should be
estimated based on twilight or hard defined by the known location \verb{stap$known_l**}}
}
\value{
Pressure data.frame without flight and discarded values, on a 1 hour resolution.
}
\description{
Performs the follow pre-processing for the pressure data.frame:
\enumerate{
\item Remove discarded and flight pressure measurements.
\item Group pressure measurements by stationary period and elevation level.
\item Smooth and downscale pressure to a 1 hour resolution to match ERA5 data.
}

This function is used within \code{\link[=geopressure_map_mismatch]{geopressure_map_mismatch()}} but can be useful to check the
pressure data sent to the GeoPressureAPI.
}
\examples{
owd <- setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)
setwd(owd)

pressure_processed <- geopressure_map_preprocess(tag)
str(pressure_processed)
}
