% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_prob_map}
\alias{geopressure_prob_map}
\title{Compute probability raster}
\usage{
geopressure_prob_map(
  pressure_maps,
  s = 1,
  thr = 0.9,
  fun_w = function(n) {
     log(n)/n
 }
)
}
\arguments{
\item{pressure_maps}{List of raster built with \code{\link[=geopressure_map]{geopressure_map()}}.}

\item{s}{Standard deviation of the pressure error.}

\item{thr}{Threshold of the percentage of data point outside the elevation range to be considered
not possible.}

\item{fun_w}{function taking the number of sample of the timeseries used to compute the
probability map and return the log-linear pooling weight (see the
\href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation })}
}
\value{
List of the probability raster map
}
\description{
This function convert the raster of normalized MSE and altitude threshold \eqn{z_{thr}} computed
by \code{\link[=geopressure_map]{geopressure_map()}} into a probability map with,
\eqn{p = \exp \left(-w \frac{MSE}{s} \right) \left[z_{thr}>thr \right],}
where \eqn{s} is the standard deviation of pressure and \eqn{thr} is the threshold. Because the
auto-correlation of the timeseries is not accounted for in this equation, we use a log-linear
pooling weight \eqn{w=\log(n) - 1}, with \eqn{n} is the number of data point in the timeserie.
This operation is describe in \doi{10.21203/rs.3.rs-1381915/v1}.
}
\examples{
# See `geopressure_map()` for generating pressure_maps
\dontrun{
pressure_prob <- geopressure_prob_map(
  pressure_maps,
  s = 0.4,
  thr = 0.9
)
pressure_prob_1 <- pressure_prob[[1]]
}
pressure_prob_1 <- readRDS(system.file("extdata/1_pressure/", "18LX_pressure_prob_1.rda",
  package = "GeoPressureR"
))
raster::metadata(pressure_prob_1)
raster::plot(pressure_prob_1,
  main = "Probability",
  xlim = c(5, 20), ylim = c(42, 50)
)
}
\seealso{
\code{\link[=geopressure_map]{geopressure_map()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map }
}
