% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_prob_map}
\alias{geopressure_prob_map}
\title{Compute probability raster}
\usage{
geopressure_prob_map(pressure_maps, s = 1, thr = 0.9)
}
\arguments{
\item{pressure_maps}{list of raster built with \code{geopressure_map()}}

\item{s}{standard deviation of the pressure error}

\item{thr}{threshold of the percentage of data point outside the elevation range to be considered
not possible}
}
\value{
List of the probability raster map
}
\description{
This function convert the raster of noramlized MSE and altitude threshold \eqn{z_{thr}} computed
by \code{geopressure_map()} into a probability map with,
\eqn{p = \exp \left(-w \frac{MSE}{s} \right) \left[z_{thr}>thr \right],}
where \eqn{s} is the standard deviation of pressure and \eqn{thr} is the threashold. Because the
auto-correlation of the timeseries is not accounted for in this equation, we use a log-linear
pooling weight \eqn{w=\log(n) - 1}, with \eqn{n} is the number of data point in the timeserie.
This operation is describe in ...
}
\examples{
\dontrun{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
pam_data <- trainset_read(pam_data,
  pathname = system.file("extdata", package = "GeoPressureR")
)
pam_data <- pam_sta(pam_data)
pressure_maps <- geopressure_map(
  pam_data$pressure,
  extent = c(50, -16, 0, 20),
  scale = 10
)
pressure_prob <- geopressure_prob_map(
  pressure_maps,
  s = 0.4,
  thr = 0.9
)
}
load(system.file("extdata", "18LX_pressure_prob.rda", package = "GeoPressureR"))
raster::metadata(pressure_prob[[1]])
raster::plot(pressure_prob[[1]],
  main = "Probability",
  xlim = c(5, 20), ylim = c(42, 50)
)
}
