% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_timeseries.R
\name{geopressure_timeseries}
\alias{geopressure_timeseries}
\title{Request and download pressure timeseries at a given location}
\usage{
geopressure_timeseries(
  lat,
  lon,
  pressure = NULL,
  start_time = NULL,
  end_time = NULL,
  timeout = 60 * 5,
  quiet = FALSE,
  debug = FALSE
)
}
\arguments{
\item{lat}{Latitude to query (0° to 90°).}

\item{lon}{Longitude to query (-180° to 180°).}

\item{pressure}{A data.frame of pressure timeseries, containing at least a \code{"date"} and \code{"value"}
column.}

\item{start_time}{If \code{pressure} is not provided, \code{start_time} defines the start time of
the timeseries as POSIXlt.}

\item{end_time}{If \code{pressure} is not provided, \code{end_time} defines the end time of
the timeseries as POSIXlt.}

\item{timeout}{Duration before the code is interrupted both for the request on
GeoPressureAPI and on GEE (in seconds, see \code{httr2::req_timeout()}).}

\item{quiet}{logical to hide messages about the progress}

\item{debug}{logical to display additional information to debug a request}
}
\value{
A data.frame containing
\itemize{
\item \code{date} POSIXct date time
\item \code{pressure_era5} pressure (hPa)
\item \code{lon} same as input \code{lon} except if over water
\item \code{lat} same as input \code{lat} except if over water.
\item \code{pressure_era5_norm} only if \code{pressure} is provided as input
\item \code{altitude} only if \code{pressure} is provided as input
}
}
\description{
This function returns the surface atmospheric pressure timeseries from ERA5 at any requested
location.

If the location queried is over water, the location will be moved to the closest onshore
location.

The ERA5 pressure timeseries of the response \eqn{P_{ERA}} will be provided on a hourly basis
between \code{start_time} and \code{end_time} or the same as \code{pressure$date} if \code{pressure} is supplied.

If you supply the \code{pressure} of the geolocator \eqn{P_{gl}}, the function will
additionally return the altitude of the geolocator above sea level \eqn{z_{gl}} using the
barometric equation,
\deqn{ z_{{gl}}(x)=z_{ERA5}(x) + \frac{T_{ERA5}(x)}{L_b}  \left( \frac{P_{gl}}{P_{ERA5}(x)}
\right)^{\frac{RL_b}{g M}-1},}
where \eqn{z_{ERA}}, \eqn{T_{ERA}}, and \eqn{P_{ERA}} respectively correspond to the ground level
elevation, temperature at 2m, and ground level pressure of ERA5, \eqn{L_b}  is the standard
temperature lapse rate, \eqn{R} is the universal gas constant, \eqn{g} is the gravity constant
and  \eqn{M} is the molar mass of air. See more information at
\href{https://raphaelnussbaumer.com/GeoPressureAPI/#description-1}{the GeoPressureAPI documentation}.

To be able to compare the temporal variation of the retrieved pressure of ERA5 \eqn{P_{ERA}} to
the geolocator pressure \eqn{P_{gl}}, the function also returns the ERA pressure normalized with
the geolocator mean pressure measurement as \code{pressure_era5_norm}.
\deqn{ P_{ERA5,0}(\boldsymbol{x})[t] = \left( P_{ERA5}(\boldsymbol{x})[t]-P_{gl}[t]\right) -
\left( \frac{1}{n}\sum_{i=1}^{n} P_{ERA5}(\boldsymbol{x})[i]-P_{gl}[i] \right).}
}
\examples{
# Request pressure at a given location
pressurepath <- geopressure_timeseries(
  lat = 46, lon = 6,
  start_time = "2017-01-01 00:00",
  end_time = "2017-01-02 00:00",
  quiet = TRUE
)

str(pressurepath)

plot(pressurepath$date, pressurepath$pressure,
  type = "b", ylab = "Pressure (hPa)", xlab = "Datetime"
)

# Retrieve the altitude of a bird being at this location adding random noise on the sensor.
pressurepath <- geopressure_timeseries(
  lat = 46, lon = 6,
  pressure = data.frame(
    data.frame(
      date = pressurepath$date,
      value = pressurepath$pressure_era5 + rnorm(nrow(pressurepath))
    )
  ),
  quiet = TRUE
)

str(pressurepath)

plot(pressurepath$date, pressurepath$altitude,
  type = "b", ylab = "Altitude (m)", xlab = "Datetime"
)
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, and Felix Liechti. 2023.
Global Positioning with Animal‐borne Pressure Sensors. \emph{Methods in Ecology and Evolution}, 14,
1118–1129 \url{https://doi.org/10.1111/2041-210X.14043}.}
}
\seealso{
Other pressurepath: 
\code{\link{plot_pressurepath}()},
\code{\link{pressurepath2altitude}()},
\code{\link{pressurepath_create}()},
\code{\link{pressurepath_update}()}
}
\concept{pressurepath}
