% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure_timeseries.R
\name{geopressure_timeseries}
\alias{geopressure_timeseries}
\title{Download ERA5 pressure timeseries along a path}
\usage{
geopressure_timeseries(path, pressure, include_flight = FALSE, workers = 90)
}
\arguments{
\item{path}{A data.frame of the position containing latitude (\code{lat}), longitude  (\code{lon}), and the
stationary period id (\code{stap}) as columns.}

\item{pressure}{Pressure data.frame from data logger.}

\item{include_flight}{Extend request to also query the pressure and altitude during the previous
and/or next flight. Flights are defined by a \code{stap_id = 0}. Accept Logical or vector of -1 (
previous flight), 0 (stationary) and/or 1 (next flight). (e.g. \code{include_flight = c(-1, 1)} will
only search for the flight before and after but not the stationary period). Note that next and
previous flights are defined by the +/1 of the \code{stap_id} value (and not the previous/next
\code{stap_id} value).}

\item{workers}{Number of parallel requests on GEE. Between 1 and 99.}
}
\value{
A data.frame containing the row bind of the data.frames of each stationary period
computed with \code{geopressure_timeseries_latlon()}:
\itemize{
\item \code{date} same as \code{pressure$date}
\item \code{pressure_tag}: same as \code{pressure$value}
\item \code{label} same as \code{pressure$label}
\item \code{stap_id} same as \code{pressure$stap_id} (i.e., including 0)
\item \code{pressure_era5}
\item \code{altitude}
\item \code{lat}
\item \code{lon}
\item \code{pressure_era5_norm}
\item \code{stap_ref} stap_id grouping of the requests (same as \code{stap_id} except for flights where
\code{stap_id = 0} and \code{stap_id>0} is to the stationary period of the position used from the path).
}

See \code{\link[=geopressure_timeseries_latlon]{geopressure_timeseries_latlon()}} for more details
}
\description{
This function downloads the ERA5 pressure timeseries matching geolocator pressure data along a
path. The function essentially splits the pressure of the geolocator into stationary periods and
downloads the ERA5 timeseries of pressure for the corresponding location of the path by calling
\code{geopressure_timeseries_latlon()}. The computation is optimized by paralleling the call of
each stationary period.

The location of the path is matched with the geolocator pressure using \code{stap_id}.

You can include previous and/or next flight period in each query, for instance to estimate
flight altitude.

Read more about how this is computed with \code{\link[=geopressure_timeseries_latlon]{geopressure_timeseries_latlon()}}
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |>
  tag_label()

path <- data.frame(
  stap_id = tag$stap$stap_id,
  lat = c(48.5, 32.5, 30.5, 49.5, 41.6),
  lon = c(17.5, 13.5, 16.5, 21.5, 12.7)
)

path_pres <- geopressure_timeseries(path, tag$pressure)

str(path_pres)

plot(path_pres$date, path_pres$pressure_tag,
  col = path_pres$stap_id,
  ylab = "Pressure (hPa)", xlab = "Datetime"
)
lines(path_pres$date, path_pres$pressure_era5_norm, col = "red")


path_pres <- geopressure_timeseries(
  path[c(2, 3, 4), ], tag$pressure,
  include_flight = TRUE
)

plot(path_pres$date, path_pres$altitude,
  type = "l", ylab = "Pressure (hPa)", xlab = "Datetime"
)
}
\seealso{
\code{\link[=geopressure_timeseries_latlon]{geopressure_timeseries_latlon()}}, \code{\link[=geostap2path]{geostap2path()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map}
}
