% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_timeseries}
\alias{geopressure_timeseries}
\title{Request and download surface pressure timeseries at location}
\usage{
geopressure_timeseries(
  lon,
  lat,
  pressure = NULL,
  end_time = NULL,
  start_time = NULL,
  timeout = 60 * 5,
  verbose = TRUE
)
}
\arguments{
\item{lon}{Longitude to query (-180째 to 180째).}

\item{lat}{Latitude to query (0째 to 90째).}

\item{pressure}{Pressure list from data logger dataset list (optional).}

\item{end_time}{If \code{pressure} is not provided, then \code{end_time} define the ending time of
the timeseries as POSIXlt.}

\item{start_time}{If \code{pressure} is not provided, then \code{start_time} define the starting time of
the timeseries as POSIXlt.}

\item{timeout}{duration (sec) before the code is interrupted both for the request on
GeoPressureAPI and GEE. See \code{\link[httr:timeout]{httr::timeout()}}.}

\item{verbose}{Display (or not) the progress of the query (logical).}
}
\value{
A data.frame containing the timeseries of ERA5 pressure (date, pressure) as well as
longitude  and latitude (different if over water). If \code{pressure} is provided, the return
data.frame is the same as \code{pressure} with altitude and pressure0.
}
\description{
This function return the surface atmospheric pressure timeseries from ERA5 at a queried location.
}
\details{
If you supply the pressure (and time) of the geolocator \eqn{P_{gl}}, the function will
additionally return the altitude of the geolocator above sea level \eqn{z_{gl}} using the
barometric equation,
\deqn{ z_{{gl}}(x)=z_{ERA5}(x) + \frac{T_{ERA5}(x)}{L_b}  \left( \frac{P_{gl}}{P_{ERA5}(x)}
\right)^{\frac{RL_b}{g M}-1},}
where \eqn{z_{ERA}}, \eqn{T_{ERA}} and \eqn{P_{ERA}} respectively correspond to the ground level
elevation, temperature at 2m and ground level pressure of ERA5, \eqn{L_b}  is the standard
temperature lapse rate, \eqn{R} is the universal gas constant, \eqn{g} is the gravity constant
and  \eqn{M} is the molar mass of air. See more information on
\href{https://raphaelnussbaumer.com/GeoPressureAPI/#description-1}{the GeoPressureAPI documentation}.

The timeseries of the response will be on the same as time if supply, otherwise, it will return
on a hourly basis between \code{start_time} and \code{end_time}.

If the location query is over water, the location will be moved to the closest onshore location.

To be able to compare the temporal variation of the retrieved pressure of ERA5 \eqn{P_{ERA}} to
the geolocator pressure \eqn{P_{gl}}, the function also return the ERA pressure normalized with
the geolocator mean pressure measurement as \code{pressure0}.
\deqn{ P_{0}(\boldsymbol{x})[t] = \left( P_{ERA5}(\boldsymbol{x})[t]-P_{gl}[t]\right) -
\left( \frac{1}{n}\sum_{i=1}^{n} P_{ERA5}(\boldsymbol{x})[i]-P_{gl}[i] \right).}

See \href{https://raphaelnussbaumer.com/GeoPressureManual/probability-aggregation.html}{GeoPressureManual | Probability aggregation }
for more information on the meaning of this value.
}
\seealso{
\code{\link[=geopressure_timeseries_path]{geopressure_timeseries_path()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map },
}
