% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_ts}
\alias{geopressure_ts}
\title{Request and download surface pressure timeseries at location}
\usage{
geopressure_ts(
  lon,
  lat,
  pressure = NULL,
  end_time = NULL,
  start_time = NULL,
  verbose = T
)
}
\arguments{
\item{lon}{longitude to query (-180째 to 180째).}

\item{lat}{latitude to query (0째 to 90째).}

\item{pressure}{pressure list from PAM logger dataset list}

\item{end_time}{same as start_time}

\item{start_time}{if pressure not provided, then the start_time of the timeserie return is needed}

\item{verbose}{logical display progress of the query}
}
\value{
Timeserie of date, pressure, latitude, longitude and optionally altitude. Latitude and
longitude differs from the requested coordinates if over water.
}
\description{
This function return the surface atmospheric pressure timeseries from ERA5 at a queried location.
}
\details{
If you supply the pressure (and time) of the geolocator, it will additionally return the
elevation of the geolocator above sea level.

The timeserie of the response will be on the same as time if supply, otherwise, it will return
on a hourly basis between \code{start_time} and \code{end_time}.

If the location query is over water, the location will be moved to the closest onshore location.
}
\examples{
\dontrun{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
pam_data <- trainset_read(pam_data,
  pathname = system.file("extdata", package = "GeoPressureR")
)
pam_data <- pam_sta(pam_data)
pressure_timeserie[[1]] <- geopressure_ts(
  lon = 16.85,
  lat = 48.75,
  pressure = subset(pam_data$pressure, sta_id == 1)
)
}
pressure_timeserie <- readRDS(system.file("extdata", "18LX_pressure_timeserie.rda",
package = "GeoPressureR"))
par(mfrow = c(2, 1), mar = c(2, 5, 1, 1))
plot(pressure_timeserie[[1]]$date,
  pressure_timeserie[[1]]$pressure,
  ylab = "Pressure [hPa]", xlab = ""
)
plot(pressure_timeserie[[1]]$date,
  pressure_timeserie[[1]]$altitude,
  ylab = "Altitude [m asl]", xlab = ""
)
}
