% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressure.R
\name{geopressure_ts_path}
\alias{geopressure_ts_path}
\title{Query the timeserie of pressure from a path and geolocator pressure
observation}
\usage{
geopressure_ts_path(path, pressure)
}
\arguments{
\item{path}{a data.frame of the position containing latitude (\code{lat}),
longitude  (\code{lon}) and the stationay period id (\code{sta_id}) as column.}

\item{pressure}{pressure list from PAM logger dataset list}
}
\value{
list of data.frame containing for each stationay period, the date,
pressure, altitude (same as \code{geopressure_ts()}) but also sta_id, lat, lon and
pressure0, pressure normalized to match geolocator pressure measurement.
}
\description{
Query the timeserie of pressure from a path and geolocator pressure
observation
}
\examples{
\dontrun{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
pam_data <- trainset_read(pam_data,
  pathname = system.file("extdata", package = "GeoPressureR")
)
pam_data <- pam_sta(pam_data)
data("pressure_prob", package = "GeoPressureR")
path <- geopressure_map2path(pressure_prob)
pressure_timeserie <- geopressure_ts_path(path, pam_data$pressure)
}
data("pressure_timeserie", package = "GeoPressureR")
par(mfrow = c(2, 1), mar = c(2, 5, 1, 1))
plot(pressure_timeserie[[1]]$date,
  pressure_timeserie[[1]]$pressure,
  ylab = "Pressure [hPa]", xlab = ""
)
plot(pressure_timeserie[[1]]$date,
  pressure_timeserie[[1]]$altitude,
  ylab = "Altitude [m asl]", xlab = ""
)
}
