% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoPressureR-package.R
\docType{package}
\name{GeoPressureR-package}
\alias{GeoPressureR-package}
\title{GeoPressureR: Global Positioning by Atmospheric Pressure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package to determine the position and trajectory of a bird based on light-weight data-logger measuring at lease atmospheric pressure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Rafnuss/GeoPressureR}
  \item \url{https://raphaelnussbaumer.com/GeoPressureR/}
  \item Report bugs at \url{https://github.com/Rafnuss/GeoPressureR/issues}
}

}
\author{
\strong{Maintainer}: Raphaël Nussbaumer \email{rafnuss@gmail.com} (\href{https://orcid.org/0000-0002-8185-1020}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mathieu Gravey \email{mathieu@mgravey.com} (\href{https://orcid.org/0000-0002-0871-1507}{ORCID})
}

}
\keyword{internal}
