% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressuretemplate.R,
%   R/geopressuretemplate_config.R, R/geopressuretemplate_graph.R,
%   R/geopressuretemplate_pressurepath.R, R/geopressuretemplate_tag.R
\name{geopressuretemplate}
\alias{geopressuretemplate}
\alias{geopressuretemplate_config}
\alias{geopressuretemplate_graph}
\alias{geopressuretemplate_pressurepath}
\alias{geopressuretemplate_tag}
\title{Workflow for GeoPressureR}
\usage{
geopressuretemplate(
  id,
  config = config::get(config = id),
  quiet = FALSE,
  file = glue::glue("./data/interim/{id}.RData"),
  ...
)

geopressuretemplate_config(
  id,
  config = config::get(config = id),
  assert_graph = FALSE,
  ...
)

geopressuretemplate_graph(
  id,
  config = config::get(config = id),
  quiet = FALSE,
  file = glue::glue("./data/interim/{id}.RData"),
  ...
)

geopressuretemplate_pressurepath(
  id,
  config = config::get(config = id),
  quiet = FALSE,
  file = glue::glue("./data/interim/{id}.RData"),
  ...
)

geopressuretemplate_tag(
  id,
  config = config::get(config = id),
  quiet = FALSE,
  file = glue::glue("./data/interim/{id}.RData"),
  assert_graph = FALSE,
  ...
)
}
\arguments{
\item{id}{A unique identifier for the tag being processed.}

\item{config}{A configuration object specifying workflow parameters, which is loaded by default
using \code{config::get(config = id)}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses informational messages during execution. The default
value is \code{FALSE}.}

\item{file}{A file path to save the intermediate results (e.g., tag, graph, and pressure paths).
Default is \verb{./data/interim/\{id\}.RData}.}

\item{...}{Additional parameters to overwrite default or config values. Always prefer to modify
\code{config.yml} if possible.}

\item{assert_graph}{Logical. If \code{TRUE}, check that the config is compatible for the creation of
a graph. The default value is \code{TRUE}. Set to \code{FALSE} only if you don't want to create a graph
model}
}
\value{
The function returns nothing. Instead, it saves the processed outputs (tag, graph,
pressure paths, etc.) to the specified \code{file}.
}
\description{
The \code{geopressuretemplate} function manages the complete workflow for modeling bird trajectories
using geospatial pressure and light data. It includes the creation of tag data, likelihood map
computation (pressure and/or light), graph model construction, trajectory estimation, and
pressure path computation. The function follows a standard structure as defined by the GitHub
repository \href{https://github.com/Rafnuss/GeoPressureTemplate}{geopressuretemplate}, which
separates data and analysis for improved readability, sharability, and reproducibility.
}
\details{
The \code{geopressuretemplate} function is a high-level entry point that coordinates multiple
steps for processing geospatial animal movement data. It relies on underlying child
functions for each step:
\itemize{
\item \strong{Tag Creation \code{\link[=geopressuretemplate_tag]{geopressuretemplate_tag()}}}: Initializes and labels the \code{tag} object. It
also generates light and pressure likelihood maps:
}
\enumerate{
\item \code{tag_create()}: Initializes the tag object.
\item \code{tag_label()}: Adds labels.
\item \code{tag_set_map()}: Sets the spatial and temporal parameters.
\item If \code{"map_pressure"} is in the \code{config$geopressuretemplate$likelihood}:
\itemize{
\item \code{geopressure_map()}: Computes the pressure likelihood.
}
\item If \code{"map_light"} is in the \code{config$geopressuretemplate$likelihood}:
\itemize{
\item \code{twilight_create() |> twilight_read() |> geolight_map()}: Computes the light likelihood.
}
}
\itemize{
\item \strong{Graph Creation \code{\link[=geopressuretemplate_graph]{geopressuretemplate_graph()}}}: Builds a movement model graph
based \code{tag} data, and can include wind effects if specified. Outputs such as
marginal distributions, most likely paths, and simulation paths can be computed.
}
\enumerate{
\item \code{graph_create()}: Creates the graph based on tag.
\item If \code{config$graph_set_movement$type == "gs"} (i.e., no wind):
\itemize{
\item \code{graph_set_movement()}: Sets the movement model without wind.
}
\item If \code{config$graph_set_movement$type == "as"} (i.e., with wind):
\itemize{
\item \code{graph_add_wind()}: Adds wind data to the graph.
\item \code{graph_set_movement()}: Sets the movement model with wind.
}
\item If \code{"marginal"} is in \code{config$geopressuretemplate$outputs}:
\itemize{
\item \code{graph_marginal()}: Computes the marginal distribution map.
}
\item If \code{"most_likely"} is in \code{config$geopressuretemplate$outputs}:
\itemize{
\item \code{graph_most_likely()}: Computes the most likely path based on the movement model.
}
\item If \code{"simulation"} is in \code{config$geopressuretemplate$outputs}:
\itemize{
\item \code{graph_simulation()}: Runs simulations to model multiple possible paths.
}
\item \code{save()}: Saves the computed graph and associated objects in \verb{data/interim/\{id\}.Rdata}
}
\itemize{
\item \strong{Pressure Path Processing \code{\link[=geopressuretemplate_pressurepath]{geopressuretemplate_pressurepath()}}}: Computes pressurepaths
(\code{pressurepath_create}) using the content of the \code{Rdata} file and appending the pressurepath
data.frame to the same file.
}
\enumerate{
\item If \code{"most_likely"} is in \code{config$geopressuretemplate$pressurepath}, computes the pressure path
for \code{path_most_likely}.
\item If \code{"geopressureviz"} is in \code{config$geopressuretemplate$pressurepath} computes the pressure
path for \code{path_geopressureviz}
}

Each of these child functions can be called individually or automatically as part of
the \code{geopressuretemplate} workflow.
}
\examples{
\dontrun{
# Run the complete geopressuretemplate workflow
geopressuretemplate("18LX", quiet = TRUE)
}

}
\concept{geopressuretemplate}
