% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressureviz.R
\name{geopressureviz}
\alias{geopressureviz}
\title{Start the GeoPressureViz shiny app}
\usage{
geopressureviz(
  x,
  path = NULL,
  marginal = NULL,
  launch_browser = TRUE,
  run_bg = TRUE
)
}
\arguments{
\item{x}{a GeoPressureR \code{tag} object, a \code{.Rdata} file or the
unique identifier \code{id} with a \code{.Rdata} file located in \code{"./data/interim/{id}.RData"}.}

\item{path}{a GeoPressureR \code{path} or \code{pressurepath} data.frame.}

\item{marginal}{map of the marginal probability computed with \code{graph_marginal()}. Overwrite the
\code{path} or \code{pressurepath} contained in the \code{.Rdata} file.}

\item{launch_browser}{If true (by default), the app runs in your browser, otherwise it runs on
Rstudio.}

\item{run_bg}{If true, the app runs in a background R session using the \code{callr} package. This
allows you to continue using your R session while the app is running.}
}
\value{
When \code{run_bg = FALSE}: The updated path visualized in the app. Can also be retrieved with
\code{shiny::getShinyOption("path_geopressureviz")} after the app completes.
When \code{run_bg = TRUE}: Returns the background process object.
}
\description{
GeoPressureViz is a shiny app designed to help you visualize the overall trajectory of the bird
as well as each step-by-step move. This app is particularly useful to check the correspondence
between pressure map, light map and flight distance. You can edit the path and query pressure
time series for individual stationary period to test manual what seems the optimal path.
}
\details{
GeoPressureViz can be started based on a \code{.Rdata} file containing at least \code{tag}, but also
optionally \code{marginal} and/or \code{path} (\code{path_most_likely} is also accepted).

You can retrieve the edited path from the return value of this function (when \code{run_bg = FALSE})
or with \code{shiny::getShinyOption("path_geopressureviz")} after the app completes.

Learn more about GeoPressureViz in the \href{https://raphaelnussbaumer.com/GeoPressureManual/geopressureviz.html}{GeoPressureManual } or with
this \href{https://rafnuss.shinyapps.io/GeoPressureViz/}{demo of the Great Reed Warbler (18LX)}.
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/geopressureviz.html}{GeoPressureManual }
}
