% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geopressureviz.R
\name{geopressureviz}
\alias{geopressureviz}
\title{Start the GeoPressureViz shiny app}
\usage{
geopressureviz(geostap, path_pres = NA, launch_browser = TRUE)
}
\arguments{
\item{geostap}{List of the geographical and stationary period information. See
\code{\link[=geostap_create]{geostap_create()}} for details.}

\item{path_pres}{geopressure_timeseries}

\item{launch_browser}{If true (by default), the app runs in your browser, otherwise it runs on Rstudio.}
}
\value{
The updated path visualized in the app.
}
\description{
GeoPressureViz is a shiny app designed to help you visualize the overall trajectory of the bird
as well as each step-by-step move. Learn more about GeoPressureViz in the \href{https://raphaelnussbaumer.com/GeoPressureManual/geopressureviz.html}{GeoPressureManual | GeoPressureViz} or with
this \href{https://rafnuss.shinyapps.io/GeoPressureViz/}{demo of the Great Reed Warbler (18LX)}.
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/geopressureviz.html}{GeoPressureManual | GeoPressureViz }
}
