% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_add_wind.R
\name{graph_add_wind}
\alias{graph_add_wind}
\title{Add windspeed and airspeed}
\usage{
graph_add_wind(
  graph,
  pressure,
  directory = glue::glue("data/5_wind_graph/{graph$id}/"),
  thr_as = Inf
)
}
\arguments{
\item{graph}{graph constructed with \code{\link[=graph_create]{graph_create()}}}

\item{pressure}{pressure data from a data logger. This data.frame needs to contain \code{date} as
POSIXt and \code{value} in hPa. It is good practice to use \code{\link[=tag_read]{tag_read()}} and \code{\link[=tag_label_stap]{tag_label_stap()}} to build
this data.frame.}

\item{directory}{Character of the path where the netCDF files can be found.}

\item{thr_as}{Threshold of airspeed (km/h).}
}
\value{
Graph as a list with windspeed and airspeed as \code{ws} and \code{as} respectively (see
\code{\link[=graph_create]{graph_create()}} for more details on the graph returned).
}
\description{
Reads the NetCDF files downloaded in \code{directory} and adds the average windspeed experienced by the
bird, and the corresponding airspeed for each edge of the graph.

See the \href{https://raphaelnussbaumer.com/GeoPressureManual/wind-graph.html#download-wind-data}{GeoPressureManual | Wind graph} for
explanations and an example on how to download the \code{NetCDF} files from ERA-5.
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/wind-graph.html#add-wind-to-graph}{GeoPressureManual | Wind graph}
}
