% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_add_wind.R
\name{graph_add_wind}
\alias{graph_add_wind}
\title{Compute windspeed and airspeed on a \code{graph}}
\usage{
graph_add_wind(
  graph,
  pressure,
  thr_as = Inf,
  file = function(stap_id)
    glue::glue("./data/wind/{graph$param$id}/{graph$param$id}_{stap_id}.nc"),
  quiet = FALSE
)
}
\arguments{
\item{graph}{a GeoPressureR graph object.}

\item{pressure}{pressure measurement of the associated \code{tag} data used to estimate the pressure
level (i.e., altitude) of the bird during the flights. This data.frame needs to contain \code{date} as
POSIXt and \code{value} in hPa.}

\item{thr_as}{threshold of airspeed (km/h).}

\item{file}{Absolute or relative path of the ERA5 wind data file to be downloaded. Function
taking as single argument the stationary period identifier.}

\item{quiet}{logical to hide messages about the progress}
}
\value{
a \code{graph} object with windspeed and airspeed as \code{ws} and \code{as} respectively.
}
\description{
Reads the NetCDF files downloaded and interpolate the average windspeed experienced by the
bird on each possible edge, as well as the corresponding airspeed.

In addition, the graph can be further pruned based on a threshold of airspeed \code{thr_as}.

See section \href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.14082#mee314082-sec-0008-title}{2.2.4 in Nussbaumer (2023b)}
for more technical details and the \href{https://raphaelnussbaumer.com/GeoPressureManual/trajectory-with-wind.html}{GeoPressureManual} for an
illustration on how to use it.
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/trajectory-with-wind.html}{GeoPressureManual}

Other graph: 
\code{\link{graph_create}()},
\code{\link{graph_marginal}()},
\code{\link{graph_most_likely}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_simulation}()},
\code{\link{print.graph}()}

Other movement: 
\code{\link{bird_create}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_transition}()},
\code{\link{plot_graph_movement}()},
\code{\link{speed2prob}()},
\code{\link{tag_download_wind}()}
}
\concept{graph}
\concept{movement}
