% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_create}
\alias{graph_create}
\title{Create graph}
\usage{
graph_create(static_prob, thr_prob_percentile = 0.99, thr_gs = 150)
}
\arguments{
\item{static_prob}{list of raster containing probability map of each
stationary period. The metadata of \code{static_prob} needs to include the flight
duration to the next stationary period in the variable
\code{next_flight_duration} as a numeric in hours.}

\item{thr_prob_percentile}{threshold of percentile (see explanation above)}

\item{thr_gs}{threashold of groundspeed (km/h)  (see explanation above)}
}
\value{
graph as a list (see description above)
}
\description{
This function return a graph representing the trajectory of a bird based on
filtering and triming the probability maps provided.
}
\details{
In the final graph, we only keep the most likely node (position in time)
defined as:
\enumerate{
\item those which cumulative probability reach up to \code{thr_prob_percentile} for
each stationary period.
\item those which average ground speed is lower than \code{thr_gs} km/h.
}

The graph returned is a list of the edges of the graph containing:
\itemize{
\item \code{s}: source node (index in the 3d grid lat-lon-sta),
\item \code{t}: target node (index in the 3d grid lat-lon-sta),
\item \code{gs}: average ground speed required to make that transition (km/h)
\item \code{ps}: static probability of each target node
\item \code{sz}: size of the 3d grid lat-lon-sta
\item \code{equipement}: node(s) of the first sta (index in the 3d grid lat-lon-sta)
\item \code{retrival}: node(s) of the last sta (index in the 3d grid lat-lon-sta)
\item \code{flight_duration}: list of flight duration to next sta in hours
\item \code{lat}: list of the \code{static_prob} latitude in cell center
\item \code{lon}: list of the \code{static_prob} longitude in cell center
\item \code{extent}: raster geographical extent of the `static_prob``
\item \code{resolution}: raster res of the \code{static_prob}
\item \code{temporal_extent}: start and end date time retrieved from the metadata of
\code{static_prob}
}

The vignette \verb{How to use the graph} provided an example how to prepare the
data for the function and the output of this function
}
