% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_create}
\alias{graph_create}
\title{Create graph}
\usage{
graph_create(likelihood, thr_prob_percentile = 0.99, thr_gs = 150)
}
\arguments{
\item{likelihood}{List of likelihood map of each stationary period. The
metadata of \code{likelihood} needs to include the flight information to the next stationary period
in \code{likelihood[[1]]$flight}.}

\item{thr_prob_percentile}{Threshold of percentile (see details).}

\item{thr_gs}{Threshold of groundspeed (km/h)  (see details).}
}
\value{
Graph as a list (see details).
}
\description{
This function return a graph representing the trajectory of a bird based on filtering and triming
the probability maps provided.
}
\details{
In the final graph, we only keep the most likely node (position in time) defined as: 1. those
which cumulative probability reach up to \code{thr_prob_percentile} for each stationary period. 2.
those which average ground speed is lower than \code{thr_gs} km/h.

The graph returned is a list of the edges of the graph containing:
\itemize{
\item \code{s}: source node (index in the 3d grid lat-lon-sta),
\item \code{t}: target node (index in the 3d grid lat-lon-sta),
\item \code{gs}: average ground speed required to make that transition (km/h) as complex number
representing the E-W as real and S-N as imaginary.
\item \code{ps}: static probability of each target node,
\item \code{sz}: size of the 3d grid lat-lon-sta,
\item \code{equipment}: node(s) of the first stap (index in the 3d grid lat-lon-sta),
\item \code{retrieval}: node(s) of the last stap (index in the 3d grid lat-lon-sta),
\item \code{flight_duration}: list of flight duration to next stap in hours,
\item \code{lat}: list of the \code{likelihood$map} latitude in cell center,
\item \code{lon}: list of the \code{likelihood$map} longitude in cell center,
\item \code{extent}: SpatRaster geographical extent of \code{likelihood$map},
\item \code{resolution}: resolution of \code{likelihood$map},
\item \code{temporal_extent}: start and end date time retrieved from \code{likelihood}.
}

The \href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#create-the-graph}{GeoPressureManual | Basic graph} provided an
example how to prepare the
data for the function and the output of this function.
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#create-the-graph}{GeoPressureManual | Basic graph}
}
