% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_create.R
\name{graph_create}
\alias{graph_create}
\title{Create graph}
\usage{
graph_create(geostap, thr_likelihood = 0.99, thr_gs = 150, likelihood = NA)
}
\arguments{
\item{geostap}{List of the geographical and stationary period information. See
\code{\link[=geostap_create]{geostap_create()}} for details.}

\item{thr_likelihood}{Threshold of percentile (see details).}

\item{thr_gs}{Threshold of groundspeed (km/h)  (see details).}

\item{likelihood}{Field of the \code{geostap} list containing the likelihood map (character). Default
\code{NA} is to take the product of \code{map_pressure} and \code{map_light} if available.}
}
\value{
Graph as a list
\itemize{
\item \code{s}:   source node (index in the 3d grid lat-lon-stap)
\item \code{t}:   target node (index in the 3d grid lat-lon-stap)
\item \code{gs}:  average ground speed required to make that transition (km/h) as complex number
representing the E-W as real and S-N as imaginary
\item \code{obs}:   observation model, corresponding to the normalized likelihood in a 3D matrix of size
\code{sz}
\item \code{sz}:  size of the 3d grid lat-lon-stap
\item \code{stap}: data.frame of all stationary periods
\item \code{equipment}: node(s) of the first stap (index in the 3d grid lat-lon-sta)
\item \code{retrieval}: node(s) of the last stap (index in the 3d grid lat-lon-sta)
\item \code{extent}: same as \code{geostap$extent}
\item \code{scale}: same as \code{geostap$scale}
\item \code{mask_water}: logical matrix of water-land
\item \code{param}: parameter used to create the graph, including \code{thr_likelihood} and \code{thr_gs}
}
}
\description{
This function returns a graph representing the trajectory of a bird based on filtering and
trimming the likelihood maps provided.

In the final graph, we only keep the most likely nodes (position in time) defined as (1) those
whose cumulative probability reaches up to \code{thr_likelihood} for each stationary period and (2)
those which are connected to edges whose average ground speed is lower than \code{thr_gs} km/h.
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. “Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model.” \emph{Methods in Ecology and Evolution}.
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#create-the-graph}{GeoPressureManual | Basic graph}
}
