% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_download_wind}
\alias{graph_download_wind}
\title{Download wind data}
\usage{
graph_download_wind(
  pam,
  area,
  sta_id = seq_len(nrow(pam$sta) - 1),
  cds_key = Sys.getenv("cds_key"),
  cds_user = Sys.getenv("cds_user"),
  path = paste0("data/5_wind_graph/", pam$id, "/")
)
}
\arguments{
\item{pam}{PAM logger dataset list with \code{pam$sta} computed. See \code{\link[=pam_read]{pam_read()}} and \code{\link[=pam_sta]{pam_sta()}}.}

\item{area}{Geographical extent of the map to query. Either a raster (e.g. \code{static_prob}) or a
list ordered by North, West, South, East  (e.g. \code{c(50,-16,0,20)}).}

\item{sta_id}{stationary period identifier of the \emph{start} of the flight to query as defined in
\code{pam$sta}.}

\item{cds_key}{user (email address) used to sign up for the ECMWF data service. See
\code{\link[=wf_set_key]{wf_set_key()}}.}

\item{cds_user}{token provided by ECMWF. See \code{\link[=wf_set_key]{wf_set_key()}}.}

\item{path}{path were to store the downloaded data.}
}
\value{
the path of the downloaded (requested file)
}
\description{
This function wrap the function Wind data is available at high resolution (1hr, 0.25°, 37 pressure level) on ERA5 hourly data on pressure levels. And this data is easily accessible through the ecmwfr package. As the flight are of short duration, we suggest to download a file for each flight.
}
\details{
area

Be default download data for all the flight, thus sta_id is all exetpt the last one (no flight starting from the last stationary period)

cds_key and cds_user
}
\seealso{
\code{\link[=wf_request]{wf_request()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/wind-graph.html#download-wind-data}{GeoPressureManual | Wind graph }
}
