% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_product.R
\name{graph_marginal}
\alias{graph_marginal}
\title{Marginal probability map}
\usage{
graph_marginal(graph)
}
\arguments{
\item{graph}{graph constructed with \code{\link[=graph_create]{graph_create()}}.}
}
\value{
A list of the marginal maps for each stationary period (even those not modelled). Best to
include within \code{geostap}.
}
\description{
Compute the marginal probability map from a graph. The graph needs to have a movement model
defined (see \code{graph_add_movement()}). The computation uses the \href{https://en.wikipedia.org/wiki/Forward\%E2\%80\%93backward_algorithm}{forward backward algorithm}.
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. “Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model.” \emph{Methods in Ecology and Evolution}.
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \code{\link[=graph_add_movement]{graph_add_movement()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-2-marginal-probability-map}{GeoPressureManual | Basic graph}
}
