% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_marginal}
\alias{graph_marginal}
\title{Marginal Probability Map}
\usage{
graph_marginal(grl)
}
\arguments{
\item{grl}{graph constructed with \code{\link[=graph_create]{graph_create()}}}
}
\value{
list of map of the marginal probability at each stationary period
}
\description{
This function return the marginal probability map as raster from a graph.
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-2-marginal-probability-map}{GeoPressureManual | Basic graph}
}
