% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_marginal.R
\name{graph_marginal}
\alias{graph_marginal}
\title{Compute marginal probability map}
\usage{
graph_marginal(graph)
}
\arguments{
\item{graph}{a GeoPressureR \code{graph} with defined movement model \code{graph_set_movement()}.}
}
\value{
A list of the marginal maps for each stationary period (even those not modelled). Best to
include within \code{tag}.
}
\description{
Compute the marginal probability map from a graph. The computation uses the \href{https://en.wikipedia.org/wiki/Forward\%E2\%80\%93backward_algorithm}{forward backward algorithm}. For more
details, see \href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.14082#mee314082-sec-0012-title}{section 2.3.2 of Nussbaumer et al. (2023b)}
and the \href{https://bit.ly/3sd20vC}{GeoPressureManual}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  twilight_create() |>
  twilight_label_read() |>
  tag_set_map(
    extent = c(-16, 23, 0, 50),
    known = data.frame(stap_id = 1, known_lon = 17.05, known_lat = 48.9)
  ) |>
  geopressure_map(quiet = TRUE) |>
  geolight_map(quiet = TRUE)

# Create graph
graph <- graph_create(tag, quiet = TRUE)

# Define movement model
graph <- graph_set_movement(graph)

# Compute marginal
marginal <- graph_marginal(graph)

plot(marginal)

}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\href{https://bit.ly/3sd20vC}{GeoPressureManual}

Other graph: 
\code{\link{graph_add_wind}()},
\code{\link{graph_create}()},
\code{\link{graph_most_likely}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_simulation}()},
\code{\link{print.graph}()}
}
\concept{graph}
