% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_most_likely.R
\name{graph_most_likely}
\alias{graph_most_likely}
\title{Compute the most likely trajectory}
\usage{
graph_most_likely(graph)
}
\arguments{
\item{graph}{a graph object.}
}
\value{
Path data.frame containing the columns
-\code{stap_id} stationary period
\itemize{
\item \code{j} unique ID for each path, here always 1 as there is a single path.
\item \code{ind} indices of the coordinate in the 2D grid. Useful to retrieve map or graph information
\item \code{lat} latitude,
\item \code{lon} longitude
\item \code{start} datetime of the start of the stationary period (same as in \code{stap})
\item \code{end} datetime of the end of the stationary period (same as in \code{stap})
\item \code{include} logical if stationary period was modeled (same as in \code{stap})
\item \verb{nb_sample known} number of datapoint used to compute pressure (same as in \code{stap})
}
}
\description{
Compute the trajectory which maximizes the joint probability using the \href{https://en.wikipedia.org/wiki/Viterbi_algorithm}{Viterbi algorithm} on the graph structure. See section 2.3.1 in
Nussbaumer (2023) for more information.
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-3-simulate-path}{GeoPressureManual | Basic graph}

Other graph: 
\code{\link{graph_add_wind}()},
\code{\link{graph_create}()},
\code{\link{graph_marginal}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_simulation}()},
\code{\link{print.graph}()}
}
\concept{graph}
