% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_product.R
\name{graph_most_likely}
\alias{graph_most_likely}
\title{Most likely trajectory}
\usage{
graph_most_likely(graph)
}
\arguments{
\item{graph}{Graph constructed with \code{\link[=graph_create]{graph_create()}}.}
}
\value{
Path data.frame. See \code{\link[=ind2path]{ind2path()}} for the structure.
}
\description{
Compute the trajectory which maximizes the overall probability using the \href{https://en.wikipedia.org/wiki/Viterbi_algorithm}{Viterbi algorithm} on the graph structure. The graph needs to have
a movement model defined (see \code{graph_add_movement()}).
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. “Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model.” \emph{Methods in Ecology and Evolution}.
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-3-simulate-path}{GeoPressureManual | Basic graph}
}
