% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_path2edge}
\alias{graph_path2edge}
\title{Find the edge index from a path index}
\usage{
graph_path2edge(path_id, grl)
}
\arguments{
\item{path_id}{List or matrix of node index (\verb{nj x nsta}).}

\item{grl}{Graph constructed with \code{\link[=graph_create]{graph_create()}}.}
}
\value{
List or matrix of the edge \verb{nj x (nsta-1)}.
}
\description{
Very inefficient way to find the edges...
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureR/articles/basic-graph.html}{Vignette Basic Graph }
}
