% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_path2lonlat}
\alias{graph_path2lonlat}
\title{Find the latitude and longitude from a path index}
\usage{
graph_path2lonlat(path_id, grl)
}
\arguments{
\item{path_id}{List or matrix of node index.}

\item{grl}{Graph constructed with \code{\link[=graph_create]{graph_create()}}.}
}
\value{
List of the path with latitude and longitude and index fo the the path provided.
}
\description{
Find the latitude and longitude from a path index
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-1-shortest-path}{GeoPressureManual | Basic graph}
}
