% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_set_movement.R
\name{graph_set_movement}
\alias{graph_set_movement}
\title{Define the movement model of a \code{graph}}
\usage{
graph_set_movement(
  graph,
  type = ifelse("ws" \%in\% names(graph), "as", "gs"),
  method = ifelse("ws" \%in\% names(graph), "power", "gamma"),
  shape = 7,
  scale = 7,
  location = 40,
  bird = NULL,
  power2prob = function(power) (1/power)^3,
  low_speed_fix = 15
)
}
\arguments{
\item{graph}{a GeoPressureR \code{graph} object.}

\item{type}{Ground speed \code{"gs"} or airspeed \code{"as"}}

\item{method}{method used to convert the speed to probability ("gamma", "logis" or "power")}

\item{shape}{parameter of the gamma distribution}

\item{scale}{parameter of the gamma and logistic distribution}

\item{location}{parameter for the logistic distribution}

\item{bird}{A GeoPressureR \code{bird} object containing the basic morphological traits necessary:
mass, wing span, wing aspect ratio, and body frontal area. See \code{bird_create()}.}

\item{power2prob}{function taking power as a single argument and returning a probability}

\item{low_speed_fix}{speed below which the probability remains the same. This parameter is used
to allow short flights covering small distances.}
}
\value{
Graph list with a new list \code{graph$movement} storing all the parameters needed to compute
the transition probability
}
\description{
Configure the movement model of a \code{graph} by defining the value of the parameters needed to build
the transition \code{graph_transition()} through \code{speed2prob()}.

Three methods are currently implemented with two parametric function \code{"gamma"} and \code{"logis"}
suitable when wind data is not available and are thus defining the probability of a groundspeed.

If wind data is available, it is recommended to use the \code{"power"} method which rely on the power
curve equation (energy vs airspeed) to estimate the probability of a airspeed. Read more about
this approach in \href{https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.14082#mee314082-sec-0009-title}{section 2.2.5. of Nussbaumer et al. (2023b)}
}
\examples{
owd <- setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  twilight_create() |>
  twilight_label_read() |>
  tag_set_map(
    extent = c(-16, 23, 0, 50),
    known = data.frame(stap_id = 1, known_lon = 17.05, known_lat = 48.9)
  ) |>
  geopressure_map(quiet = TRUE) |>
  geolight_map(quiet = TRUE)
setwd(owd)

graph <- graph_create(tag, quiet = TRUE)

graph <- graph_set_movement(graph,
  method = "gamma",
  shape = 4,
  scale = 6,
  low_speed_fix = 10
)
plot_graph_movement(graph)

graph <- graph_set_movement(graph,
  method = "logis",
  shape = 4,
  location = 60,
  low_speed_fix = 10
)
plot_graph_movement(graph)

}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model. \emph{Methods in Ecology and Evolution}, 14, 1118–1129
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
Other graph: 
\code{\link{graph_add_wind}()},
\code{\link{graph_create}()},
\code{\link{graph_marginal}()},
\code{\link{graph_most_likely}()},
\code{\link{graph_simulation}()},
\code{\link{print.graph}()}

Other movement: 
\code{\link{bird_create}()},
\code{\link{graph_add_wind}()},
\code{\link{graph_transition}()},
\code{\link{plot_graph_movement}()},
\code{\link{speed2prob}()},
\code{\link{tag_download_wind}()}
}
\concept{graph}
\concept{movement}
