% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_product.R
\name{graph_simulation}
\alias{graph_simulation}
\title{Simulation of trajectories}
\usage{
graph_simulation(graph, nj = 10)
}
\arguments{
\item{graph}{Graph constructed with \code{\link[=graph_create]{graph_create()}}.}

\item{nj}{Number of simulations.}
}
\value{
Data.frame of simulated paths. See \code{\link[=ind2path]{ind2path()}} for the structure.
}
\description{
This function randomly simulates multiple trajectories from a graph using the forward filtering
backward sampling algorithm.
}
\references{
{ Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, Felix Liechti, and Daniel
Sheldon. 2023. “Reconstructing bird trajectories from pressure and wind data using a highly
optimized hidden Markov model.” \emph{Methods in Ecology and Evolution}.
\url{https://doi.org/10.1111/2041-210X.14082}.}
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-3-simulate-path}{GeoPressureManual | Basic graph}
}
