% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_simulation}
\alias{graph_simulation}
\title{Simulation of trajectory}
\usage{
graph_simulation(grl, nj = 10)
}
\arguments{
\item{grl}{graph constructed with \code{geopressure_graph_create()}}

\item{nj}{number of simulation}
}
\value{
list of the simulated path
}
\description{
This function generate simulated path from a graph
}
