% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_simulation}
\alias{graph_simulation}
\title{Simulation of trajectory}
\usage{
graph_simulation(grl, nj = 10)
}
\arguments{
\item{grl}{Graph constructed with \code{\link[=graph_create]{graph_create()}}.}

\item{nj}{Number of simulation.}
}
\value{
List of simulated paths.
}
\description{
This function simulates multiple trajectory fomr a graph. The trajectories consist of the
positions at each stationary periods.
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/basic-graph.html#output-3-simulate-path}{GeoPressureManual | Basic graph}
}
