% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grl}
\alias{grl}
\title{Wind graph for 18LX}
\format{
An object of class \code{list} of length 16.
}
\usage{
grl
}
\description{
This dataset was generated with the vignette
\href{articles/wind-graph.html}{Improving the graph with wind}. \code{grl} represents
the graph as described in \code{graph_create()} and \code{graph_add_wind()}
}
\keyword{datasets}
