% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind2path.R
\name{ind2path}
\alias{ind2path}
\title{Create a \code{path} from indices of coordinates}
\usage{
ind2path(ind, tag_graph, use_known = TRUE)
}
\arguments{
\item{ind}{vector of indices of the 2D map (vector or matrix)}

\item{tag_graph}{either a \code{tag} or a \code{graph} object, must contains \code{stap}, \code{extent} and \code{scale}.}

\item{use_known}{If true, enforce the known position in the path created. The known positions are
approximated to the map resolution in order to corresponds to integer index.}
}
\value{
A path data.frame
\itemize{
\item \code{stap_id} stationary period identification
\item \code{j} identification for each trajectory (\code{1} to \code{nj}).
\item \code{lat} Latitude,
\item \code{lon} longitude
\item \code{...} other columns from \code{stap}
}
}
\description{
Convert a vector of 2D coordinate indices into a \code{path} data.frame. If \code{ind} is a matrix,
each row represents the path of a different trajectory \code{j}.

This function is the main way to create \code{path} in GeoPressureR, and is used by
\code{\link{graph_most_likely}}, \code{\link{graph_simulation}} and \code{\link{tag2path}}.
}
\examples{
setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  tag_set_map(c(-16, 23, 0, 50), scale = 1)

# ind generated by `graph_most_likely` (single trajectory - vector)
ind <- c(1652, 1603, 1755, 1708, 1607)
path <- ind2path(ind, tag)
knitr::kable(path)

# ind generated by `graph_simulation` (10 trajectories - matrix)
ind <- matrix(c(
  1652, 1652, 1652, 1652, 1652, 1652, 1652, 1652, 1652, 1652,
  1653, 1606, 1653, 1504, 1604, 1704, 1504, 1653, 1753, 1701,
  1805, 1457, 1408, 1657, 1609, 1903, 1506, 1757, 1856, 1804,
  1607, 1514, 1611, 1457, 1860, 1802, 1759, 1609, 1760, 1208,
  1505, 1314, 1559, 1357, 1758, 1852, 1661, 1706, 1862, 1358
), nrow = 10)
path <- ind2path(ind, tag)
knitr::kable(head(path, n = 15))

}
\seealso{
Other path: 
\code{\link{path2edge}()},
\code{\link{path2elevation}()},
\code{\link{plot_path}()}
}
\concept{path}
