% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_interim.R
\name{load_interim}
\alias{load_interim}
\title{Load an interim RData object}
\usage{
load_interim(id, envir = parent.frame(), verbose = FALSE)
}
\arguments{
\item{id}{A character string of length 1 corresponding to the tag identifier used in the
GeoPressureTemplate (e.g., \code{"18LX"}). The function will look for a file named
\verb{./data/interim/\{id\}.RData}.}

\item{envir}{the environment where the data should be loaded.}

\item{verbose}{should item names be printed during loading?}
}
\value{
Invisibly returns the names of the objects loaded (as in \code{\link[base:load]{base::load()}}).
}
\description{
Loads an object from the \verb{./data/interim/} directory created during the GeoPressure analysis.
This is a convenience function to quickly restore saved objects (e.g., \code{tag}, \code{graph},
\code{path_most_likely}) for a given tag \code{id}.
}
