% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map2path}
\alias{map2path}
\title{Return the most likely path from a likelihood map}
\usage{
map2path(likelihood, interp = 0, format = "lonlat")
}
\arguments{
\item{likelihood}{List of likelihood map of each stationary period. See
\code{\link[=geopressure_likelihood]{geopressure_likelihood()}} or \code{\link[=geolight_likelihood]{geolight_likelihood()}}.}

\item{interp}{The position of the stationary period shorter than \code{interp} will be
replace by a linear average from other position (in days) .}

\item{format}{One of \code{"lonlat"}, \code{"ind"}, \code{"arr.ind"}. return the path in lon-lat or indices}
}
\value{
a data.frame of the position containing latitude (\code{lat}), longitude (\code{lon}) and the
stationary period id (\code{stap}) as column. Optionally, if indexes were requested, it will be
return. You will need to use \code{which.max(as.matrix(map))} and not \code{which.max(map)} to get
the correct location.
}
\description{
Find the latitude and longitude of the highest value in wach map from a list of likelihood map:
\code{pressure_likelihood} or \code{light_likelihood}.
}
\details{
\code{interp} can be used to interpolate unrealistic position from short stationary period based on
the position of the longer ones. The interpolation assumes that the first and last stationary
period can be safely estimated from the probability map.
}
\examples{
# See `geopressure_likelihood()` for generating pressure_likelihood
pressure_likelihood_1 <- readRDS(system.file("extdata/1_pressure/",
  "18LX_pressure_likelihood_1.rda",
  package = "GeoPressureR"
))
map2path(list(pressure_likelihood_1))
}
\seealso{
\code{\link[=geopressure_likelihood]{geopressure_likelihood()}}, \code{\link[=geopressure_timeseries_path]{geopressure_timeseries_path()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#compute-altitude}{GeoPressureManual | Pressure Map }
}
