% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_create.R
\name{map_create}
\alias{map_create}
\title{Create a \code{map} object}
\usage{
map_create(data, extent, scale, stap, id = NA, type = "unknown")
}
\arguments{
\item{data}{list of matrices of the same size, one for each stationary period.}

\item{extent}{geographical extent of the map on which the likelihood and graph model will be
computed. Vector of length 4 \code{c(xmin, xmax, ymin, ymax)} or \code{c(W, E, S, N)}.}

\item{scale}{number of pixels per 1° latitude-longitude. For instance, \code{scale = 10} for a
resolution of 0.1° (~10km) and \code{scale=4} for a resolution of 0.25° (~30km). To avoid
interpolating the ERA5 data, the scale should be equal to or smaller than 10. Read more about
scale on the \href{https://developers.google.com/earth-engine/guides/scale}{Google earth Engine documentation }.}

\item{stap}{a data.frame of stationary periods.}

\item{id}{tag identifier.}

\item{type}{map type used for default palettes in plots. One of:
\code{"unknown"}, \code{"pressure"}, \code{"light"}, \code{"magnetic"}, \code{"pressure_mse"}, \code{"pressure_mask"},
\code{"mask_water"}, \code{"magnetic_intensity"}, \code{"magnetic_inclination"}, \code{"marginal"}, \code{"twilight"}.}
}
\value{
A GeoPressure \code{map} object
}
\description{
This function creates a GeoPressureR \code{map} object from a spatio-temporal dataset.
The data needs to be discretized according to \code{scale}, \code{extent} (space) and \code{stap} (time).

This function is used by \code{geopressure_map} and \code{graph_marginal}.
}
\examples{
data <- lapply(1:10, \(x) matrix(runif(5000), nrow = 50, ncol = 100))
scale <- 10
extent <- c(0, 10, 0, 5)
seq(as.Date("2023-01-01"), as.Date("2023-01-10"), by = "day")
stap <- data.frame(
  stap_id = 1:10,
  start = seq(as.POSIXct("2023-01-01", tz = "UTC"),
    as.POSIXct("2023-01-10 UTC", tz = "UTC"),
    by = "day"
  ),
  include = TRUE
)
stap$end <- stap$start + sample(1:10) * 10000

# Create the map
map <- map_create(
  data = data,
  extent = extent,
  scale = scale,
  stap = stap,
  id = "18LX",
  type = "pressure"
)

print(map)

plot(map)

}
\seealso{
Other map:
\code{\link[=map_add_mask_water]{map_add_mask_water()}},
\code{\link[=print.map]{print.map()}},
\code{\link[=rast.map]{rast.map()}}
}
\concept{map}
