% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_classify}
\alias{pam_classify}
\title{Automatic classification of pam}
\usage{
pam_classify(pam, min_duration = 30)
}
\arguments{
\item{pam}{logger dataset list. See \code{\link[=pam_read]{pam_read()}}.}

\item{min_duration}{duration in minutes}
}
\value{
pam
}
\description{
This function uses activity data to classify migratory flapping flight. It returns the same data
list \code{pam} adding a column \code{ismig} to the data.frame \code{acceleration}. This fonction is inspired by
the function \code{classify_flap} from the \href{https://github.com/KiranLDA/pamlr}{PAMLr package}.
}
\examples{
pam <- pam_read(
  pathname = system.file("extdata/0_PAM/18LX", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
pam <- pam_classify(pam, min_duration = 30)
head(pam$acceleration)
}
\seealso{
\code{\link[=pam_read]{pam_read()}}, \href{https://kiranlda.github.io/PAMLrManual/flapping.html}{flapping chapter of the PAMLr manual}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#automatic-classification-of-activity}{GeoPressureManual | Pressure Map }
}
