% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_classify}
\alias{pam_classify}
\title{Automatic classification of pam}
\usage{
pam_classify(pam, min_duration = 30)
}
\arguments{
\item{pam}{data list}

\item{min_duration}{duration in minutes}
}
\value{
pam
}
\description{
This function uses activity data to classify migratory flapping flight. It
returns the same dats list \code{pam} adding a column \code{class} to the data.frame
\code{acceleration}.
}
\details{
This fonction is inspired by the function \code{classify_flap} from the
\href{https://github.com/KiranLDA/pamLr}{pamLr package}.
}
\examples{
pam_data = pam_read(
   pathname = system.file("extdata", package = "GeoPressureR"),
   crop_start = "2017-06-20", crop_end = "2018-05-02")
pam_data = pam_classify(pam_data, min_duration = 30)
head(pam_data$acceleration)
}
