% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_read}
\alias{pam_read}
\title{Read pam data}
\usage{
pam_read(
  pathname,
  extension = c("pressure", "glf", "acceleration", "temperature", "magnetic"),
  crop_start = "1900-01-01",
  crop_end = "2100-01-01"
)
}
\arguments{
\item{pathname}{path of the directory where the files are stored}

\item{extension}{list of file extensions to read (e.g., ".pressure", ".glf", ".gle",
".acceleration", ".temperature" and ".magnetic")}

\item{crop_start}{Remove all date before this date}

\item{crop_end}{Remove all date after this date}
}
\value{
a list of data.frames of all measurements type from the extension list (see example)
}
\description{
Imports multi-sensor logger data from a folder and optionally crop at specific date. Read all
available file from the extension list provided.
}
\examples{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
summary(pam_data)
head(pam_data$id)
head(pam_data$acceleration)
head(pam_data$light)
head(pam_data$pressure)
head(pam_data$temperature)
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureR/articles/pressure-map.html}{Vignette Pressure Map }
}
