% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_read}
\alias{pam_read}
\title{Read pam data}
\usage{
pam_read(
  pathname,
  extension = c("pressure", "glf", "acceleration", "temperature", "magnetic"),
  crop_start = "1900-01-01",
  crop_end = "2100-01-01"
)
}
\arguments{
\item{pathname}{path where files are stored}

\item{extension}{list of file extentions to read (e.g., ".pressure",
".glf", ".gle",".acceleration", ".temperature" and ".magnetic")}

\item{crop_start}{is the date that pam data should start}

\item{crop_end}{is the date that pam data should end}
}
\value{
a list of all measurements
}
\description{
Imports multi-sensor logger data from a folder and optionally crop at
specific date. Read all available file from the extension list AND which
exist in the folder.
}
