% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_read}
\alias{pam_read}
\title{Read pam data}
\usage{
pam_read(
  pathname,
  pressure_file = "*.pressure",
  light_file = "*.glf",
  acceleration_file = "*.acceleration",
  crop_start = "1900-01-01",
  crop_end = "2100-01-01",
  id = NA
)
}
\arguments{
\item{pathname}{path of the directory where the files are stored}

\item{pressure_file}{file with pressure data. Extension must be \code{.pressure}, \code{.deg} or \code{NA} if
absent.}

\item{light_file}{file with light data. Extension must be \code{.glf}, \code{.lux} or \code{NA} if absent.}

\item{acceleration_file}{file with acceleration data. Extension must be \code{.acceleration}  or \code{NA}
if absent.}

\item{crop_start}{Remove all date before this date}

\item{crop_end}{Remove all date after this date}

\item{id}{Unique identifier of the track. Default (\code{NA}) is to take the part of
\code{pressure_file} up to a character \verb{_} (e.g. \verb{18LX} for \verb{18LX_20180725.pressure}).}
}
\value{
a list of data.frames of pressure, light and acceleration.
}
\description{
Imports multi-sensor logger data from a folder (\code{pathname}) and optionally crop at specific date.
The file provided can be exact (e.g., \code{"18LX_20180725.pressure"}) or using regex expression
(e.g., \code{"*.pressure"}).
}
\details{
Create \href{https://github.com/Rafnuss/GeoPressureR/issues/new}{an issue on github} if you have data
in a format not supported yet.
}
\examples{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
summary(pam_data)
head(pam_data$id)
head(pam_data$pressure)
head(pam_data$light)
head(pam_data$acceleration)
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureR/articles/pressure-map.html}{Vignette Pressure Map }
}
