% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_read}
\alias{pam_read}
\title{Read pam data}
\usage{
pam_read(
  pathname,
  pressure_file = "*.pressure",
  light_file = "*.glf",
  acceleration_file = "*.acceleration",
  crop_start = "1900-01-01",
  crop_end = "2100-01-01",
  id = NA
)
}
\arguments{
\item{pathname}{path of the directory where the files are stored}

\item{pressure_file}{file with pressure data. Extension must be \code{.pressure} or \code{.deg} (required).}

\item{light_file}{file with light data. Extension must be \code{.glf}, \code{.lux} or \code{NA} if absent.}

\item{acceleration_file}{file with acceleration data. Extension must be \code{.acceleration}, \code{.deg}
or \code{NA} if absent.}

\item{crop_start}{Remove all date before this date (in UTC).}

\item{crop_end}{Remove all date after this date (in UTC).}

\item{id}{Unique identifier of the track. Default (\code{NA}) is to take the part of
\code{pressure_file} up to a character \verb{_} (e.g. \verb{18LX} for \verb{18LX_20180725.pressure}). If \code{basename},
take the basename of \code{pressure_file}.}
}
\value{
a list of data.frames of pressure, light and acceleration.
}
\description{
Imports multi-sensor logger data from a folder (\code{pathname}) and optionally crop at specific date.
The \verb{*_file} arguments are matched using a regex expression (e.g., \code{"*.pressure"} matches any
files with the extension \code{pressure}).
}
\details{
The current implementation can read the files from:
\itemize{
\item \href{https://www.vogelwarte.ch/en/projects/bird-migration/tracking-devices-miniaturized-geolocators}{Swiss Ornithological Institute (SOI) }
(default) with \code{pressure_file = "*.pressure"}, \code{light_file = "*.glf"} and
\code{acceleration_file = "*.acceleration"}
\item \href{http://www.migratetech.co.uk/}{Migrate Technology} with \code{pressure_file = "*.deg"},
\code{light_file = "*.lux"} and \code{acceleration_file = "*.deg"}
}

Create \href{https://github.com/Rafnuss/GeoPressureR/issues/new}{an issue on github} if you have data
in a format not supported yet.
}
\examples{
pam <- pam_read(
  pathname = system.file("extdata/0_PAM/18LX", package = "GeoPressureR")
)
summary(pam)
head(pam$id)
head(pam$pressure)
head(pam$light)
head(pam$acceleration)
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#read-geolocator-data}{GeoPressureManual | Pressure Map }
}
