% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_sta}
\alias{pam_sta}
\title{Compute stationary periods}
\usage{
pam_sta(pam)
}
\arguments{
\item{pam}{pam logger dataset list (see \code{\link[=pam_read]{pam_read()}})}
}
\value{
Same as input \code{pam} but with (1) a new data.frame of stationary periods \code{pam$sta} and (2)
a new column \code{sta_id} for pressure and light data.
}
\description{
This function computes the stationary periods from classified acceleration data
(\code{pam$acceleration$ismig}).
}
\examples{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
pam_data <- trainset_read(pam_data, pathname = system.file("extdata", package = "GeoPressureR"))
pam_data <- pam_sta(pam_data)
head(pam_data$pressure)
head(pam_data$light)
}
\seealso{
\code{\link{pam_read}}, \code{\link{pam_classify}}, \href{https://raphaelnussbaumer.com/GeoPressureR/articles/pressure-map.html}{Vignette Pressure Map }
}
