% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{pam_sta}
\alias{pam_sta}
\title{Compute stationary periods}
\usage{
pam_sta(pam)
}
\arguments{
\item{pam}{PAM logger dataset list. See \code{\link[=pam_read]{pam_read()}}.}
}
\value{
Same as input \code{pam} but with (1) a new data.frame of stationary periods \code{pam$sta} and (2)
a new column \code{sta_id} for pressure and light data.
}
\description{
This function computes the stationary periods from classified acceleration data
(\code{pam$acceleration$ismig}).
}
\examples{
pam <- pam_read(
  pathname = system.file("extdata/0_PAM/18LX", package = "GeoPressureR")
)
pam <- trainset_read(pam,
  pathname = system.file("extdata/1_pressure/labels", package = "GeoPressureR")
)
pam <- pam_sta(pam)
head(pam$pressure)
head(pam$light)
}
\seealso{
\code{\link{pam_read}}, \code{\link{pam_classify}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#identify-stationary-periods}{GeoPressureManual | Pressure Map }
}
