% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_create.R
\name{param_create}
\alias{param_create}
\title{Create a \code{param} list}
\usage{
param_create(id, default = FALSE, ...)
}
\arguments{
\item{id}{Unique identifier of a tag.}

\item{default}{logical to initiate param with default value of the package.}

\item{...}{arguments passed to other methods.}
}
\value{
A GeoPressureR \code{param} list
}
\description{
Create the list of parameter for GeoPressureR \code{tag} and \code{graph} objects.

\code{param} list are mostly used to archived the actual value of parameters used to create a \code{tag}
and/or a \code{graph}, thus allowing for examination of parameters post-creation. This function should
therefore not be used to set/define parameters ahead of computation. In reality, there are very
few external case of use for this function.
}
\examples{
param <- param_create("18LX", default = TRUE)
print(param)

}
\seealso{
Other param: 
\code{\link{print.param}()}
}
\concept{param}
