% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path2df}
\alias{path2df}
\title{Path to data.frame}
\usage{
path2df(pam, path)
}
\arguments{
\item{pam}{pam logger dataset list with \code{pam$sta} computed (see \code{pam_sta})}

\item{path}{data.frame containtings the path(s) of the bird with column \code{lat}, \code{lon} and \code{sta_id}
at least. Path can be generated with \code{geopressure_map2path}, \code{graph_simulation}, \code{geopressureviz}
.}
}
\description{
This function convert a GeoPressureR path to a data.frame which can be read in \href{https://movevis.org/index.html}{movevis} for instance.
}
\details{
The function basically duplicate location position at the start and end time of each stationary
period.
}
\examples{
pam <- pam_read(
  pathname = system.file("extdata/0_PAM/18LX", package = "GeoPressureR")
)
pam <- trainset_read(pam,
  pathname = system.file("extdata/1_pressure/labels", package = "GeoPressureR")
)
pam <- pam_sta(pam)
pressure_likelihood_1 <- readRDS(system.file("extdata/1_pressure/", "18LX_pressure_likelihood_1.rda",
  package = "GeoPressureR"
))
path <- geopressure_map2path(list(pressure_likelihood_1))
path2df(pam, path)
}
\seealso{
\href{https://movevis.org/reference/df2move.html}{\code{movevis::df2move}},
\code{\link{geopressure_map2path}}, \code{\link{graph_simulation}}, \code{\link{geopressureviz}}
}
