% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2edge.R
\name{path2edge}
\alias{path2edge}
\title{Edges flight info from path}
\usage{
path2edge(path, graph)
}
\arguments{
\item{path}{Data.frame containing at least \code{stap_id} and \code{ind}.}

\item{graph}{graph constructed with \code{\link[=graph_create]{graph_create()}}.}
}
\value{
Data.frame of the edge containing:
\itemize{
\item \code{stap_s} : stationary period of the origin (source).
\item \code{stap_t} : stationary period of the destination (target).
\item \code{gs} : groundspeed vector expressed as a complex number. You can compute the groundspeed value
(km/h) with \code{abs(gs)}, the W-E and S-N component of the flight with \code{Re(gs)} and \code{Im(gs)}, and
the angle/direction with \code{Arg(gs)}.
\item \code{dist} : Distance (in km) of the flight.
\item \code{ws}: if computed with \code{graph_add_wind()}, same value as \code{gs}. Airspeed is computed with
\code{as = gs - ws} in complex number to keep the vectorial additive properties.
}
}
\description{
Return the flight information of the edges composing a given path.
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/wind-graph.html#energy}{GeoPressureManual | Wind graph}
}
