% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2edge.R
\name{path2edge}
\alias{path2edge}
\title{Extract the edges of a \code{path} from a \code{graph}}
\usage{
path2edge(path, graph)
}
\arguments{
\item{path}{a GeoPressureR \code{path} data.frame}

\item{graph}{a GeoPressureR \code{graph} object}
}
\value{
Data.frame of the edge containing:
\itemize{
\item \code{stap_s} : stationary period of the origin (source).
\item \code{stap_t} : stationary period of the destination (target).
\item \code{gs} : groundspeed vector expressed as a complex number. You can compute the groundspeed value
(km/h) with \code{abs(gs)}, the W-E and S-N component of the flight with \code{Re(gs)} and \code{Im(gs)}, and
the angle/direction with \code{Arg(gs)}.
\item \code{distance} : Distance (in km) of the flight.
\item \code{ws}: if computed with \code{graph_add_wind()}, same value as \code{gs}. Airspeed is computed with
\code{as = gs - ws} in complex number to keep the vectorial additive properties.
}
}
\description{
Retrieve the edges in a \code{graph} corresponding to the flight transition defined by a \code{path}. These
edges can be useful to extract flight information specific to a path.
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/wind-graph.html#energy}{GeoPressureManual | Wind graph}

Other path: 
\code{\link{ind2path}()},
\code{\link{plot_path}()}
}
\concept{path}
