% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2twilight.R
\name{path2twilight}
\alias{path2twilight}
\title{Compute astronomical twilight from positions and dates}
\usage{
path2twilight(path, date = NULL)
}
\arguments{
\item{path}{a GeoPressureR \code{path} or \code{pressurepath} data.frame}

\item{date}{a vector of POSIXt datetime.}
}
\value{
a \code{twilight} data.frame with columns:
\itemize{
\item \code{twilight} (date-time of twilight)
\item \code{rise} (logical) indicating sunrise (\code{TRUE}) or sunset (\code{FALSE}).
\item \code{stap_id}.
}
}
\description{
This function compute the theorical twilight (i.e., datetime of sunrise and sunset) at given
locations and for specific date.

We use the \href{https://github.com/adokter/suntools}{suntools} package for this.
}
\examples{
owd <- setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)
setwd(owd)

# Create twilight data.frame
tag <- twilight_create(tag)

str(tag$twilight)

plot(tag, type = "twilight")

}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/light-map.html}{GeoPressureManual }, \href{https://github.com/adokter/suntools}{suntools }

Other geolight: 
\code{\link{geolight_map}()},
\code{\link{twilight_create}()},
\code{\link{twilight_label_read}()},
\code{\link{twilight_label_write}()}
}
\concept{geolight}
