% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.map.R
\name{plot.map}
\alias{plot.map}
\title{Plot a \code{map} object}
\usage{
\method{plot}{map}(
  x,
  path = NULL,
  thr_likelihood = 1,
  plot_leaflet = TRUE,
  provider = NULL,
  provider_options = leaflet::providerTileOptions(),
  palette = NULL,
  opacity = 0.8,
  legend = FALSE,
  fac_res_proj = 4,
  ...
)
}
\arguments{
\item{x}{a GeoPressureR \code{map} object}

\item{path}{a GeoPressureR \code{path} data.frame}

\item{thr_likelihood}{Threshold to display likelihood values.}

\item{plot_leaflet}{logical to use an interactive \code{leaflet} map instead of \code{terra::plot}}

\item{provider}{optional tile provider name (see \code{leaflet::providers}). When \code{NULL}, uses all
available GeoPressureR basemap layers; see the \emph{Basemap API keys} section.}

\item{provider_options}{tile options. See leaflet::addProviderTiles() and
leaflet::providerTileOptions().}

\item{palette}{color palette name or vector of colors.}

\item{opacity}{opacity of the raster layer in leaflet.}

\item{legend}{logical to display the legend.}

\item{fac_res_proj}{Factor of the resolution of the reprojection (see details above). A value of
\code{1} will roughly reproject on a map with similar size resulting in relatively high inaccuracy of
the pixel displayed. Increasing this factor will reduce the uncertainty but might also increase
the computational cost of the reprojection.}

\item{...}{additional parameters passed to \code{leaflet::addRasterImage()} or \code{terra::plot()}.}
}
\value{
a plot or leaflet object.
}
\description{
This function plots a GeoPressureR \code{map} object.

You can plot on top of the \code{map} a \code{path}, this uses the \code{plot_path()} function.

Our maps are defined in lat-lon (i.e., EPSG:4326), but the display of maps on web map are
nearly always in \href{https://en.wikipedia.org/wiki/Web_Mercator_projection}{web mercator} (i.e.,
EPSG:3857). We therefore need to reproject our map for display.
However, we don't really want to interpolate the map as each pixel might be important to
visualize. We therefore re-project with a near-neighbour interpolation (\code{method = "near"}
in \code{\link[terra:project]{terra::project()}}). Yet to avoid having pixel misplaced, we generally need to use a
projection with a finer resolution. The argument \code{fac_res_proj} controls the relative change of
resolution between the original map to the projected map.
}
\section{Basemap API keys}{

Interactive maps always offer Esri basemaps (gray canvas, imagery and topographic). CARTO,
Mapbox and MapTiler basemaps are added to the layer control only when their API key is stored
in your system keyring. Get a key from each provider:
\itemize{
\item CARTO: "Get your API key" on \href{https://carto.com/basemaps}{CARTO Basemaps}.
\item Mapbox: \href{https://account.mapbox.com/access-tokens/}{Access tokens} page of your account
(see \href{https://docs.mapbox.com/help/getting-started/access-tokens/}{Mapbox documentation}).
\item MapTiler: \href{https://cloud.maptiler.com/account/keys/}{API keys} page of your MapTiler Cloud
account (see \href{https://docs.maptiler.com/cloud/api/authentication-key/}{MapTiler documentation}).
}

Then store each key once:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{keyring::key_set_with_value("CARTO_API_KEY", password = "<your CARTO API key>")
keyring::key_set_with_value("MAPBOX_ACCESS_TOKEN", password = "<your Mapbox token>")
keyring::key_set_with_value("MAPTILER_API_KEY", password = "<your MapTiler API key>")
}\if{html}{\out{</div>}}

Mapbox access tokens are visible in the browser when a map is rendered. Use a public, read-only
token rather than a secret token.
}

\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |>
    tag_label(quiet = TRUE) |>
    tag_set_map(
      extent = c(-16, 23, 0, 50),
      scale = 4
    ) |>
    geopressure_map(quiet = TRUE)
})

plot(tag$map_pressure)

plot(tag$map_pressure, plot_leaflet = FALSE)

# `thr_likelihood` can be used to visualize its effect in `graph_create`
plot(tag$map_pressure,
  thr_likelihood = 0.9,
  palette = "viridis",
  opacity = 1,
  provider = "Esri.WorldGrayCanvas"
)

}
\seealso{
\code{\link[=plot_path]{plot_path()}} \code{\link[=plot.tag]{plot.tag()}}
}
\concept{map plot_tag}
