% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot.tag}
\alias{plot.tag}
\title{Plot a \code{tag} object}
\usage{
\method{plot}{tag}(x, type = NULL, ...)
}
\arguments{
\item{x}{a GeoPressureR \code{tag} object.}

\item{type}{type of the plot to display. One of \code{"pressure"}, \code{"acceleration"}, \code{"light"},
\code{"twilight"}, \code{"map"}, \code{"map_pressure"}, \code{"map_light"}, \code{"map_pressure_mse"},
\code{"map_pressure_mask"}, \code{"mask_water"}}

\item{...}{additional parameters for \code{plot_tag_pressure()}, \code{plot_tag_acceleration()},
\code{plot_tag_light()}, \code{plot_tag_twilight()} or \code{plot.map()}}
}
\description{
This function plot a GeoPressureR \code{tag} object as a timeseries or a map.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE)

# Plot pressure by default
plot(tag)
plot(tag, type = "acceleration")
plot(tag, type = "light")

tag <- twilight_create(tag) |>
  twilight_label_read()

plot(tag, type = "twilight")

tag <- tag_set_map(tag,
  extent = c(-16, 23, 0, 50),
  scale = 4,
  known = data.frame(
    stap_id = 1,
    known_lon = 17.05,
    known_lat = 48.9
  )
) |>
  geopressure_map(quiet = TRUE)

# default is now to take "map"
plot(tag)

tag <- geolight_map(tag)

plot(tag)
plot(tag, type = "map_light")

}
\seealso{
Other tag: 
\code{\link{print.tag}()},
\code{\link{tag_create}()},
\code{\link{tag_update}()}
}
\concept{tag}
