% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph_movement.R
\name{plot_graph_movement}
\alias{plot_graph_movement}
\title{Plot movement model of a \code{graph}}
\usage{
plot_graph_movement(graph, speed = seq(1, 120), plot_plotly = FALSE)
}
\arguments{
\item{graph}{a GeoPressureR \code{graph} object.}

\item{speed}{Vector of speed value (km/h) used on the x-axis.}

\item{plot_plotly}{logical to use \code{plotly}}
}
\description{
This function display a plot of pressure timeseries recorded by a tag
}
\seealso{
Other movement: 
\code{\link{bird_create}()},
\code{\link{graph_add_wind}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_transition}()},
\code{\link{speed2prob}()},
\code{\link{tag_download_wind}()}
}
\concept{movement}
