% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path.R
\name{plot_path}
\alias{plot_path}
\title{Plot \code{path}}
\usage{
plot_path(path)
}
\arguments{
\item{path}{A \code{path} list}
}
\description{
This function display the basic information on a \code{path} list.
}
\seealso{
map2path
}
