% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pressurepath.R
\name{plot_pressurepath}
\alias{plot_pressurepath}
\title{Plot pressure pressurepath}
\usage{
plot_pressurepath(
  pressurepath,
  type = "ts",
  pressure_err_warning = 2,
  pressure_err_danger = 5
)
}
\arguments{
\item{pressurepath}{rerer}

\item{type}{\code{"ts"} or \code{"hist"}}

\item{pressure_err_warning}{Threshold of pressure difference marking as ️warning (hPa)}

\item{pressure_err_danger}{Threshold of pressure difference marking as ️danger (hPa)}
}
\description{
description of check
}
