% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_acceleration}
\alias{plot_tag_acceleration}
\title{Plot acceleration data of a \code{tag}}
\usage{
plot_tag_acceleration(tag, plot_plotly = TRUE)
}
\arguments{
\item{plot_plotly}{Logical to display interactive plot with \code{plotly}}
}
\description{
This function display a plot of acceleration timeseries recorded by a tag
}
