% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_acceleration}
\alias{plot_tag_acceleration}
\title{Plot acceleration data of a \code{tag}}
\usage{
plot_tag_acceleration(
  tag,
  variable = "activity",
  plot_plotly = TRUE,
  label_auto = TRUE,
  min_duration = 30
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{variable}{type of acceleration variable to plot \code{"activity"} (or \code{"value"}) or \code{"pitch"}}

\item{plot_plotly}{logical to use \code{plotly}}

\item{label_auto}{logical to compute and plot the flight label using \code{tag_label_auto()}. Only if
labels are not already present on tag$acceleration$label}

\item{min_duration}{Minimal duration (in minutes) to consider a high activity as migratory
flight.}
}
\value{
a plot or ggplotly object.
}
\description{
This function display a plot of acceleration time series recorded by a tag
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)
})

plot_tag_acceleration(tag)

}
\seealso{
Other plot_tag: 
\code{\link{plot_tag_actogram}()},
\code{\link{plot_tag_light}()},
\code{\link{plot_tag_pressure}()},
\code{\link{plot_tag_temperature}()},
\code{\link{plot_tag_twilight}()}
}
\concept{plot_tag}
