% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_actogram}
\alias{plot_tag_actogram}
\title{Plot Actogram data of a \code{tag}}
\usage{
plot_tag_actogram(tag, twl_offset = NULL, plot_plotly = FALSE)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{twl_offset}{Shift of the middle of the night compared to 00:00 UTC (in hours). If not
provided, it uses the middle of all nights.}

\item{plot_plotly}{logical to use \code{plotly}}
}
\value{
a plot object.
}
\description{
This function display a plot of the acceleration time series recorded by a tag
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)

  plot_tag_actogram(tag, plot_plotly = TRUE)
})
}
\seealso{
Other plot_tag: 
\code{\link{plot_tag_acceleration}()},
\code{\link{plot_tag_light}()},
\code{\link{plot_tag_pressure}()},
\code{\link{plot_tag_temperature}()},
\code{\link{plot_tag_twilight}()}
}
\concept{plot_tag}
