% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_light}
\alias{plot_tag_light}
\title{Plot light data of a \code{tag}}
\usage{
plot_tag_light(tag, transform_light = TRUE, plot_plotly = TRUE)
}
\arguments{
\item{plot_plotly}{Logical to display interactive plot with \code{plotly}}
}
\description{
This function display a plot of light timeseries recorded by a tag
}
