% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_pressure}
\alias{plot_tag_pressure}
\title{Plot pressure data of a \code{tag}}
\usage{
plot_tag_pressure(
  tag,
  plot_plotly = TRUE,
  display_check = TRUE,
  pressure_diff_danger = 5,
  pressure_diff_warning = 3,
  stap_length_danger = 6,
  stap_length_warning = 12,
  ...
)
}
\arguments{
\item{plot_plotly}{Logical to display interactive plot with \code{plotly}}

\item{pressure_diff_danger}{Threshold of pressure hourly difference marking as ️danger (hPa)}

\item{pressure_diff_warning}{Threshold of pressure hourly difference marking as ️warning (hPa)}

\item{stap_length_danger}{Threshold number of pressure datapoints flagged as ️danger (hourly).}

\item{stap_length_warning}{Threshold number of pressure datapoints flagged as ️warning (hourly).}
}
\description{
This function display a plot of pressure timeseries recorded by a tag
}
