% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_temperature}
\alias{plot_tag_temperature}
\title{Plot temperature data of a \code{tag}}
\usage{
plot_tag_temperature(
  tag,
  variable = "external",
  plot_plotly = TRUE,
  label_auto = TRUE,
  min_duration = formals(tag_label_auto)$min_duration
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{variable}{temperature variable to plot \code{"external"} or \code{"internal"}}

\item{plot_plotly}{logical to use \code{plotly}}

\item{label_auto}{logical to compute and plot the flight label using \code{tag_label_auto()}. Only if
labels are not already present on tag$temperature$label}

\item{min_duration}{Minimal duration (in minutes) to consider a high activity as migratory
flight.}
}
\value{
a plot or ggplotly object.
}
\description{
This function display a plot of temperature time series recorded by a tag
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)
})

plot_tag_temperature(tag)

}
\seealso{
Other plot_tag: 
\code{\link{plot_tag_acceleration}()},
\code{\link{plot_tag_actogram}()},
\code{\link{plot_tag_light}()},
\code{\link{plot_tag_pressure}()},
\code{\link{plot_tag_twilight}()}
}
\concept{plot_tag}
