% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tag.R
\name{plot_tag_twilight}
\alias{plot_tag_twilight}
\title{Plot twilight data of a \code{tag}}
\usage{
plot_tag_twilight(tag, transform_light = TRUE, plot_plotly = FALSE)
}
\arguments{
\item{plot_plotly}{Logical to display interactive plot with \code{plotly}}
}
\description{
This function display a plot of twilight timeseries recorded by a tag
}
