% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressurepath_create.R
\name{pressurepath_create}
\alias{pressurepath_create}
\title{Create a \code{pressurepath}}
\usage{
pressurepath_create(
  tag,
  path = tag2path(tag),
  variable = c("altitude", "surface_pressure"),
  preprocess = FALSE,
  timeout = 60 * 5,
  workers = "auto",
  quiet = FALSE,
  debug = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{path}{a GeoPressureR \code{path} data.frame.}

\item{variable}{ERA5 variable/parameters available to download. See
\href{https://confluence.ecmwf.int/display/CKB/ERA5\%3A+data+documentation#heading-Parameterlistings}{Parameter listings in ERA5 doc}}

\item{preprocess}{logical to use \code{geopressure_map_preprocess}.}

\item{timeout}{duration before the code is interrupted both for the request on
GeoPressureAPI and on GEE (in seconds, see \code{httr2::req_timeout()}).}

\item{workers}{number of parallel requests on GEE. Integer between 1 and 99.}

\item{quiet}{logical to hide messages about the progress}

\item{debug}{logical to display additional information to debug a request}
}
\value{
A GeoPressureR \code{pressurepath} data.frame with columns:
\itemize{
\item \code{date} same as \code{pressure$date}
\item \code{stap_id} same as \code{pressure$stap_id}
\item \code{pressure_tag} same as \code{pressure$value}
\item \code{label} same as \code{pressure$label}
\item \code{j} same as \code{path$j}
\item \code{ind} same as \code{path$ind}
\item \code{lat} same as \code{path$lat}
\item \code{lon} same as \code{path$lon}
\item \code{include} same as \code{path$include}
\item \code{known} same as \code{path$known}
\item \code{altitude} altitude of the bird along the path (see detail)
\item \code{surface_pressure} pressure retrieved from ERA5.
\item \code{surface_pressure_norm} pressure retrieved from ERA5 normalized to the average of
\code{pressure_tag} over the stationary period.
\item \code{...} any other ERA5 variable requested by \code{variable}
}
}
\description{
A \code{pressurepath} is a data.frame merging \code{path} and \code{tag$pressure}. It can be thought as the
equivalent of the measurement of variables by sensors equipped on the bird moving along a
specific \code{path}.

Any ERA variables can be retrieve but the functions is most notably used to retrieve ERA5
surface pressure to then be able to compare it to the pressure measured by the tag, and
potentially adjust labeling tag data.

We use the \href{https://doi.org/10.24381/cds.e2161bac}{ERA5 LAND } or \href{https://doi.org/10.24381/cds.adbb2d47}{ERA5 surface level } if position over water. Available variable can be found
on the \href{https://confluence.ecmwf.int/display/CKB/ERA5\%3A+data+documentation#heading-Parameterlistings}{Parameter listings in ERA5 doc}

Positions during the flight are estimated by linear interpolation of the position of the stap
before and after. \code{pressurepath_create()} does not return measurement for stationary periods
which are not provided in \code{path} as well as the flight before and after.

\code{pressurepath_create()} also return the altitude of the bird along its trajectory. The altitude
\eqn{z_{tag}} (a.s.l.) is computed from the tag pressure \eqn{P_{tag}}, using the barometric
equation \deqn{ z_{{tag}}(x)=z_{ERA5}(x) + \frac{T_{ERA5}(x)}{L_b}  \left(
\frac{P_{tag}}{P_{ERA5}(x)} \right)^{\frac{RL_b}{g M}-1},}
where \eqn{z_{ERA}}, \eqn{T_{ERA}}, and \eqn{P_{ERA}} respectively correspond to the ground level
elevation, temperature at 2m, and ground level pressure of ERA5, \eqn{L_b}  is the standard
temperature lapse rate, \eqn{R} is the universal gas constant, \eqn{g} is the gravity constant
and  \eqn{M} is the molar mass of air. See more information at
\href{https://raphaelnussbaumer.com/GeoPressureAPI/#description-1}{the GeoPressureAPI documentation}.

To be able to compare the temporal variation of the retrieved pressure of ERA5 \eqn{P_{ERA}} to
the tag pressure \eqn{P_{tag}}, the function also returns the ERA pressure normalized with
the tag mean pressure measurement as \code{surface_pressure_norm}.
\deqn{ P_{ERA5,0}(\boldsymbol{x})[t] = \left( P_{ERA5}(\boldsymbol{x})[t]-P_{tag}[t]\right) -
\left( \frac{1}{n}\sum_{i=1}^{n} P_{ERA5}(\boldsymbol{x})[i]-P_{tag}[i] \right).}
}
\examples{
setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)

path <- data.frame(
  stap_id = tag$stap$stap_id,
  lat = c(48.5, 32.5, 30.5, 49.5, 41.6),
  lon = c(17.5, 13.5, 16.5, 21.5, 12.7)
)

pressurepath <- pressurepath_create(tag, path = path, quiet = TRUE)

str(pressurepath)

plot_pressurepath(pressurepath)

pressurepath <- pressurepath_create(
  tag,
  path[c(2, 3, 4), ],
  quiet = TRUE
)

plot_pressurepath(pressurepath)
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map}

Other pressurepath: 
\code{\link{geopressure_timeseries}()},
\code{\link{plot_pressurepath}()},
\code{\link{pressurepath_update}()}
}
\concept{pressurepath}
