% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressurepath_create.R
\name{pressurepath_create}
\alias{pressurepath_create}
\title{Create a \code{pressurepath}}
\usage{
pressurepath_create(
  tag,
  path = tag2path(tag),
  include_flight = FALSE,
  workers = "auto",
  preprocess = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{path}{a GeoPressureR \code{path} data.frame.}

\item{include_flight}{Extend request to also query the pressure and altitude during the previous
and/or next flight. Flights are defined by a \code{stap_id = 0}. Accept logical or vector of -1 (
previous flight), 0 (stationary) and/or 1 (next flight). (e.g. \code{include_flight = c(-1, 1)} will
only search for the flight before and after but not the stationary period). Note that next and
previous flights are defined by the +/1 of the \code{stap_id} value (and not the previous/next
\code{stap_id} value).}

\item{workers}{Number of parallel requests on GEE. Between 1 and 99. \code{"auto"} adjust the number
of workers to the number of \code{stap_elev} to query.}

\item{preprocess}{logical to use \code{geopressure_map_preprocess}.}

\item{quiet}{logical to hide messages about the progress}
}
\value{
A GeoPressureR \code{pressurepath} data.frame with columns:
\itemize{
\item \code{date} same as \code{pressure$date}
\item \code{pressure_tag}: same as \code{pressure$value}
\item \code{label} same as \code{pressure$label}
\item \code{stap_id} same as \code{pressure$stap_id} (i.e., including 0)
\item \code{pressure_era5} pressure retrieved from ERA5.
\item \code{altitude} Altitude computed
\item \code{lat} same as \code{path$lat}
\item \code{lon} same as \code{path$lon}
\item \code{pressure_era5_norm} pressure retrieved from ERA5 normalized to the average of \code{pressure_tag}
over the stationary period.
\item \code{stap_ref} stap_id grouping of the requests (same as \code{stap_id} except for flights where
\code{stap_id = 0} and \code{stap_ref} is to the stationary period of the position used from the path).
}
}
\description{
A \code{pressurepath} is a data.frame representing a timeseries of ERA5 reanalysis pressure along a
\code{path} (position over time). It can be thought as the equivalent of the hypothetical measurement
of a pressure sensor equiped on a bird moving along a specific \code{path}.

\code{pressurepath} are useful for labeling tag data by allowing a direct comparison of the actual
pressure measured by the sensor to the one equivalent to the best estimate of the path.

You can include previous and/or next flight period in each query, for instance to estimate
flight altitude.

The function essentially splits the pressure of the geolocator into stationary periods and
downloads the ERA5 timeseries of pressure for the corresponding location of the path by calling
\code{geopressure_timeseries()} at each stationary period. The computation is optimized by
paralleling the call of each stationary period.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)

path <- data.frame(
  stap_id = tag$stap$stap_id,
  lat = c(48.5, 32.5, 30.5, 49.5, 41.6),
  lon = c(17.5, 13.5, 16.5, 21.5, 12.7)
)

pressurepath <- pressurepath_create(tag, path = path, quiet = TRUE)

str(pressurepath)

plot_pressurepath(pressurepath)

pressurepath <- pressurepath_create(
  tag,
  path[c(2, 3, 4), ],
  include_flight = TRUE,
  quiet = T
)

plot_pressurepath(pressurepath)
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html}{GeoPressureManual | Pressure Map}

Other pressurepath: 
\code{\link{geopressure_timeseries}()},
\code{\link{plot_pressurepath}()},
\code{\link{pressurepath2altitude}()},
\code{\link{pressurepath_update}()}
}
\concept{pressurepath}
