% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressurepath_update.R
\name{pressurepath_update}
\alias{pressurepath_update}
\title{Update a \code{pressurepath}}
\usage{
pressurepath_update(pressurepath, tag, path = tag2path(tag), quiet = FALSE)
}
\arguments{
\item{pressurepath}{a GeoPressureR \code{pressurepath} data.frame}

\item{tag}{a GeoPressureR \code{tag} object.}

\item{path}{a GeoPressureR \code{path} data.frame.}

\item{quiet}{logical to hide messages about the progress}
}
\value{
a list containing the new \code{pressurepath}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because (1) few people use it, (2) hard to maintain and (3) not
so slow to use the alternative of creating of new \code{pressurepath} altogether.

When updating the labelling file of a tag, often, only a few stationary periods are changing. To
avoid recomputing the entire workflow, this function figure out which stationary period have
been changing on only update those in \code{tag$map_pressure} and \code{pressurepath}.
}
\keyword{internal}
