% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressurepath_update.R
\name{pressurepath_update}
\alias{pressurepath_update}
\title{Update a \code{pressurepath}}
\usage{
pressurepath_update(pressurepath, tag, path = tag2path(tag))
}
\arguments{
\item{pressurepath}{a GeoPressureR \code{pressurepath} data.frame}

\item{tag}{a GeoPressureR \code{tag} object.}

\item{path}{a GeoPressureR \code{path} data.frame.}
}
\value{
a list containing the new \code{pressurepath}.
}
\description{
When updating the labelling file of a tag, often, only a few stationary periods are changing. To
avoid recomputing the entire workflow, this function figure out which stationary period have
been changing on only update those in \code{tag$map_pressure} and \code{pressurepath}.
}
\seealso{
Other pressurepath: 
\code{\link{geopressure_timeseries}()},
\code{\link{plot_pressurepath}()},
\code{\link{pressurepath2altitude}()},
\code{\link{pressurepath_create}()}
}
\concept{pressurepath}
