% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bird.R
\name{print.bird}
\alias{print.bird}
\title{Print a \code{bird} object}
\usage{
\method{print}{bird}(x, ...)
}
\arguments{
\item{x}{a GeoPressureR \code{bird} object.}

\item{...}{arguments passed to other methods}
}
\value{
\code{bird} is returned invisibly and unchanged
}
\description{
This function displays the information of a \code{bird} object.
}
\examples{
# Using AVONET dataset
bird_create("Acrocephalus arundinaceus")

}
\seealso{
Other bird: 
\code{\link{bird_create}()}
}
\concept{bird}
