% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird.R
\name{print.bird}
\alias{print.bird}
\title{Print \code{bird}}
\usage{
\method{print}{bird}(x, ...)
}
\arguments{
\item{x}{A \code{bird} object}

\item{...}{arguments passed from other methods}
}
\value{
\code{bird} is returned invisibly and unchanged
}
\description{
This function display the basic information on a \code{bird}
}
\seealso{
Other bird: 
\code{\link{bird_create}()},
\code{\link{is.bird}()}
}
\concept{bird}
