% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{print.graph}
\alias{print.graph}
\title{Print \code{graph}}
\usage{
\method{print}{graph}(x, ...)
}
\arguments{
\item{x}{A \code{graph} list}
}
\value{
\code{graph} is returned invisibly and unchanged
}
\description{
This function display the basic information on a \code{graph} list.
}
\seealso{
graph_create
}
