% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.graph.R
\name{print.graph}
\alias{print.graph}
\title{Print a \code{graph} object}
\usage{
\method{print}{graph}(x, ...)
}
\arguments{
\item{x}{a GeoPressureR \code{graph} object.}

\item{...}{arguments passed to other methods}
}
\value{
\code{graph} is returned invisibly and unchanged
}
\description{
This function displays the information of a \code{graph} object.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  twilight_create() |>
  twilight_label_read() |>
  tag_set_map(
    extent = c(-16, 23, 0, 50),
    known = data.frame(stap_id = 1, known_lon = 17.05, known_lat = 48.9)
  ) |>
  geopressure_map(quiet = TRUE) |>
  geolight_map(quiet = TRUE)

graph <- graph_create(tag, quiet = TRUE)

print(graph)

}
\seealso{
Other graph: 
\code{\link{graph_add_wind}()},
\code{\link{graph_create}()},
\code{\link{graph_marginal}()},
\code{\link{graph_most_likely}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_simulation}()}
}
\concept{graph}
