% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.map.R
\name{print.map}
\alias{print.map}
\title{Print a \code{map} object}
\usage{
\method{print}{map}(x, ...)
}
\arguments{
\item{x}{a GeoPressureR \code{map} object}

\item{...}{arguments passed to other methods}
}
\value{
\code{map} is returned invisibly and unchanged
}
\description{
This function displays the information of a \code{map} object.
}
\examples{
setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  tag_set_map(
    extent = c(-16, 23, 0, 50),
    scale = 4
  ) |>
  geopressure_map(quiet = TRUE)

print(tag$map_pressure)

}
\seealso{
Other map: 
\code{\link{map_create}()},
\code{\link{rast.map}()}
}
\concept{map}
