% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tag.R
\name{print.tag}
\alias{print.tag}
\title{Print a \code{tag} object}
\usage{
\method{print}{tag}(x, ...)
}
\arguments{
\item{x}{a GeoPressureR \code{tag} object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{tag} is returned invisibly and unchanged
}
\description{
This function displays the information of a \code{tag} object.
}
\examples{
setwd(system.file("extdata", package = "GeoPressureR"))

tag <- tag_create("18LX", quiet = TRUE)

print(tag)

}
\seealso{
Other tag: 
\code{\link{tag_create}()},
\code{\link{tag_set_map}()},
\code{\link{tag_update}()}
}
\concept{tag}
