% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tag.R
\name{print.tag}
\alias{print.tag}
\title{Print \code{tag}}
\usage{
\method{print}{tag}(x, ...)
}
\arguments{
\item{x}{A \code{tag} object}

\item{...}{arguments passed from other methods}
}
\value{
\code{tag} is returned invisibly and unchanged
}
\description{
This function display the basic information on a \code{tag}
}
\seealso{
Other tag: 
\code{\link{is.tag}()},
\code{\link{plot.tag}()},
\code{\link{tag_create}()}
}
\concept{tag}
