% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed2prob.R
\name{speed2prob}
\alias{speed2prob}
\title{Probability of flight}
\usage{
speed2prob(
  speed,
  method = "gamma",
  shape = 7,
  scale = 7,
  location = 40,
  bird = NA,
  power2prob = function(power) (1/power)^3,
  low_speed_fix = 15
)
}
\arguments{
\item{speed}{airspeed or ground speed in km/h}

\item{method}{method used to convert the speed to probability ("gamma", "logis" or "power")}

\item{shape}{parameter of the gamma distribution}

\item{scale}{parameter of the gamma and logistic distribution}

\item{location}{parameter for the logistic distribution}

\item{bird}{list of basic morphological traits necessary: mass, wing span, wing aspect ratio, and
body frontal area. It is good practice to create a bird with \code{\link[=bird_create]{bird_create()}}.}

\item{power2prob}{function taking power as a single argument and returning a probability}

\item{low_speed_fix}{speed below which the probability remains the same. This parameter is used
to allow short flights covering small distances.}
}
\value{
Probability values corresponding to the speed provided.
}
\description{
Converts a speed (airspeed or ground speed) to a probability using different parametric functions.
}
\examples{
speed <- seq(1, 120)
low_speed_fix <- 20 # minimum speed allowed
prob <- speed2prob(speed,
  method = "gamma", shape = 7, scale = 7,
  low_speed_fix = low_speed_fix
)
plot(speed, prob,
  type = "l", xlab = "Groundspeed [km/h]", ylab =
    "Probability"
)
abline(v = low_speed_fix)
bird <- bird_create("Acrocephalus arundinaceus")
prob <- speed2prob(speed, method = "power", bird = bird)
plot(speed, prob, type = "l", xlab = "Airspeed [km/h]", ylab = "Probability")
}
\seealso{
\code{\link[=bird_create]{bird_create()}}
}
