% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2likelihood.R
\name{tag2likelihood}
\alias{tag2likelihood}
\title{Return a likelihood map from a \code{tag}}
\usage{
tag2likelihood(tag, likelihood = NA)
}
\arguments{
\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character). Default
\code{NA} is to take the product of \code{map_pressure} and \code{map_light}, or only \code{map_pressure} if
\code{map_light} is not available.}
}
\value{
likelihood map
}
\description{
Return a likelihood map from a \code{tag}
}
