% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag2path.R
\name{tag2path}
\alias{tag2path}
\title{Extract a \code{path} from a \code{tag}}
\usage{
tag2path(tag, likelihood = NULL, interp = FALSE, use_known = TRUE)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{likelihood}{Field of the \code{tag} list containing the likelihood map (character). Possible
value are \code{map_pressure}, \code{map_light}, \code{map_pressure_mse}, \code{map_pressure_mse},
\code{map_pressure_mse}, \code{mask_water}. Default \code{NA} is to take the product of \code{map_pressure} and
\code{map_light}, or if not available, taking the first of the possible values.}

\item{interp}{The position of the stationary period shorter than \code{interp} will be
replace by a linear average from other position accounting for flight duration (in days) .}

\item{use_known}{If true, enforce the known position defined in \code{tag} in the path created. Known
position are not interpolated (even if shorter than \code{interp}) and used in the interpolation. In
most (all?) case, the likelihood map was computed using known, and therefore will result in the
same position (approx. to the map resolution).}
}
\value{
A path data.frame
\itemize{
\item \code{stap_id} stationary period
\item \code{ind} indices of the coordinate in the 2D grid. Useful to retrieve map or graph information.
\item \code{lat} Latitude,
\item \code{lon} longitude
}
}
\description{
Find the position of the highest value in a map, typically most probable value in a likelihood
map.
}
\details{
Note that this path is the most likely, considering only the observation model and ignoring the
movement model. Prefer to use \code{graph_most_likely()} for the most realistic path.

\code{interp} can be used to interpolate unrealistic position from short stationary period based on
the position of the longer ones. The interpolation assumes that the first and last stationary
period can be safely estimated from the probability map.
}
\examples{
owd <- setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  tag_set_map(
    extent = c(-16, 23, 0, 50),
    scale = 2
  ) |>
  geopressure_map(quiet = TRUE)
setwd(owd)

# Extract a path from pressure map
path <- tag2path(tag)
plot_path(path, plot_leaflet = FALSE)

# Short stationary periods can be unreliably estimated, so
# interpolating them might be better
path <- tag2path(tag, interp = 1)
plot_path(path, plot_leaflet = FALSE)

}
