% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_assert.R
\name{tag_assert}
\alias{tag_assert}
\title{Assert the status of a \code{tag}}
\usage{
tag_assert(tag, condition = "tag", type = "abort")
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{condition}{condition to assert \code{tag} for. One of \code{"tag"} (default), \code{"pressure"},
\code{"light"}, \code{"acceleration"}, \code{"label"}, \code{"stap"}, \code{"setmap"}, \code{"map_pressure"}, \code{"map_light"}
\code{"map_pressure_mismatch"} and \code{"twilight"}}

\item{type}{Message type to display. One of \code{"abort"} (default), \code{"warn"} or \code{"inform"}}
}
\value{
logical indicating the \code{tag} object has the relevant element
}
\description{
This function check the condition of a \code{tag} object.
}
\examples{
setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)

tag_assert(tag)

tag_assert(tag, "stap")

tag_assert(tag, "setmap", type = "warn")

tag_assert(tag, "map_pressure", type = "inform")

}
