% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_assert.R
\name{tag_assert}
\alias{tag_assert}
\title{Assert status of a \code{tag}}
\usage{
tag_assert(tag, cond = "tag")
}
\arguments{
\item{tag}{a \code{tag} object}
}
\value{
logical indicating the \code{tag} object has the relevant element
}
\description{
These functions return logical about the contents of a \code{tag} object.
}
