% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_create.R
\name{tag_create}
\alias{tag_create}
\title{Create a \code{tag} object}
\usage{
tag_create(
  id,
  manufacturer = NULL,
  crop_start = NULL,
  crop_end = NULL,
  directory = glue::glue("./data/raw-tag/{id}"),
  pressure_file = NULL,
  light_file = NULL,
  acceleration_file = NULL,
  temperature_external_file = NULL,
  temperature_internal_file = NULL,
  magnetic_file = NULL,
  assert_pressure = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{id}{unique identifier of a tag.}

\item{manufacturer}{One of \code{NULL}, \code{"soi"}, \code{"migratetech"}, \code{"lund"} or \code{"manual"}}

\item{crop_start}{remove all data before this date (POSIXct or character in UTC).}

\item{crop_end}{remove all data after this date (POSIXct or character in UTC).}

\item{directory}{path of the directory where the tag files can be read.}

\item{pressure_file}{name of the file with pressure data. Full pathname  or finishing
with extensions (e.g., \code{"*.pressure"}, \code{"*.deg"} or \code{"*_press.xlsx"}).}

\item{light_file}{name of the file with light data. Full pathname  or finishing
with extensions (e.g., \code{"*.glf"}, \code{"*.lux"} or \code{"*_acc.xlsx"}).}

\item{acceleration_file}{name of the file with acceleration data. Full pathname  or finishing
with extensions (e.g., \code{"*.acceleration"}, \code{"*.deg"} or \code{"*_acc.xlsx"}).}

\item{temperature_external_file}{name of the file with temperature data. Full pathname  or
finishing with extensions (e.g., \code{"*.temperature"}, \code{"*.airtemperature"} or \code{"*.deg"}). External
or air temperature is generally for temperature sensor on directed outward from the bird.}

\item{temperature_internal_file}{name of the file with temperature data . Full pathname  or
finishing with extensions (e.g., \code{"*.bodytemperature"}). Internal or body temperature is
generally for temperature sensor on directed inward (between bird and tag).}

\item{magnetic_file}{name of the file with magnetic/accelerometer data. Full pathname  or
finishing with extensions (e.g., \code{"*.magnetic"}).}

\item{assert_pressure}{logical to check that the return tag has pressure data.}

\item{quiet}{logical to hide messages about the progress.}
}
\value{
a GeoPressureR \code{tag} object containing
\itemize{
\item \code{param} parameter object (see \code{\link[=param_create]{param_create()}})
\item \code{pressure} data.frame with columns: \code{date} and \code{value}
\item \code{light} (optional) same structure as pressure
\item \code{temperature_external} (optional) same structure as pressure
\item \code{temperature_internal} (optional) same structure as pressure
\item \code{acceleration} (optional) data.frame with columns: \code{date}, \code{value}, \code{act} and \code{pit}.
\itemize{
\item \code{value} is the activity computed as the sum of the difference in acceleration on the z-axis
(i.e. jiggle). In the SOI sensor, it is summarised from 32 measurements at 10Hz
\item \code{pitch} is the relative position of the bird’s body relative to the z axis. In the SOI
sensor, it is an average over 32 measurements at 10Hz.
}
\item \code{magnetic} (optional) data.frame with columns: \code{date}, \code{magnetic_x}, \code{magnetic_y}, \code{magnetic_z}
, \code{acceleration_x}, \code{acceleration_y} and \code{acceleration_z}
}
}
\description{
Create a GeoPressureR \code{tag} object from the data collected by a tracking device. The function
can read data formatted according to three manufacturers SOI, Migratetech or Lund CAnMove, as
well as according to the GeoLocator Data package standard and also accept manual data.frame as
input. Pressure data is required for the GeoPressureR workflow but can be allowed to be missing
with \code{assert_pressure = FALSE}.
}
\details{
The current implementation can read files from the following three sources:
\itemize{
\item \href{https://raphaelnussbaumer.com/GeoLocator-DP/}{GeoLocator Data Package (\code{gldp})}
\itemize{
\item \code{pressure_file = "pressure.csv"}(optional)
\item \code{light_file = "light.csv"} (optional)
\item \code{acceleration_file = "acceleration.csv"} (optional)
\item \code{temperature_external_file = "temperature.csv"} (optional)
\item \code{temperature_external_file = "airtemperature.csv"} (optional)
\item \code{magnetic_file = "magnetic.csv"} (optional)
}
\item \href{https://bit.ly/3QI6tkk}{Swiss Ornithological Institute (\code{soi})}
\itemize{
\item \code{pressure_file = "*.pressure"}
\item \code{light_file = "*.glf"} (optional)
\item \code{acceleration_file = "*.acceleration"} (optional)
\item \code{temperature = "*.temperature"} (optional)
\item \code{airtemperature = "*.airtemperature"} (optional)
\item \code{magnetic = "*.magnetic"} (optional)
}
\item \href{http://www.migratetech.co.uk/}{Migrate Technology (\code{migratetech})}:
\itemize{
\item \code{pressure_file = "*.deg"}
\item \code{light_file = "*.lux"} (optional)
\item \code{acceleration_file = "*.deg"} (optional)
}
\item \href{https://bit.ly/3P6quyi}{Lund CAnMove (\code{lund})}
\itemize{
\item \code{pressure_file = "*_press.xlsx"}
\item \code{light_file = "*_acc.xlsx"} (optional)
\item \code{acceleration_file = "*_acc.xlsx"} (optional)
}
}

You can also enter the data manually (\code{manufacturer = "manual"}) by providing the data.frame to
\code{pressure_file}:
\itemize{
\item \code{pressure_file}: data.frame with column date and value.
\item \code{light_file}: (optional) data.frame with column date and value.
\item \code{acceleration_file}: (optional) data.frame with column date and value.
}

You can still create a \code{tag} without pressure data using \code{assert_pressure = TRUE}. This \code{tag}
won't be able to run the traditional GeoPressureR workflow, but you can still do some analysis.

By default \code{manufacturer = NULL}, the manufacturer is determined automatically from the content
of the \code{directory}. You can also specify manually the file with a full pathname or the file
extension using a regex expression (e.g., \code{"*.pressure"} matches any file ending with
\code{pressure}).

Please create \href{https://github.com/Rafnuss/GeoPressureR/issues/new}{an issue on Github} if you
have data in a format that is not yet supported.

This function can be used to crop the data at specific date, for instance to remove pre-equipment
or post-retrieval data.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  # Read all sensor file
  tag <- tag_create("18LX")

  print(tag)

  # Read only pressure and crop date
  tag <- tag_create("18LX",
    light_file = NULL,
    acceleration_file = NULL,
    crop_start = "2017-08-01",
    crop_end = "2017-08-05"
  )

  print(tag)

  # You can also specify the exact file in case multiple files with the
  # same extension exist in your directory (migratetech data)
  tag <- tag_create("CB621",
    pressure_file = "CB621_BAR.deg",
    light_file = "CB621.lux",
    acceleration_file = NULL
  )

  print(tag)

  # You can specify the data manually with
  pressure <- data.frame(
    date = as.POSIXct(c(
      "2017-06-20 00:00:00 UTC", "2017-06-20 01:00:00 UTC",
      "2017-06-20 02:00:00 UTC", "2017-06-20 03:00:00 UTC"
    ), tz = "UTC"),
    value = c(1000, 1000, 1000, 1000)
  )
  tag_create(id = "xxx", pressure_file = pressure)
})

}
\seealso{
\href{https://bit.ly/4462jpr}{GeoPressureManual}

Other tag: 
\code{\link{print.tag}()},
\code{\link{tag_set_map}()}
}
\concept{tag}
