% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_create.R
\name{tag_create}
\alias{tag_create}
\title{Create a \code{tag} object}
\usage{
tag_create(
  id,
  directory = glue::glue("./data/raw-tag/{id}/"),
  pressure_file = "auto",
  light_file = "auto",
  acceleration_file = "auto",
  crop_start = NULL,
  crop_end = NULL,
  quiet = FALSE
)
}
\arguments{
\item{id}{unique identifier of a tag.}

\item{directory}{path of the directory where the tag files can be read.}

\item{pressure_file}{name of the file with pressure data. Full pathname, \code{"auto"}, or finishing
with extensions \code{.pressure} or \code{.deg} (required).}

\item{light_file}{name of the file with light data. Full pathname, \code{"auto"}, \code{NULL} to ignore or
extensions \verb{*.glf}, \verb{*.lux}.}

\item{acceleration_file}{name of the file with acceleration data. Full pathname, \code{"auto"}, \code{NULL}
to ignore or extension \code{.acceleration}, \code{.deg} (or \code{NA} to ignore).}

\item{crop_start}{remove all data before this date (POSIXct or character in UTC).}

\item{crop_end}{remove all data after this date (POSIXct or character in UTC).}

\item{quiet}{logical to hide messages about the progress.}
}
\value{
a GeoPressureR \code{tag} object containing
\itemize{
\item \code{param} parameter object (see \link{param_create})
\item \code{pressure} data.frame with columns for \code{date} and \code{value}
\item \code{light} (optional) same structure as pressure
\item \code{acceleration} (optional) same structure as pressure
}
}
\description{
Create a GeoPressureR \code{tag} object from the data collected by the tracking device. This can
include one or multiple sensors (pressure, light, acceleration). This function also optionally
crops the data at a specific date.

By default \code{"auto"}, the file are determined automatically from the files present in \code{directory}.
You can also specfy the full pathname or a file extension using a \link{regex} expression (e.g.,
\code{"*.pressure"} matches any file ending with \code{pressure}).

The current implementation can read files from:
\itemize{
\item \href{https://bit.ly/3QI6tkk}{Swiss Ornithological Institute (SOI)}
(default)
\itemize{
\item \code{pressure_file = "*.pressure"}
\item \code{light_file = "*.glf"}
\item \code{acceleration_file = "*.acceleration"}
}
\item \href{http://www.migratetech.co.uk/}{Migrate Technology}:
\itemize{
\item \code{pressure_file = "*.deg"}
\item \code{light_file = "*.lux"}
\item \code{acceleration_file = "*.deg"}
}
}

Use \code{NULL} to hide the message warning about the absence of a sensor file.

Please create \href{https://github.com/Rafnuss/GeoPressureR/issues/new}{an issue on Github} if you
have data in a format that is not yet supported.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))

# Read all sensor file
tag <- tag_create("18LX")

print(tag)

# Read only pressure and crop date
tag <- tag_create("18LX",
  light_file = NULL,
  acceleration_file = NULL,
  crop_start = "2017-08-01",
  crop_end = "2017-08-05"
)

print(tag)

# You can also specify the exact file in case multiple files with the
# same extension exist in your directory
tag <- tag_create("CB621",
  pressure_file = "CB621_BAR.deg",
  light_file = "CB621.lux",
  acceleration_file = NULL
)

print(tag)

}
\seealso{
\href{https://bit.ly/4462jpr}{GeoPressureManual}

Other tag: 
\code{\link{print.tag}()},
\code{\link{tag_set_map}()},
\code{\link{tag_update}()}
}
\concept{tag}
