% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_create.R
\name{tag_create}
\alias{tag_create}
\title{Create a \code{tag} object}
\usage{
tag_create(
  id,
  directory = glue::glue("./data/raw-tag/{id}/"),
  pressure_file = "*.pressure",
  light_file = "*.glf",
  acceleration_file = "*.acceleration",
  crop_start = NULL,
  crop_end = NULL,
  quiet = FALSE
)
}
\arguments{
\item{id}{Unique identifier of a tag.}

\item{directory}{Path of the directory where the tag files can be read.}

\item{pressure_file}{Name of the file with pressure data. Extension must be \code{.pressure} or
\code{.deg} (required).}

\item{light_file}{Name of the file with light data. Extension must be \code{.glf}, \code{.lux} (or \code{NA}
to ignore).}

\item{acceleration_file}{Name of the file with acceleration data. Extension must be
\code{.acceleration}, \code{.deg} (or \code{NA} to ignore).}

\item{crop_start}{Remove all data before this date (POSIXct or character in UTC).}

\item{crop_end}{Remove all data after this date (POSIXct or character in UTC).}

\item{quiet}{logical to hide messages about the progress.}
}
\value{
A GeoPressureR \code{tag} object containing
\itemize{
\item \code{param} parameter object (see \link{param_create})
\item \code{pressure} data.frame with column \code{date} and \code{value}
\item \code{light} (optional) same structure as pressure
\item \code{acceleration} (optional) same structure as pressure
}
}
\description{
Create a GeoPressureR \code{tag} object from sensor(s) data and optionally crops at specific
date. The \verb{*_file} arguments are matched using a \link{regex} expression (e.g., \code{"*.pressure"}
matches any file with the extension \code{pressure}).

The current implementation can read the files from:
\itemize{
\item \href{https://www.vogelwarte.ch/en/projects/bird-migration/tracking-devices-miniaturized-geolocators}{Swiss Ornithological Institute (SOI) }
(default)
\itemize{
\item \code{pressure_file = "*.pressure"}
\item \code{light_file = "*.glf"}
\item \code{acceleration_file = "*.acceleration"}
}
\item \href{http://www.migratetech.co.uk/}{Migrate Technology}:
\itemize{
\item \code{pressure_file = "*.deg"}
\item \code{light_file = "*.lux"}
\item \code{acceleration_file = "*.deg"}
}
}

Use \code{NULL} to hide the message warning about the absence of a sensor file.

Please create \href{https://github.com/Rafnuss/GeoPressureR/issues/new}{an issue on Github} if you
have data in a format not supported yet.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))

tag_create("18LX")

tag_create("18LX", crop_start = "2017-08-01", crop_end = "2017-08-05")

# For Migrate Technology file, use
tag_create("CB621",
  pressure_file = "*.deg",
  light_file = "*.lux",
  acceleration_file = NULL
)

# You can also specify exactly the file in case multiple file with the same
# extension exist in your directory
tag_create("CB621",
  pressure_file = "CB621_BAR.deg",
  light_file = NA,
  acceleration_file = NA
)

}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#read-geolocator-data}{GeoPressureManual | Pressure Map }

Other tag: 
\code{\link{plot.tag}()},
\code{\link{print.tag}()},
\code{\link{tag_update}()}
}
\concept{tag}
