% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_download_wind.R
\name{tag_download_wind}
\alias{tag_download_wind}
\title{Download flight data}
\usage{
tag_download_wind(
  tag,
  extent = tag$param$tag_set_map$extent,
  include_stap_id = NULL,
  variable = c("u_component_of_wind", "v_component_of_wind"),
  file = function(stap_id, tag_id) {
    
    glue::glue("./data/wind/{tag_id}/{tag_id}_{stap_id}.nc")
 },
  overwrite = FALSE,
  workers = 19,
  cds_token = lifecycle::deprecated(),
  ...
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{extent}{geographical extent of the map on which the likelihood will be computed.
Vector of length 4 \code{c(xmin, xmax, ymin, ymax)} or \code{c(W, E, S, N)}.}

\item{include_stap_id}{stationary period identifiers of the start of the flight to download.
Default is to download all flights.}

\item{variable}{list of variables to download from \href{https://confluence.ecmwf.int/display/CKB/ERA5:+data+documentation#ERA5:datadocumentation-Table9}{the ERA5 pressure level }:
\code{"u_component_of_wind"}, \code{"v_component_of_wind"},  \code{"temperature"},
\code{"fraction_of_cloud_cover"}, \code{"relative_humidity"}, \code{"vertical_velocity"},
\code{"specific_cloud_ice_water_content"}, \code{"specific_cloud_liquid_water_content"},
\code{"specific_humidity"}, \code{"specific_rain_water_content"}, \code{"specific_snow_water_content"},
\code{"divergence"}, \code{"geopotential"}, \code{"ozone_mass_mixing_ratio"}, \code{"potential_vorticity"},
\verb{'vorticity"}.}

\item{file}{absolute or relative path of the ERA5 wind data file to be downloaded. Function
taking as arguments (1) the stationary period identifier and (2) the tag_id.}

\item{overwrite}{logical. If \code{TRUE}, file is overwritten.}

\item{workers}{maximum number of simultaneous request that will be submitted
to the service. Most ECMWF services are limited to 20 concurrent requests
(default = 2).}

\item{cds_token}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Enter the API token with
\code{\link[ecmwfr:wf_set_key]{ecmwfr::wf_set_key()}}}

\item{...}{
  Arguments passed on to \code{\link[ecmwfr:wf_request_batch]{ecmwfr::wf_request_batch}}
  \describe{
    \item{\code{request}}{nested list with query parameters following the layout
as specified on the ECMWF APIs page}
    \item{\code{user}}{user (default = "ecmwf") provided by the ECMWF data service,
used to retrieve the token set by \code{\link[ecmwfr]{wf_set_key}}}
    \item{\code{transfer}}{logical, download data TRUE or FALSE (default = TRUE)}
    \item{\code{path}}{path were to store the downloaded data}
    \item{\code{time_out}}{how long to wait on a download to start (default =
\code{3600} seconds).}
    \item{\code{retry}}{polling frequency of submitted request for downloading (default =
\code{30} seconds).}
    \item{\code{job_name}}{optional name to use as an RStudio job and as output variable
name. It has to be a syntactically valid name.}
    \item{\code{verbose}}{show feedback on processing}
    \item{\code{request_list}}{a list of requests that will be processed in parallel.}
    \item{\code{total_timeout}}{overall timeout limit for all the requests in seconds.}
  }}
}
\value{
The path of the downloaded (requested file) or the an R6 object with download/transfer
information
}
\description{
This function downloads data associated to each flight from the \href{https://doi.org/10.24381/cds.bd0915c6}{ERA5 hourly pressure levels}
with the \href{https://cds.climate.copernicus.eu/}{Climate Data Store (CDS)} and through the \href{https://bluegreen-labs.github.io/ecmwfr/index.html}{\code{ecmwfr} R package}.

\href{https://confluence.ecmwf.int/display/CKB/ERA5:+data+documentation#ERA5:datadocumentation-Table9}{Any variable available from the ERA5 pressure level}
can be downloaded.

The flights are determined from the stationary periods classified \code{tag$stap}. It requests a
single file for each flight using the exact time (hourly basis) and pressure (altitude). To make
the download more efficient, \href{https://bluegreen-labs.github.io/ecmwfr/articles/advanced_vignette.html#batch-parallel-requests}{\code{wf_request_batch()}}
is used to download all files at the same time (up to 20 requests in parallel).

To be able to download data from the Climate Data Store (CDS), you will need to create an ECMWF
account on \url{https://www.ecmwf.int/}. Once created, you can
retrieve your API Token on \url{https://cds.climate.copernicus.eu/profile} and save them in your local keychain with:
\code{
  ecmwfr::wf_set_key("abcd1234-foo-bar-98765431-XXXXXXXXXX")
}

More information \href{https://geopressure.org/GeoPressureManual/geopressuretemplate-wind.html}{in the GeoPressureManual}.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
  tag_download_wind(tag)
\dontshow{\}) # examplesIf}
}
\seealso{
\href{https://bluegreen-labs.github.io/ecmwfr/reference/wf_request.html}{\code{wf_request_batch()} },
\href{https://geopressure.org/GeoPressureManual/trajectory-with-wind.html}{GeoPressureManual }

Other movement:
\code{\link[=bird_create]{bird_create()}},
\code{\link[=graph_set_movement]{graph_set_movement()}},
\code{\link[=plot_graph_movement]{plot_graph_movement()}},
\code{\link[=speed2prob]{speed2prob()}}
}
\concept{movement}
