% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_download_wind.R
\name{tag_download_wind}
\alias{tag_download_wind}
\title{Download wind data}
\usage{
tag_download_wind(
  tag,
  extent = tag$param$extent,
  stap_id = utils::head(tag$stap$stap_id, -1),
  cds_key = Sys.getenv("cds_key"),
  cds_user = Sys.getenv("cds_user"),
  file = function(stap_id)
    glue::glue("./data/wind/{tag$param$id}/{tag$param$id}_{stap_id}.nc"),
  overwrite = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{extent}{Geographical extent of the map on which the likelihood will be computed.
Vector of length 4 \code{c(xmin, xmax, ymin, ymax)} or \code{c(W, E, S, N)}.}

\item{stap_id}{Stationary period identifier of the start of the flight to download. Be default,
download for all flights.}

\item{cds_key}{User (email address) used to sign up for the ECMWF data service. See
\code{\link[=wf_set_key]{wf_set_key()}}.}

\item{cds_user}{Token provided by ECMWF. See \code{\link[=wf_set_key]{wf_set_key()}}.}

\item{file}{Absolute or relative path of the ERA5 wind data file to be downloaded. Function
taking as single argument the stationary period identifier.}

\item{overwrite}{logical. If \code{TRUE}, file is overwritten.}
}
\value{
The folder with the downloaded file (same as the directory).
}
\description{
This function downloads the wind data from \href{https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-pressure-levels?tab=overview}{ERA5 hourly pressure levels}
with the \href{https://cds.climate.copernicus.eu}{Climate Data Store (CDS)} and through the \href{https://bluegreen-labs.github.io/ecmwfr/index.html}{\code{ecmwfr} R package}.

The flights are determined from the stationary periods classified \code{tag$sta}
(see \code{tag_label_auto()}). It request a single file for each flight using the exact time
(hourly basis) and pressure (altitude). To make the download more efficient,
\href{https://bluegreen-labs.github.io/ecmwfr/articles/advanced_vignette.html#batch-parallel-requests}{\code{wf_request_batch()}}
is used to download all wind files at the same time (up to 20 requests in parallel).

To be able to download data from the Climate Data Store (CDS), you will need to create an account
on \url{https://cds.climate.copernicus.eu}. You can then save
your credentials (\code{cds_key} and \code{cds_user}) in your \code{.Rprofile} (see
\href{https://raphaelnussbaumer.com/GeoPressureManual/wind-graph.html#download-wind-data}{GeoPressureManual | Wind graph}).
}
\seealso{
\href{https://bluegreen-labs.github.io/ecmwfr/reference/wf_request.html}{\code{wf_request()}},
\href{https://raphaelnussbaumer.com/GeoPressureManual/wind-graph.html#download-wind-data}{GeoPressureManual | Wind graph }

Other movement: 
\code{\link{bird_create}()},
\code{\link{graph_add_wind}()},
\code{\link{graph_set_movement}()},
\code{\link{graph_transition}()},
\code{\link{plot_graph_movement}()},
\code{\link{speed2prob}()}
}
\concept{movement}
