% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label.R
\name{tag_label}
\alias{tag_label}
\title{Add label to tag}
\usage{
tag_label(tag, file = glue::glue("data/1_labels/{tag$id}-labeled.csv"))
}
\arguments{
\item{tag}{List containing the data logger dataset, this needs to contain at least a \code{pressure}
data.frame, but can also have a \code{light} and \code{acceleration} data.frame (see \code{\link[=tag_read]{tag_read()}}).}

\item{file}{Absolute or relative path of the label file.}
}
\value{
Same \code{tag} list with

(1) a \code{stap} data.frame describing the STAtionary Period:
\itemize{
\item \code{stap_id} unique identifier in increasing order 1,...,n
\item \code{start} start date of each stationary period
\item \code{end} end date of each stationary period
}

(2) an additional \code{label} and \code{stap_id} column on the sensor data.frame:
\itemize{
\item \code{date} datetime of measurement as POSIXt
\item \code{value} sensor measurement
\item \code{label} indicates the observation to be discarded (\code{"discard"} and \code{"flight"}) as well as
grouped by elevation layer (\verb{elev_*})
\item \code{stap_id} stationary period of the measurement matching the \code{tag$stap}.
}
}
\description{
This function performs the following operations:
\enumerate{
\item Read label file with \code{\link[=tag_label_read]{tag_label_read()}} and assign the label to a new column in each sensor
data.frame
\item Compute the stationary period \code{tag$stap} from the label and assign the corresponding \code{stap_id}
on all sensors data.frame with \code{\link[=tag_label_stap]{tag_label_stap()}}
}

If the label file does not exist, the function will suggest to create it with
\code{\link[=tag_label_write]{tag_label_write()}} and use \code{\link[=tag_label_classify]{tag_label_classify()}} if acceleration data exists.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX")

tag <- tag_label(tag)
str(tag)

}
\seealso{
\code{\link[=tag_read]{tag_read()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#edit-activity-on-trainset}{GeoPressureManual | Pressure Map }

Other tag_label: 
\code{\link{tag_label_classify}()},
\code{\link{tag_label_read}()},
\code{\link{tag_label_write}()}
}
\concept{tag_label}
