% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label.R
\name{tag_label}
\alias{tag_label}
\title{Label a \code{tag} object}
\usage{
tag_label(
  tag,
  file = glue::glue("./data/tag-label/{tag$param$id}-labelled.csv"),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{tag}{a GeoPressure \code{tag} object.}

\item{file}{Absolute or relative path of the label file.}

\item{quiet}{logical to display message.}

\item{...}{
  Arguments passed on to \code{\link[=tag_label_stap]{tag_label_stap}}
  \describe{
    \item{\code{warning_flight_duration}}{Threshold of flight duration to display warning for (hours)}
    \item{\code{warning_stap_duration}}{Threshold of stationary period duration to display warning for
(hours)}
  }}
}
\value{
Same \code{tag} list with

(1) a \code{stap} data.frame describing the STAtionary Period:
\itemize{
\item \code{stap_id} unique identifier in increasing order 1,...,n
\item \code{start} start date of each stationary period
\item \code{end} end date of each stationary period
}

(2) an additional \code{label} and \code{stap_id} column on the sensor data.frame:
\itemize{
\item \code{date} datetime of measurement as POSIXt
\item \code{value} sensor measurement
\item \code{label} indicates the observation to be discarded (\code{"discard"} and \code{"flight"}) as well as
grouped by elevation layer (\verb{elev_*})
\item \code{stap_id} stationary period of the measurement matching the \code{tag$stap}.
}
}
\description{
This function performs the following operations:
\enumerate{
\item Read label file with \code{tag_label_read()} and assign the label to a new column in each sensor
data.frame
\item Compute the stationary period \code{tag$stap} from the label and assign the corresponding \code{stap_id}
on all sensors data.frame with \code{tag_label_stap()}
}

If the label file does not exist, the function will suggest to create it with
\code{tag_label_write()} and use \code{tag_label_auto()} if acceleration data exists.
}
\examples{
setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE)

print(tag)

tag <- tag_label(tag)

print(tag)

# The labelled `tag` contains additional column on the sensor data.frame
str(tag)

}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/labelling-tracks.html}{GeoPressureManual }

Other tag_label: 
\code{\link{tag_label_auto}()},
\code{\link{tag_label_read}()},
\code{\link{tag_label_stap}()},
\code{\link{tag_label_write}()}
}
\concept{tag_label}
