% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_auto.R
\name{tag_label_auto}
\alias{tag_label_auto}
\title{Automatic labelling of tag from acceleration data}
\usage{
tag_label_auto(tag, min_duration = 30)
}
\arguments{
\item{tag}{List containing the data logger dataset, this needs to contain at least a \code{pressure}
data.frame, but can also have a \code{light} and \code{acceleration} data.frame (see \code{\link[=tag_create]{tag_create()}}).}

\item{min_duration}{Minimal duration (in minutes) to consider a high activity as migratory
flight.}
}
\value{
Same data logger list than input \code{tag}, but with the column \code{label} filled with
\code{"flight"} in the acceleration data.frame when a sustained high-activity period is detected.
}
\description{
This function uses acceleration data to classify migratory flights. The function uses a
\code{k=2} mean clustering (\code{\link[=kmeans]{kmeans()}}) to identify high activity periods. Periods lasting more than
\code{min_duration} are then considered to be migratory flight.
}
\details{
This function is inspired by the function \code{classify_flap} from the
\href{https://github.com/KiranLDA/PAMlr}{PAMlr package}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX")

tag <- tag_label_auto(tag, min_duration = 15)
str(tag$acceleration)

}
\seealso{
\href{https://kiranlda.github.io/tagLrManual/flapping.html}{flapping chapter of the tagLr manual}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#automatic-classification-of-activity}{GeoPressureManual | Pressure Map }

Other tag_label: 
\code{\link{tag_label_read}()},
\code{\link{tag_label_write}()},
\code{\link{tag_label}()}
}
\concept{tag_label}
