% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_read.R
\name{tag_label_read}
\alias{tag_label_read}
\title{Read classification of activity and pressure}
\usage{
tag_label_read(tag, file = glue::glue("data/1_labels/{tag$id}-labeled.csv"))
}
\arguments{
\item{tag}{List containing the data logger dataset, this needs to contain at least a \code{pressure}
data.frame, but can also have a \code{light} and \code{acceleration} data.frame (see \code{\link[=tag_read]{tag_read()}}).}

\item{file}{Absolute or relative path of the label file.}
}
\value{
Same data logger list as input, updated with the labels \code{tag$pressure$label} and
optionally \code{tag$acceleration$label}.
}
\description{
This function reads an exported csv file from \href{https://trainset.geocene.com/}{TRAINSET} and updates
the data logger dataset \code{tag}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX")

tag <- tag_label_read(tag)
str(tag)
}
\seealso{
Other tag_label: 
\code{\link{tag_label_classify}()},
\code{\link{tag_label_write}()},
\code{\link{tag_label}()}
}
\concept{tag_label}
