% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_label_stap.R
\name{tag_label_stap}
\alias{tag_label_stap}
\title{Create stationary periods from tag label}
\usage{
tag_label_stap(tag)
}
\arguments{
\item{tag}{List containing the data logger dataset, this needs to contain at least a \code{pressure}
data.frame, but can also have a \code{light} and \code{acceleration} data.frame (see \code{\link[=tag_read]{tag_read()}}).}
}
\value{
Same data logger list as input \code{tag} but with (1) a new data.frame of stationary periods
\code{tag$stap} and (2) a new column \code{stap_id} for each sensor data.
}
\description{
This function computes the stationary periods from the pressure and/or acceleration label data.

If an acceleration data.frame is present and contains a column \code{label}, the stationary period will
be computed from it, otherwise, it uses the pressure data.frame.

Stationary periods are defined by being separated by a flight (label \code{"flight"}). Flights of any
duration will be considered.

The function also adds the corresponding \code{stap_id} column to each sensor data.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |>
  tag_label_read()

tag <- tag_label_stap(tag)
str(tag)
str(tag$stap)

}
\seealso{
\code{\link[=tag_read]{tag_read()}}, \code{\link[=tag_label]{tag_label()}}, \code{\link[=tag_label_read]{tag_label_read()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#identify-stationary-periods}{GeoPressureManual | Pressure Map }
}
