% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_update.R
\name{tag_update}
\alias{tag_update}
\title{Update a \code{tag} object}
\usage{
tag_update(
  tag,
  file = glue::glue("./data/tag-label/{tag$param$id}-labelled.csv"),
  known = NULL,
  include_stap_id = NULL,
  quiet = FALSE
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object.}

\item{file}{absolute or relative path of the label file.}

\item{known}{data.frame containing the known positions of the bird (e.g., equipment or retrieval
site). The default is to use \code{tag$stap}, which assumes that the \code{stap_id} has not changed for the
known stationary periods.}

\item{include_stap_id}{vector of \code{stap_id} defining which stationary period to model, that is,
to compute in the likelihood map and use in the graph. The default is to use the same original
value.}

\item{quiet}{logical to hide messages about the progress}
}
\value{
The updated \code{tag} object
}
\description{
When updating the labelling file of a \code{tag}, we often change only a few stationary periods. To
avoid recomputing the entire workflow, this function identifies which stationary periods have
been modified and updates only these ones in \code{tag$map_pressure} and \code{pressurepath}.
}
\examples{
setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  tag_set_map(extent = c(-16, 23, 0, 50), scale = 1) |>
  geopressure_map(quiet = TRUE)

print(tag)

tag <- tag_update(tag,
  file = "./data/tag-label/18LX-labeled-updated.csv",
  quiet = TRUE
)

print(tag)
}
\seealso{
Other tag: 
\code{\link{print.tag}()},
\code{\link{tag_create}()},
\code{\link{tag_set_map}()}
}
\concept{tag}
