% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{trainset_edit}
\alias{trainset_edit}
\title{Edit classification of activity and pressure}
\usage{
trainset_edit(
  pam,
  pathname,
  filename = paste0(pam$id, "_act_pres-labeled.csv")
)
}
\arguments{
\item{pam}{pam logger dataset list}

\item{pathname}{Path to the folder where the labeled files should be
saved}

\item{filename}{Name for the file.}
}
\value{
pam logger dataset list updated with the new label named \code{class}
(\code{pam$pressure$class} and \code{pam$acceleration$class})
}
\description{
This function perform three steps: (1) write the \code{csv} file of the
automatically labeled activity and pressure with
\code{\link{trainset_write}}, (2) open trainset in your broweser
(\url{https://trainset.geocene.com/}) so that you can edit
the labels and (3) read the exported \code{csv} file from trainset with
\code{\link{trainset_read}}.
}
