% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{trainset_read}
\alias{trainset_read}
\title{Read classification of activity and pressure}
\usage{
trainset_read(
  pam,
  pathname,
  filename = paste0(pam$id, "_act_pres-labeled.csv")
)
}
\arguments{
\item{pam}{pam logger dataset list}

\item{pathname}{Path to the folder where the labeled file is.}

\item{filename}{Name of the file.}
}
\value{
pam logger dataset list updated with the new label named \code{class}
(\code{pam$pressure$class} and \code{pam$acceleration$class})
}
\description{
This function read an exported csv file from trainset
(https://trainset.geocene.com/) and update the pam logger dataset
}
\examples{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
pam_data <- trainset_read(pam_data,
  pathname = system.file("extdata", package = "GeoPressureR")
)
head(pam_data$pressure)
head(pam_data$acceleration)
}
