% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainset.R
\name{trainset_read}
\alias{trainset_read}
\title{Read classification of activity and pressure}
\usage{
trainset_read(
  pam,
  pathname = "data/1_pressure/labels/",
  filename = paste0(pam$id, "_act_pres-labeled.csv")
)
}
\arguments{
\item{pam}{pam logger dataset list}

\item{pathname}{Path to the folder where the labeled file is.}

\item{filename}{Name of the file.}
}
\value{
pam logger dataset list updated with the labels (\code{pam$pressure$isoutlier} and
\code{pam$acceleration$ismig})
}
\description{
This function read an exported csv file from trainset (https://trainset.geocene.com/) and update
the pam logger dataset
}
\examples{
pam <- pam_read(pathname = system.file("extdata/0_PAM/18LX", package = "GeoPressureR"))
pam <- pam_classify(pam)
pam <- trainset_read(pam, pathname = system.file("extdata/1_pressure/labels/",
  package = "GeoPressureR"
))
head(pam$pressure)
head(pam$acceleration)
}
\seealso{
\code{\link{pam_read}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#edit-activity-on-trainset}{GeoPressureManual | Pressure Map }
}
