% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainset.R
\name{trainset_write}
\alias{trainset_write}
\title{Write classification of activity and pressure}
\usage{
trainset_write(
  tag,
  pathname = "data/1_pressure/labels/",
  filename = paste0(tag$id, "_act_pres.csv")
)
}
\arguments{
\item{tag}{data logger dataset list}

\item{pathname}{Path to the folder where the labeled files should be saved}

\item{filename}{Name for the file.}
}
\description{
This function writes the csv file of the automatically labeled activity and pressure which can
be read with TRAINSET (https://trainset.geocene.com/).
}
\details{
Optionally, it can also export a reference dataset for pressure \code{tag$pressure$ref} as another
series to be visualized on TRAINSET, but without impacting the labeling process.
}
\examples{
tag <- tag_read(
  pathname = system.file(
    "extdata/0_tag/18LX",
    package = "GeoPressureR"
  )
)
trainset_write(tag, pathname = system.file("extdata/1_pressure/labels/",
  package = "GeoPressureR"
))
}
\seealso{
\code{\link{tag_read}}, \code{\link[=trainset_read]{trainset_read()}}, \href{https://raphaelnussbaumer.com/GeoPressureManual/pressure-map.html#edit-activity-on-trainset}{GeoPressureManual | Pressure Map }
}
