% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam.R
\name{trainset_write}
\alias{trainset_write}
\title{Write classification of activity and pressure}
\usage{
trainset_write(pam, pathname, filename = paste0(pam$id, "_act_pres"))
}
\arguments{
\item{pam}{pam logger dataset list}

\item{pathname}{Path to the folder where the labeled files should be saved}

\item{filename}{Name for the file.}
}
\description{
This function writes the csv file of the automatically labeled activity and pressure which can
be read with TRAINSET (https://trainset.geocene.com/).
}
\examples{
\dontrun{
pam_data <- pam_read(
  pathname = system.file("extdata", package = "GeoPressureR"),
  crop_start = "2017-06-20", crop_end = "2018-05-02"
)
pam_data <- pam_classify(pam_data)
trainset_write(pam_data, pathname = system.file("extdata", package = "GeoPressureR"))
}
}
