% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2mat.R
\name{ts2mat}
\alias{ts2mat}
\title{Format time series data frame into a matrix}
\usage{
ts2mat(
  ts,
  sensor = "time series",
  twl_offset = 0,
  value = "value",
  twl_time_tolerance = 180
)
}
\arguments{
\item{ts}{data.frame of a \code{tag}, containing at least \code{date} and \code{value}.}

\item{sensor}{sensor name used in messages.}

\item{twl_offset}{Shift of the middle of the night compared to 00:00 UTC (in hours). If not
provided, it uses the middle of all nights.}

\item{value}{column name to extract}

\item{twl_time_tolerance}{Maximum allowed time difference in seconds between observations
and the regular grid. Observations beyond this threshold will be set to NA. Default is 180
seconds (3 minutes).}
}
\value{
A data.frame with columns \code{date} and \code{value}.
}
\description{
Format time series data frame into a matrix
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
  mat <- ts2mat(tag$pressure)
\dontshow{\}) # examplesIf}
}
