% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2mat.R
\name{ts2mat}
\alias{ts2mat}
\title{Format timeserie dataframe into a matrix}
\usage{
ts2mat(ts, value = "value", twl_offset = 0)
}
\arguments{
\item{ts}{data.frame of a \code{tag}, containing at least \code{date} and \code{value}.}

\item{value}{column name to extract}
}
\value{
A data.frame with columns \code{date} and \code{value}.
}
\description{
Format timeserie dataframe into a matrix
}
\concept{twits}
