% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts2mat.R
\name{ts2mat}
\alias{ts2mat}
\title{Format timeserie data.frame into a matrix}
\usage{
ts2mat(ts, twl_offset = 0, value = "value")
}
\arguments{
\item{ts}{data.frame of a \code{tag}, containing at least \code{date} and \code{value}.}

\item{twl_offset}{Shift of the middle of the night compared to 00:00 UTC (in hours). If not
provided, it uses the middle of all nights.}

\item{value}{column name to extract}
}
\value{
A data.frame with columns \code{date} and \code{value}.
}
\description{
Format timeserie data.frame into a matrix
}
