% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_create.R
\name{twilight_create}
\alias{twilight_create}
\title{Estimate twilights from light data}
\usage{
twilight_create(
  tag,
  twl_thr = NULL,
  twl_offset = NULL,
  transform_light = TRUE,
  twl_time_tolerance = 180
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{twl_thr}{Light threshold that defines twilight. By default (\code{NULL}), it uses the smallest
value of light (i.e, first and last light of day).}

\item{twl_offset}{Shift of the middle of the night compared to 00:00 UTC (in hours). If not
provided, it uses the middle of all nights.}

\item{transform_light}{logical to use a log transformation of light}

\item{twl_time_tolerance}{Maximum allowed time difference in seconds between observations
and the regular grid. Observations beyond this threshold will be set to NA. Default is 180
seconds (3 minutes).}
}
\value{
a \code{tag} list containing a new data.frame \code{twilight} with columns:
\itemize{
\item \code{twilight} (date-time of twilight)
\item \code{rise} (logical) indicating sunrise (\code{TRUE}) or sunset (\code{FALSE}).
\item \code{stap_id} if \code{stap_id} is present in \code{light}.
}
}
\description{
This function estimates twilight (i.e., datetime of sunrise and sunset) by searching for the
first and last light of the day which is defined by the light exceeds or falls below the light
threshold \code{twl_thr}.
}
\examples{
withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)
})

# Create twilight data.frame
tag <- twilight_create(tag)

str(tag$twilight)

plot(tag, type = "twilight")

}
\seealso{
\href{https://geopressure.org/GeoPressureManual/light-map.html}{GeoPressureManual }, \href{https://rdrr.io/github/slisovski/TwGeos/man/findTwilights.html}{\code{TwGeos::findTwilights()}}

Other geolight:
\code{\link[=geolight_fit_location]{geolight_fit_location()}},
\code{\link[=geolight_map]{geolight_map()}},
\code{\link[=geolight_solar]{geolight_solar()}},
\code{\link[=plot_twl_calib]{plot_twl_calib()}},
\code{\link[=twilight_label_read]{twilight_label_read()}},
\code{\link[=twilight_label_write]{twilight_label_write()}}
}
\concept{geolight}
