% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_create.R
\name{twilight_create}
\alias{twilight_create}
\title{Estimate twilight from continuous light data}
\usage{
twilight_create(tag, twl_thr = NA, twl_offset = NA)
}
\arguments{
\item{tag}{List containing the data logger dataset, this needs to contain a \code{light}data.frame
with columns \code{date} and \code{value} and optionally \code{stap_id} (see \code{\link[=tag_read]{tag_read()}}).}

\item{twl_thr}{Light threshold that defines twilight. By default, it uses the smallest
value of light (i.e, first and last light of day).}

\item{twl_offset}{Shift of the middle of the night compared to 00:00 UTC (in seconds). If not
provided, it uses the middle of all nights.}
}
\value{
A \code{tag} list contianing a new data.frame \code{twilight} with columns:
\itemize{
\item \code{twilight} (date-time of twilight)
\item \code{rise} (logical) indicating sunrise (\code{TRUE}) or sunset (\code{FALSE}).
\item \code{stap_id} if \code{stap_id} is present in \code{light}.
}
}
\description{
Search for the time of sunset and sunrise, corresponding to the first time light exceeds or falls below
a given light threshold.

Function inspired from \href{https://rdrr.io/github/slisovski/TwGeos/man/findTwilights.html}{\code{TwGeos::findTwilights()}}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |> tag_label()
tag <- twilight_create(tag)
str(twilight)
}
\seealso{
\href{https://raphaelnussbaumer.com/GeoPressureManual/light-map.html}{GeoPressureManual | Light Map }

Other twilight: 
\code{\link{twilight_label_read}()},
\code{\link{twilight_label_write}()}
}
\concept{twilight}
