% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_label_read.R
\name{twilight_label_read}
\alias{twilight_label_read}
\title{Read label file of light}
\usage{
twilight_label_read(
  tag,
  file = glue::glue("data/2_light/{tag$id}-labeled.csv")
)
}
\arguments{
\item{tag}{List containing the data logger dataset (see \code{\link[=tag_read]{tag_read()}}) It needs to contain
a \code{twilight} data.frame created with \code{\link[=twilight_create]{twilight_create()}}.}

\item{file}{csv file of the labels to be read.}
}
\value{
Same twilight data.frame as input, updated with the labels \code{tag$twilight$label}.
}
\description{
This function read an exported csv file from TRAINSET \url{https://trainset.geocene.com/} and update
the data.frame twilight.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_read("18LX") |> tag_label()
tag <- twilight_create(tag)

twilight <- twilight_label_read(twilight)
str(twilight)
}
\seealso{
Other twilight: 
\code{\link{twilight_create}()},
\code{\link{twilight_label_write}()}
}
\concept{twilight}
