% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_label_read.R
\name{twilight_label_read}
\alias{twilight_label_read}
\title{Read a twilight label file}
\usage{
twilight_label_read(
  tag,
  file = glue::glue("./data/twilight-label/{tag$param$id}-labeled.csv")
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{file}{csv file of the labels to be read.}
}
\value{
Same \code{tag} object, updated with the labels \code{tag$twilight$label}.
}
\description{
This function read an exported csv file from TRAINSET \url{https://trainset.raphaelnussbaumer.com/}
and update the twilight data \code{tag$twilight}.
}
\examples{
owd <- setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  twilight_create()

tag_labeled <- twilight_label_read(tag)
setwd(owd)

plot(tag, type = "twilight") + ggplot2::ggtitle("Before label")

str(tag_labeled$twilight)

plot(tag_labeled, type = "twilight") + ggplot2::ggtitle("After label")
}
\seealso{
Other geolight: 
\code{\link{geolight_map}()},
\code{\link{twilight_create}()},
\code{\link{twilight_label_write}()}
}
\concept{geolight}
