% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilight_label_write.R
\name{twilight_label_write}
\alias{twilight_label_write}
\title{Write a twilight label file}
\usage{
twilight_label_write(
  tag,
  file = glue::glue("./data/twilight-label/{tag$param$id}.csv")
)
}
\arguments{
\item{tag}{a GeoPressureR \code{tag} object}

\item{file}{Name of the twilight label file to be saved.}
}
\description{
This function writes the csv file of the labelled twilight which can be read with
TRAINSET \url{https://trainset.raphaelnussbaumer.com/}.
}
\examples{
setwd(system.file("extdata/", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |>
  tag_label(quiet = TRUE) |>
  twilight_create()

label_file <- twilight_label_write(tag)

str(read.csv(label_file))
}
\seealso{
Other geolight: 
\code{\link{geolight_map}()},
\code{\link{twilight_create}()},
\code{\link{twilight_label_read}()}
}
\concept{geolight}
