% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsupport.R
\name{windsupport}
\alias{windsupport}
\title{Compute wind support and drift}
\usage{
windsupport(windspeed, groundspeed, drift = FALSE)
}
\arguments{
\item{windspeed}{windspeed as complex value}

\item{groundspeed}{groundspeed as complex value}

\item{drift}{return drift instead of windsupport}
}
\value{
wind support vector as complex value (same unit as \code{gs} and \code{ws})
}
\description{
Compute the wind support (or wind profit) from groundspeed and windspeed.

All speed vectors are expressed as complex value where the real part corresponds to the east-west
component and imaginary part to the north-south component. You can use \code{abs()} (or \code{Mod()}) to
compute the norm (or absolute value) of the speed vector and \code{speed_to_bearing()} to compute the
bearing/orientation of the speed vector. The latter simply compute the trigonometric angle of the
speed vector with \code{Arg()} and convert this angle to a bearing (North = 0° instead of 90°).

Alternatively, if \code{drift = TRUE}, the function return the drift instead of wind support.
}
