% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windsupport.R
\name{windsupport}
\alias{windsupport}
\title{Compute wind support and drift}
\usage{
windsupport(windspeed, groundspeed, drift = FALSE)
}
\arguments{
\item{windspeed}{windspeed as complex value}

\item{groundspeed}{groundspeed as complex value}

\item{drift}{return drift instead of windsupport}
}
\value{
wind support as scalar value (same unit as \code{gs} and \code{ws})
}
\description{
Wind support, or wind profit, is the projection of the wind vector $\vec{v}_w$ onto the
groundspeed vector $\vec{v}_g$,
\text{Wind Support} = \frac{\vec{v}_g \cdot \vec{v}_w}{|\vec{v}_g|},
with $\cdot$ being the \href{https://en.wikipedia.org/wiki/Dot_product}{cross product}.

A positive value indicates that wind was blowing in the direction of movement, while a negative
one indicates a head wind.

\code{windspeed} and \code{groundspeed} should be expressed as complex value where the real part
corresponds to the east-west component and imaginary part to the north-south component.

Alternatively to wind support, you can compute the drift (or crosswind component) value with
\code{drift = TRUE},
\text{Drift} = \frac{|\vec{v}_g \times \vec{v}_w|}{|\vec{v}_g|},
with $\times$ being the \href{https://en.wikipedia.org/wiki/Cross_product}{cross product}.

A positive value of drift indicates that the wind is pushing the bird to the right of its
intended path, while a negative indicates that the wind is pushing the bird to the left.

You can use \code{abs()} (or \code{Mod()}) to compute the norm (or absolute value) of the speed vector and
\code{speed_to_bearing()} to compute the bearing/orientation of the speed vector. The latter
computes the trigonometric angle of the speed vector with \code{Arg()} and convert this angle to a
bearing (North = 0° instead of 90°).
}
